[test] Add OpenOCD compliance test
Based on @imphil's CI system tests infrastructure. This commit adds a
test that runs the OpenOCD compliance test against a Verilator target.
The test takes in external parameters to make it easy to integrate with
CI.

Sample run (takes almost 2m to complete):

```console
$ pytest -s -v test/systemtest/openocd_verilator_test.py \
  --test_bin=sw/build/sw.vmem \
  --rom_bin=sw/build/rom.vmem \
  --verilator_model=build/lowrisc_systems_top_earlgrey_verilator_0.1/sim-verilator/Vtop_earlgrey_verilator
```

Integration with CI is blocked until issue #413 is resolved.