[doc] Document installation of Vivado 2019.1
Vivado 2019.1 has slighly different installation steps as it can now
use the batch installation flow.

In addition to the installation of Vivado, we now need to install point
release updates as well, so document that.

Finally, the previous instructions didn't document how to run Vivado.
It's not fully obvious from the command line, hence a new paragraph is
added on that as well.

Fixes #89