[reggen] Correct calculation of interrupt width
Issue was found during the code review #516.

Error Symptom:

    If a IP has interrupts having more than one bit width, the validate
    has a chance to not raises ERROR for the interrupt exceeding full
    register width (currently 32).

auto interrupt register generation was added to create INTERRUPT
enable/status/test registers by the tool. It does have a limitation that
only supports up to 32bit interrupts in a IP. For instance, if IP has 33
interrupts, it raises an ERROR and stop generating.

However, GPIO has only one interrupt signal but the width of the
interrupt is 32bit width. If it is set to 33 bit width, tool doesn't
create an error. as the checker only checks:

```python
if len(m['interrupt_list']) > fullwidth:
```

It should check the `width` field and account it to compare with
`fullwidth`.

Results:

    Changed to sum the width of the interrupt_list

```python
sum([int(x['width']) if 'width' in x else 1 for x in
m['interrupt_list'])
```