[reggen] Wrap output to meet code standards
Fixes #578

regtool could output comments and #defines that went over the 80
character line limit leading to a failure of C/C++ lint checking. This
fixes all instances of that currently seen but doesn't entirely prevent
it.

Included with the commit are the new results of running regtool which
pass the C/C++ lint checks.