[rv_plic] Implement a binary tree variant of the rv_plic_target module
This implements a solution that uses a binary tree to gather the
interrupts and forward the highest priority request. As such, it
offers O(N) area and O(log(N)) delay complexity. This solution is
similar in area complexity as the SEQUENTIAL implementation, and
it offers similar delay as the MATRIX implementation, but without
the O(N^2) area cost.

Signed-off-by: Michael Schaffner <msf@google.com>