[nmi_gen] Add the nmi_gen peripheral for alert_handler testing purposes
The nmi_gen peripheral is a simple wrapper module that receives the
escalation signals from the alert handler and converts them to
interrupts for the core. These are currently not true non-maskable
interrupts. The module is only used for testing the alert handler
mechanism at the moment, and will be adapted and/or replaced in the future.