[ci] Enhance run_sw_build.sh create dist dir
A distribution directory is the "output directory" where we stage build
artifacts which will become part of a binary release. Enhance the
run_sw_build.sh script to copy the build output into such a directory.

This fix is temporary until we have built-in support in the build
system, where this functionality should be.