[doc] Document the installing of all required packages
We have a section at the beginning of the "Install Build Requirements"
outlining the packages which need to be installed. This section included
not all packages, some others were distributed across the guide. Worse,
the library dependencies for building the spiflash tool were missing
(they are only listed in the README of the tool itself).

I've restructured the section to include all packages which we need.