[hugo] Install hugo if not present
Developers would like to have an easy way to preview the documentation.
Since hugo is distributed as a static binary, we can simply download it
and drop it into the build folder if needed.

Implementation note: The curl/tar option isn't overly beautiful, but
using Python to download and then calling tar didn't seem significantly
nicer to invest too much time into it.