[ci] Don't run full builds for doc-only changes in PRs
Avoid running the full set of builds if only markdown files change.
That's a very rough heuristic and will need fine-tuning, but should be a
first step to reduce the load.

Note that this applies only to pull request triggers, the master changes
will always receive a full set of builds, catching things which might
have been overlooked in the PR pipeline.

Note 2: This commit removes the stages we have currently, and uses
explitit dependencies instead. That's to work around a missing feature
in Azure Pipelines, where variables cannot be passed between stages
(https://stackoverflow.com/a/57488169/2948140).