[tlul] Fixed TL-UL error response of d_size
Problem:

    d_size of unmapped address was tied to 0

`tlul_err_resp` module is used to return the error response inside TL-UL
crossbar when a host tries to access unmapped address space. It tied the
`d_size` to 0 regardless of the request size. But TL specification
requires the device to return the `d_size` same as `a_size`.

Resolution:

    Latch `a_size` and return the value to `d_size`.

The issue is reported by @weicaiyang and issue #822 was created.