[checklist] Change D1 checklist to instantiable
Changed the IP_INSTANTIATED to IP_INSTANTIABLE.

Previous term isn't appropriate. IP in D1 stage doesn't have to be
included in the top level. The intention of the checklist item is to
check whether it breaks the existing test items in the top.

Conforming to the [comportability spec][] doesn't mean that the IP
itself doesn't affect outside of the IP. For instance, some interrupt
keeps bugging the interrupt controller. This checklist item is to make
sure it is ready to test the main datapath in the top level simulation
or on the FPGA.

Signed-off-by: Eunchan Kim <eunchan@google.com>