[sw/device] Drop all linker sections not specified
ld will sometimes (I guess it depends on how it's built) insert a
.note.gnu.build-id symbol into the ELF. This defaults to being at the
start of the ELF. If this happens we get non-code at the start of the
ELF which causes an illigal instruction to be executed when jumping from
ROM to a sw binary loaded into flash.

The PR https://github.com/lowRISC/opentitan/pull/966 attempted to fix
the issue using the build-id command line. This PR instead disacrds all
non specified linker sections. We have to add .text.startup to ensure
the build does not fail.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>