[dashboard] Add multiple revisions
As IP grows and updated, it is needed to have multiple versions of IPs
to be managed in the dashboard. For instance, as UART, GPIO, RV_TIMER
were signed off, their versions shall be increased when the logic design
is changed.

This change introduces `revisions` field in the project Hjson file. If
this field is defined, the dashboard prints out new format.

Revised `uart` IP to show the example.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>