[util] get-toolchain.py: Support JSON buildinfo files
Current toolchain builds from
https://github.com/lowRISC/lowrisc-toolchains contain a human-readable
"buildinfo" file with the version of the build, among other things.
Since https://github.com/lowRISC/lowrisc-toolchains/pull/6 the toolchain
builds also contain a buildinfo.json file with a superset of this
information in JSON, which is much better to read from a script.

This commit updates the get-toolchain.py script to make use of this new
JSON build information, if available, and fall back to the old plaintext
file otherwise.

To help the implementation of that feature I also did the following
cleanups:

- (Behavior) Remove the `--force` option, and imply it with `--update`
  if a existing toolchain installation was detected; bail out with an
  error if the target path doesn't a toolchain yet. This gives us
  roughly the same safety net as before, while simplifying the logic.
- (Fix) Fix buildinfo regex to work with all known buildinfo files.
- (Refactoring) Use pathlib.Path consistently.
- (Refactoring) Shuffle around functions to get build information, and
  improve their naming.

Overall we get a slight decrease in lines-of-code and more
functionality.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>