[doc] Mandate C11 in the C/C++ Coding Style Guide
This matches the C version available when C++14 was standardised.

It is prudent to avoid nonstandard language extensions so that using
different compilers is easier. The style guide now includes a short
whitelist of allowed nonstandard extensions.

The style guide now also includes a more comprehensive set of
requirements and suggestions around builtin usage.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>