[systemtest] Avoid time increments after compare
We didn't use the same time value for a comparision and later on in a
statement, causing a negative value to be passed to select.select().
This leads to "ValueError: timeout must be non-negative".

Avoid the problem by buffering the current time for the check and the
socket.select() call.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>