Change version text to git hash
`OT_VERSION` will now contain something like "opentitan-43c98be" instead
of "opentitan-snapshot-20191101-1-307-g43c98be".
The tag based name is now used only for the name of the archive for a
release.
Removing the tag reference from the version reduces the likelihood of
confusion that a version is actually a snapshot.

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>