[aes] Correct file type for .c/h files
These files were originally defined as systemVerilogSource instead of
cSource to avoid problems with Synopsys VCS. In fact doing this leads to
other issues when VCS needs to compile c/h files located in different
directories, as the include paths cannot be set up correctly.

Setting the file type for all c/h files to cSource and leaving only the
.sv files in the DPI model as systemVerilogSource solves these problems.
The design can be compiled and run without any issues.

This is related to lowRISC/opentitan#1150 and lowRISC/opentitan#953.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>