[dif_uart] Introduce UART DIF library
This change introduces an UART DIF, a low level library for UART
hardware manipulation. The set of functions in this library can be
used in an ISR context, as well as normal execution context (It is up
to the higher level driver or OS to guard against the race conditions).

The init function is responsible for initialising the UART state,
which must be passed to all other DIF API.

This change is written to RFC #1043, and fixes issue #1076

Limitations:
* Implementation expects that only one thread of execution exclusively
  owns the hardware resource. This is a subject to implementation in a
  higher level driver, or OS.
* Implementation does not protect against race conditions that can
  occur between the ISR context and normal execution context. This
  responsibility also lies on a higher level driver or OS.

Current state:
* Fundamental API has been introduced, which should be sufficient for an UART
  driver needs, however it may be evolved and extended. For example, private
  user data could be added to the state structure.
* This API is largely untested

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>