[topgen/tlgen] Unify address range for device
Device ports for Xbar and leaf module have different field for address
range. Xbar has `xbar_addr` that is list of addresses. And a device port
connecting to leaf module has `base_addr` and `size_byte` field.

This commit is to merge two different address fields into `addr_range`
and let `tlgen` to handle the one entry of address (for leaf module) and
list of the addresses (for xbar).

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>