DRAFT: Add mock_mmio.c/h for DIF unit-tests (alternative)
I had the chance to take a look at @mcy's #1298 before creating this PR
and it seems like both approaches are starting to look similar. I'm
creating this PR so that we can take a look at both and come up with a
solution that combines the best parts.

This PR is a patch on top of #1254 and uses GoogleTest and GoogleMock
to provide a test fixture with `EXPECT_MMIO_READ()` and
`EXPECT_MMIO_WRITE()` similar to my previous PR. The test fixture in
`dif_gpio_test.cc` further encapsulates the `gpio` instance to simplify
tests. We can move this to `mock_mmio.h` in the form of a class template
if all device instances has the `base_addr` field.

Signed-off-by: Alphan Ulusoy <alphan@google.com>