[lint] Preliminary Verilator/AscentLint lint flows for AES
This patch adds preliminary linting support for Verilator and
AscentLint for the AES module (the lint targets and waiver files will
follow in a subsequent commit).

Support for AscentLint is preliminary, since the flow still requires a
proprietary lint policy file, which is not publicly available. However,
this lint policy will be available soon as a built-in lint policy in the
upcoming 2019.A.p3 release of the tool. The lint waivers will all be
publicly available, and the lint makefile supports a summary report option
that filters out the relevant messages from a batch run. Note that lint
report information must be filtered using this summary report script
before publishing openly.

The Verilator lint waivers are only partially populated at the moment, and
hence this linting flow still throws many warnings and messages. Further,
there is no summary report generation for Verilator lint yet.

Signed-off-by: Michael Schaffner <msf@opentitan.org>