[sw, tests] Introduce a test to handle two different consecutive IRQs
The purpose of this test is to validate external interrupt delivery from
peripheral to the target. The aim is not to test the entire PLIC or entirety
of the peripherals, but to ensure that two different IRQs can interrupt the
processor and be successfully handled. For the purpose of this test the UART
peripheral is used as the origin of the interrupts, and the interrupted target
is Ibex (the only target in current Earl Grey implementation).

This test ensures (in terms of UART RX Overflow and TX Empty IRQ sources, and
Ibex target):

* That PLIC getaway receives and handles an IRQ correctly.
* That PLIC Core and Target modules handle an IRQ correctly.
* That an IRQ can interrupt the target.
* That an IRQ can be cleared and a new IRQ can be processed.

The reason for introducing this test was an issues we had previously with the
IRQ delivery on the FPGA, see issue #1355 for more details. This test can be
run by CI, and used manually to test the IRQs on an FPGA.

Fixes: #1402

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>