[topgen] Support windowed registers in generated SVD.
SVD doesn't have a direct mapping to register windows. Instead, it
allows a <register> to contain a <dim> element, stating the size of
the register.

This seems mostly correct, except SVD also wants each element in the
<register><dim> to be uniquely accessible/have a distinct name. This
is done via including the magic string "%s" in the register name.
This is simple enough to do, but this is the only case where svdgen
mangles a name.

Signed-off-by: Nathan Holstein <nathan.holstein@gmail.com>