[topgen] Python 3.6 compatibility in SVD generation.
A few cases of the Python 3.8 walrus operator prevented compatibility
with Python <3.8. The changes are fairly small to support running on
back to version 3.6.

Note that while the codebase is now compatible back to a least Python
3.5 the output is incorrect. This is due to 3.5 and earlied randomizing
the keyword arguments dictionary used to define XML elements. The
resulting SVD file has the correct elements, but the order is random.
Unfortunately, the hardware development Docker image uses Ubuntu 16.04
which ships with Python 3.5.

Signed-off-by: Nathan Holstein <nathan.holstein@gmail.com>