[svdgen] Convert svdgen into a standalone utility.
This uses common HJSON loading logic from topgen to read and validate
the top-level configuration and IP definitions. It then converts this
to an SVD file.

This updates the rules in hw/Makefile to generate the SVD as a step
during HW build. All configuration was moved from the Python script
to the makefile to allow for greater flexibility.

Signed-off-by: Nathan Holstein <nathan.holstein@gmail.com>