[aes/rtl] Rework naming and use of cipher operation signal
This commit changes the name of this signal from "mode" to "op"
(operation) and adds a separate "op" signal to the cipher core.

In AES terms, "mode" usually refers to the block cipher mode (e.g.
ECB, CBC, CTR) whereas "operation" refers to whether encryption
or decryption is performed. Depending on "mode" and "operation",
the actual cipher operation can either be "forward" or "inverse".

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>