[doc] Add render shortcode
The 'render' shortcode takes a URL and, when the page loads, fetches it,
renders it as markdown, and inserts it into the page at the position of
the invocation of the shortcode.

Using it is fairly straightforward:

   {{< render "https://some_url" >}}

This does require that javascript be enabled in the viewer's browser and
that the referenced URL have cross-origin resource sharing enabled.

Updates #1411, #1417

Signed-off-by: Garret Kelly <gdk@google.com>