[svdgen] Support clustered window registers.
Window register HSJON don't have their "genoffset" set in their outter
object; instead this is set as a field on the inner "window" object.

Placing a "window" within a "multireg" requires computing the address
offset of the window, which requires this little bit of gymnastics to
detect and return the correct value.

Signed-off-by: Nathan Holstein <nathan.holstein@gmail.com>