[svdgen] Support running tests on Python 3.5
Older versions of Python randomize the iteration order of **kwargs.
The results the SVD tree's text nodes appearing in random order.
This doesn't impact the semantics of the SVD, and because only the
smaller, child-less text nodes are reordered legibility is nearly
on par with Python >= 3.6.

However, for the tests which compare XML structure we must ensure
to compare all the nodes in the proper order. The easiest way is
to just sort them. This has the added benefit of making the tests
slightly less finicky regarding the exact structure of the generated
SVD trees.

Signed-off-by: Nathan Holstein <nathan.holstein@gmail.com>