[rv_timer] Int clr by mtimecmp update
RISC-V Previliged Spec mentioned that the interrupt line is dropped by
writing values to `mtimecmp`. @pfmooney raised the issue that the
RV_TIMER in OpenTitan doesn't follow this (see issue #1404 )

This commit is to change the rv_timer behavior to support clearing the
interrupt by updating `mtimecmp` CSRs.

Please note that the writing 1 to the interrupt status register also
clears it to support current VIP and the firmware.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>