Add out-of-tree board definition for Tock
Adds an OpenTitan board definition outside of the Tock repository. This
is essentially a copy of the upstream board definition with dependencies
changed to the upstream Tock Git repository and layout.ld +
kernel_layout.ld squashed into a single file.

The out-of-tree definition will enable custom build configurations from
within the OpenTitan tree. This is particularly important as a means of
linking between Tock and the OpenTitan DIFs.

Signed-off-by: Jon Flatley <jflat@google.com>