[sw/tock] Add out-of-tree OpenTitan board definition
Adds an OpenTitan board definition outside of the Tock repository. This
is essentially a copy of the opentitan board from the Tock repository
with the relative paths adjusted and Makefile removed.

The out-of-tree definition will enable custom build configurations from
within the OpenTitan tree. This is particularly important as a means of
linking between Tock and the OpenTitan DIFs.

Signed-off-by: Jon Flatley <jflat@google.com>