Setup build scripts for Tock
Tock uses Make to invoke Cargo as its build systems. There's no real
reason to use Make on the out-of-tree opentitan board, but there is no
avoiding Cargo.

Attempt to draft a meson.build file that produces meaningful artifacts
from the Tock build. A small script at util/build_tock.sh is used to
wrap the actual call to Cargo as a workaround for some shortcomings of
Meson. build_tock.sh is intentionally kept as small as possible.

The build.meson file adds a few targets to build Tock and export the
resulting .elf and .bin files to build-bin/. The Tock build does not
produce a disassembled .dis file as generating it makes the build much
slower.

Signed-off-by: Jon Flatley <jflat@google.com>