[sw/dif] Add tests for HMAC DIF
This change adds some simple hardware smoke tests for the HMAC DIF.

These tests perform the following checks/actions:
1. Configure the HMAC engine.
2. Push a message into the HMAC engine.
3. Check that the HMAC engine is reporting the correct message
   length.
4. Kick off a run of HMAC/SHA256.
5. Read the digest and compare the result to the expected result.

This test is run in both HMAC and SHA256-only modes. The purpose of
these tests is to provide some basic sanity checks on hardware and
to serve as a working example of how to interface with the DIF.

Unit tests of the HMAC DIF are to follow.

Signed-off-by: Jon Flatley <jflat@google.com>