[topgen] Use OrderedDict for top, xbar objects
This is related to PR #1552.

Problem:

    In Azure Pipeline that uses Python 3.5, doesn't keep the order of
    hjson keys. It results the generated hjson has a lot of change
    randomly.

`dict` type in python is unordered dictionary. It doesn't maintain the
order of keys in the dictionary. `topgen` dumps a couple of hjson files,
such as generated tops (e.g.
`hw/top_earlgrey_data/autogen/top_earlgrey.gen.hjson`) and the crossbar
configurations (e.g.
`$TOP/ip/xbar_{name}/data/autogen/xbar_{name}.hjson`). This unordered
behavior creates random changes to the generated files, which increases
commit diffs.

From python3.7, the default dictionary type is changed to
`collections.OrderedDict`. So, if the `topgen` runs on the system >=
python3.7, it always maintain the order.

Resolution:

    Use OrderedDict when create dictionary variables.

With OrderedDict (introduced in Python2.7), the generated hjson files
are remains in order.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>