[sw,dif_plic] Rename k*Count Enum Variants
kFooCount enum variants (which come last) are useful for being able to
iterate over the domain of an enum type. However, they add an additional
variant which needs to be checked against during DIF preconditions (as
usually this `Count` does not mean anything semantically).

This replaces kFooCount enum variants with kFooLast enum variants.
kFooLast should have the same value as the last valid variant, so do
not need to be checked against in preconditions, and they can still be
used for iteration (but are off-by-one when counting).

Signed-off-by: Sam Elliott <selliott@lowrisc.org>