[CI] Optionally pass repository source to template
We have an Azure Pipelines template to install all build requirements for
OpenTitan, which assumes it is being called from the root of the
OpenTitan repository. This assumption is violated if the template is
included from another repository, as described in the docs at
https://docs.microsoft.com/en-us/azure/devops/pipelines/process/templates?view=azure-devops#use-other-repositories.

For this use case, this commit introduces a template parameter
`REPO_TOP`, which defaults to `.`. In cases where this default isn't
applicable (i.e. whenever this template is used from another repo), it
can be overwritten as needed.

The name `REPO_TOP` reflects the naming we use in the OpenTitan
documentation for the root of the opentitan repository.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>