Recommend installing python3-yaml as distro package
The Python yaml package (pyyaml) can optionally use a YAML parser/dumper
written in C called LibYAML. This library is roughly 10x faster than the
fallback implementation written in Python. Since fusesoc (and fusesoc
generators) use YAML quite extensively, switching to a faster YAML
implementation decreases build and simulation times measurably (10s of
seconds in my tests).

pip can install pyyaml with LibYAML support as long as the Python
development headers, a C compiler, and LibYAML are present. Or we can
avoid compiling the package on a user's machine, and use a precompiled
version provided by the distribution. That's what this patch is doing.
The pip version is still available as fallback in
python-requirements.txt for users not installing all requirements
from the apt-requirements.txt file (or the user guide, which uses this
information as well).

Signed-off-by: Philipp Wagner <phw@lowrisc.org>