[dif_pinmux] Introduce DIF Pinmux library
Pinmux (Pin Multiplexer) allows to connect any of the supported
peripheral inputs and outputs to any of the supported
MIO (Multiplexable IO) pins.

Pinmux is closely related to the Pad Control module, which provides
the means to control chip pin physical characteristics, and adds
some safeguards.

Pinmux inputs and outputs are connected to the chip MIO pins through the
Pad Control module. It is legal to have a peripheral input and output
connected to the same MIO pin, which means that the input will have
the value of the driven output signal.

One peripheral input can be multiplexed to one or none MIO pins by writing
to the INSEL register. Equally one MIO output can be multiplexed to one
or none peripheral inputs by writing to the OUTSEL register.

Pinmux is quite flexible. It is possible for example to connect all
peripheral inputs to a single MIO pin, or all MIO pins to a single
peripheral output.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>