[sw/crt] Initialise Global Pointer Correctly
`__global_pointer$` is used by GNU ld during linker relaxation to
represent the value in `gp`, so that some accesses can have their
addresses materialised directly in the load/store instruction, rather
than having to do `lui; addi` or similair.

The real payoff of this relaxation is for loads and stores of small
globals. For this reason, we want to ensure that the small data sections
come before their respective regular sections, so they have a higher
probability of landing within a 12-bit signed immediate of
`__global_pointer$`. This commit reorders these sections and describes
why.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>