[sw] Reformat Software Code
Recently a bug was found in CI where the code formatting hooks were not
firing correctly. This bug was fixed in
lowrisc/opentitan#ffa126f00f5cae43bd8ccd1e212550ab2d997491.

This change reformats the existing software tree, so that going forwards
we know the software tree is and remains correctly formatted.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>