[sw] Fix Software Include Guards
Recently a bug was found in CI where the code formatting hooks were not
firing correctly. This bug was fixed in
lowrisc/opentitan#ffa126f00f5cae43bd8ccd1e212550ab2d997491.

This change ensures that the include guards on existing code are
correct, so that going forwards we know the software tree is and remains
correct.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>