[doc] Generate Doxygen and add DIF listings
This commit adds Doxygen API documentation (for device code) to the docs
site and uses the Doxygen XML to create lists of DIFs for each
peripheral, which can be included on the peripheral page.

The Doxygen site still needs styling and some other work, which will
happen in a follow-up commit.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>