[dv,sw] SW -> DV tb self-checking mechanism
This commit adds SW to DV testbench self checking mechanism. It uses a
reserved location in RAM (`test_reserved_size` region) to monitor the
status of the SW test running on the Ibex core.
- Added test status lib in C for the SW to be able to write the status
(implemented as an enum). Apart from pass and fail, the enum provides
indications for these: executing boot rom, executing the test, in WFI,
etc. These are loosely defined at the moment and can change in future.

- Added sw test status interface on the DV side, hooked it up to the
testbench, connected it to the RAM (so that it can monitor the writes to
that location)

- Added `chip_sw_test_base_vseq` as an abstraction for chip level tests
run with SW running on the CPU. This enables the SW logger and the test
status monitor.

- Related fixes to supporting code, meson build files.

- Fixed incorrect calculation of number of variable args in `log.h`.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>