[chip dv] Add support for %s in DV logging
This change adds support for %s format specifier.
- Updated python script to provide information which arguments to the
format are strings based on the positions of %s in the format string.
- Updated the SV logger to process the rodata contents into a data
structure
- When the log pointer is written, it is followed by the arg values. If
the corresponding arg value is supposed to be a string, the arg value is
treated as the address in the rodata that contains the string. The
logger performs this lookup and prints the string instead.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>