[spi_device] Remove latches
Problem:

    SPI TXF state machine has had latches inside.

The state machine is coded in a way based on the state machine table
described in the specification. The last part in `StPush` state is to
move to `StUpdate` based on the condition left over (fifo_ready &&
cnt_eq_end). Mistakenly, the condition is added as `else if`. It should
be `else` to eliminate the latches. The condition is matched to what
left over anyway.

The state machine is fixed to have `else` rather than `else if` now.

This is related to #2001

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>