[topgen] Clang-format ignore top_earlgrey.{c,h}
There are issues with running different versions of clang-format on
these generated files causing conflicts (see #2035).

In order to avoid making hardware engineers have the right version of
clang-format installed when making a hardware-only change to topgen, we
instead ignore clang-formatting on the C files that are autogenerated.

This should cut down on formatting issues introduced by hardware-only
changes, while still ensuring that the software tree is kept
well-formatted.

The `.clang-format` file introduced into the sw/autogen directory is
so that utilities like `git-clang-format` will not attempt to reformat
the files in that directory.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>