[sw] Avoid clang-format-ing Vendored Software
We should not be re-formatting vendored software, even accidentally, as
we want vendored software to be as close to the original as possible.

This patch adds `.clang-format` files to the `vendor` directories, which
disables formatting of any code contained in those directories.

This patch also removes all `.clang-format` files from vendored
repositories, as these are 'closer' to the vendored files so can
re-enable formatting for the vendored repositories.

This change also ensures that `.clang-format` files are not copied into
the documentation website.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>