[aes/dv] Adapt CSR test exclusions
This commit adapts and documents the CSR exclusions for the AES module.
For the TRIGGER and the STATUS register, the exclusions are now specified
on a field basis.

The STATUS register is updated based on writes to other registers. Upon a
reset, internal operations are triggered (reseeding LFSRs, clearing
registers) that temporarily change the IDLE field in the STATUS register.
The TRIGGER register is updated by the hardware and fields with non-zero
reset values are cleared to zero by the hardware automatically.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>