[util] Update Tool Dockerfile Package Dependencies
Commit #2082 showed that the Hardware Tool Dockerfile was out of date
against the general project apt dependencies (which are now specified in
a text file).

This commit updates the Dockerfile to use the official version of the
project dependencies. It also updates the toolchain version to match CI.
The aim here is that all the tools in this Docker image match CI, so
that it is easier to get started with the project.

I have left any package names that were not in apt-requirements.txt as
I'm not sure what these may or may not affect.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>