[sw, chip dv] Convert existing tests to testlib
The existing tests are converted to a testlib that implements
`test_main()` rather than implementing its own main function as a
standalone test. The C tests are moved from their standalone directories
to one level up into the `tests` directory. The standalone directories
are removed.

The tests are also enabled to run in the DV environment (all of them passing).

To support the above updates, a few cosmetic changes had to be made,
none of which affect the tests functionally. Hence, these are bundled as
a part of the same PR.

The CI script is also updated to reflect the new paths of the C tests.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>