[site] Add menu sidebar for docs
Add a menu to left side to indicate the current position in the
documentation tree.

Increases the site width in order to accommodate the new entry.
Remove menu and toc for narrower windows.

It only shows entries which are recognized by Hugo.
This does not include all the READMEs which are available.

Fixes lowrisc/opentitan#866 lowrisc/opentitan#1420

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>