[chip dv] Single instance of SW logger interface
The existing code had a unique SW logger interface per SW image running
on the CPU - the boot rom and the test SW.
Both instances are monitoring the same address location in the RAM for
new log messages.
Both are operational from the start of time until the end of the
simulation.
This is problematic because only one SW image can run on the CPU at a
time - CPU first executes the boot rom image and then jumps to the flash
to start executing the SW test image.
Additional logic would have had to be added to ensure that only one of
them is running based on which image the CPU is currently executing.

The easier way to deal with it is to just have a single instance of the
logger instead, which is what this PR does.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>