[dv/prim] add PRESENT testbench
This PR introduces the testbench for the PRESENT block cipher,
along with integration of the associated DPI-C model.

At a high level, the TB instantiates an array of PRESENT DUTs, each
parameterized with a different number of rounds.
The DPI-C model is used to generate expected encryption and decryption
outputs.
We use the open-source "Testvectors for PRESENT" model created
by Thomas Siebert from Ruhr-Universitat Bochum, which can be found at:
lightweightcrypto.org/implementations.php.

Signed-off-by: Udi Jonnalagadda <udij@google.com>