[prim_ram_1p] Move advanced functionality into wrapper
Reduce the single-port RAM primitive (prim_ram_1p) to its basic
functionality. This functionality doesn't include a "read data valid"
singal (rvalid), as all data is expected to be returned in the next
cycle.

In line with what was done in the dual-port RAM primitive (prim_ram_2p),
we now have a wrapper for such "advanced" functionality like the rvalid
signal in the form of `prim_ram_1p_adv`.

Currently, this wrapper only provides the same functionality as
prim_ram_1p did provide before with an interface mostly identical to the
one in prim_ram_2p_adv.

Adding advanced functionality which actually lives up to the name (e.g.
ECC and parity) can be done as needed in a follow-up.

For Ibex, we have to go slightly ugly: since Ibex uses prim_ram_1p, we
need to adjust its interfaces as well. We cannot do that in Ibex
upstream directly, as we need to get the prim_ram_1p changes (which are
in this commit) first into OpenTitan. To get out of this cyclic
dependency this commit adds a patch to the vendored Ibex. This patch is
then applied with the vendor tool to give subsequent vendor imports the
same patched interface.

Fixes #2181

Signed-off-by: Philipp Wagner <phw@lowrisc.org>