[ci] Check Vendoring is up to Date
This adds a step to the CI to check whether the vendored versions of
repositories have been fully committed. If it hasn't, running
`vendor.py` (without `--update`), will cause changes to the repo which
`git diff` will show.

The intention here is to catch when someone has not committed all the
changes that an update to a vendor file may have caused.

This CI step uses the `--check-patches` option to ensure any patches
apply cleanly.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>