[doc] Use the "colorful" code highlighting theme
I understand that this is a matter of preference, but I find the tango
style really strange in that it colors equal signs and other elements,
which doesn't work well with our other style. It also adds a background
to the text boxes, which we don't have for non-highlighted code blocks.

The "colorful" style has a similar level of "colors", but no background
color and generally works better with our page style (personal
preference).

Signed-off-by: Philipp Wagner <phw@lowrisc.org>