[dvsim] Fix error detection logic in Deploy.py
The code in set_status searches through the log for error messages. If
it finds one, it adds it to self.fail_msg. If it doesn't find an error
message, the code is supposed to check the process return code. If
that is nonzero, it considers the job to have failed and adds the end
of the log to the fail message.

Unfortunately the "did I see an error message" logic was wrong,
because it assumed self.fail_msg would be empty before running
set_status. This isn't true for CompileSim, at any rate, because the
constructor adds a header.

This patch corrects the logic by adding an explicit local
"seen_fail_pattern" variable.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>