[dvsim] Run tests in "interleaved" order
This is as suggested in issue #2370. The advantage is that when you
introduce a bug, you often break (almost) all seeds for one particular
test. Ideally, you'd want to see failures for that test early in a
run, rather than running a whole load of seeds through other tests
first.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>