[prim_arbiter_ppc] Add more FPV fairness checks
This moves the RoundRobin_A out of the generate loop such that it is
more generic and can be copied for other RR arbiters as well.

Further, this adds two new FPV checks:
1) NoStarvation_A: this property checks whether all requests will be
granted eventually if they do not drop out (i.e. remain asserted until
granted)
2) Fairness_A: this property checks for strict fairness, where strict
fairness is defined as: if N requests are asserted constantly and
downstream is ready, each of these requests will be granted exactly once
if we wait for N cycles.

Also, the input request stability assertion can now optionally be
disabled using the EnReqStabA parameter if needed. The arbiter works
correctly either way, but this assertion might be useful in bus
protocols where the requests must remain stable when back pressured.

Signed-off-by: Michael Schaffner <msf@google.com>