[dv/alert_handler] temp fix interrupt triggered same cycle as clr_intr
This is a temp fix for cases where issuing clear interrupt happened
right when another interrupt is raised. In this case design won't
trigger an interrupt, but because TB process request in negedge clk
cycle, so interrupt is set in TB.

This is a temp fix. I think eventually I should constrain the sequence
to avoid this corner case.

Signed-off-by: Cindy Chen <chencindy@google.com>