[ibex] Do not vendor in DV code
Configure the vendoring tooling to not include the `dv` folder inside
the Ibex repository.

This folder contains code and core files which are shadowing ones in
OpenTitan, breaking the simulation. We can think about vendoring the DV
code from Ibex back into OT as soon as we're sure that we will not
shadow more code. This work is tracked in lowrisc/ibex#985.

Fixes #2586

See #2513 and lowrisc/ibex#985 for more background.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>