[otbn] Add python simulator
This adds the first iteration of the python simulator for OTBN. Right
now it only has RV32I, but it lays the fundamental groundwork.

Before using it, you need to run `pip install -r
python-requirements.txt` again.

The base model is an external dependency (pypi package) plus a local
package with the program that performs the required operations. Later
the local package will have the overloads for the OTBN model.

Signed-off-by: Stefan Wallentowitz <stefan.wallentowitz@gi-de.com>