[otbn] Initial OTBN python simulator
This adds the first iteration of the python instruction set
simulator (ISS) for OTBN. Right now it only has RV32I, but it lays the
fundamental groundwork.

The base model is an external dependency ("riscv-model", which is
hosted on pypi and will be installed by pip as part of
python-requirements.txt). The OTBN model is currently a small wrapper
that loads up imem and dmem contents, instantiates the base model,
runs that, and dumps some results to files. This will expand as we add
support for the rest of the OTBN instruction set.

As well as including a basic ISS, this patch has a C++ wrapper that
can be used through DPI to run the ISS inside an RTL simulation. The
wrapper grabs imem and dmem contents from the simulation and dumps
them to files. It then finds the path to the Python simulator (based
on the location of the binary) and runs that. Finally it grabs the
contents of dmem after the run, together with the number of cycles
taken by the simulation (which it returns).

Use `fusesoc ... --OTBN_MODEL` to use this OTBN model instead of the
RTL implementation.

Signed-off-by: Stefan Wallentowitz <stefan.wallentowitz@gi-de.com>
Co-authored-by: Rupert Swarbrick <rswarbrick@lowrisc.org>
Co-authored-by: Philipp Wagner <phw@lowrisc.org>