[prim_subst_perm] Add simple substitution/permutation network
This adds a simple substitution / permutation network primitive that
works for any bitwidths. It is constructed in a similar way as the
PRESENT cipher, but without a keyschedule, and parameterized
permutation works for any bitwidth.

Note that this is NOT a secure crypto primitive. Its main purpose is to
enable low-latency diffusion in small or unaligned bitvectors for data
scrambling purposes.

Signed-off-by: Michael Schaffner <msf@opentitan.org>