[prim_pad_wrapper] Update pad wrapper
This updates the pad wrapper to better reflect the ASIC pad we are going
to use. Specifically, the following changes are made:

- an input buffer enable signal is added
- a variant parameter is added which allows us to differentiate between
different pad implementations for the same technology node
- a warl_o signal is added. this signal is constant and defines which
attributes are supported by a specific pad wrapper configuration.
- the WarlOnly parameter is added for use together with warl_o
- additional pad attributes such as schmitt trigger enable and slew rate
are added
- the pull-up/pull-down attributes are re-encoded such that no invalid
configuration can be set where both pull-up and pull-down are activated

Signed-off-by: Michael Schaffner <msf@opentitan.org>