[dvsim] Fix for #2686 - missing else
Testplan milestones (V1, V2, V3) are also regression targets (one can
run all V2 tests for instance by passing the switch `-i V2`). It could
be possible that the tests added to the testplan to one of these
milestones could actually be missing (yet to be developed) in the
simulation config hjson.

When that is true, invoking dvsim results in the error messages reported
in #2686. The fix for this is a missing else clause.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>