Allow verilated top-levels to do work after a simulation completes
Before this patch, it isn't possible for a main() function that calls
VerilatorSimCtrl::Exec() to tell whether the simulation actually
ran: a return value of zero might mean a successful run or it might
mean the user passed --help.

This patch moves the registration of a signal handler into
RunSimulation() (there's no point in registering the signal handler
before we start running anything) and makes ParseCommandArgs() public.
This means that Exec() is now a simple wrapper around a collection of
public methods which can be called separately.

For example, you could have something like this:

    ...
    bool exit_app = false;
    int ret_code = sim_ctrl.ParseCommandArgs(argc, argv, exit_app);
    if (exit_app) {
      return ret_code;
    }

    sim_ctrl.RunSimulation();

    if (!sim_ctrl.WasSimulationSuccessful()) {
      return 1;
    }

    do_some_postprocessing();
    return 0;

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>