[usbdev] Fixes for I/O modes and expand their tests
The differental mode reception and conversion was being done in the iomux
and potentially not passing all information to the usb_fs_rx module. Reworked
to pass the single ended and differential data to usb_fs_rx, recover them
both and use the configured mode to forward appropriately.

This simplifies the usbdev_iomux to only pinflip I/O and pass async
rx to receive logic.

As part of debug fixed verilator UNOPTFLAT complaint and added a link
state representing the link being active but never seen SOF because this
is where a flipped link gets stuck between FPGA board and Linux. (The
flipped pullup will make it look like a LS not FS link so the 1.5Mbit
chatter is enough to get rx PID erros but also keeps resetting the
link so host_lost never fires).

Update verilator top and usbdpi to manage all the I/Os. And add code
to the usbdpi to be able to simulate pinflip and differential as well
as the normal pin mode. Note that because there is no other signal it
uses the tx_mode_se signal to control use of differential mode in both
tx and rx directions (so it only works if both phy config bits are the
same). It uses the DN PULLUP being asserted to detect flip.

Update hello_usbdev.c to set the phy config from some defines.

Add to the pin file for the FPGA to allow a bitstream to be built to
fake pinflip by swapping the FPGA pins used for D+/D- and their pullups.

Tested with verilator that all four settings of PINFLIP and DIFFIO in
hello_usb.c get the same expected output.

Tested with FPGA that PINFLIP=0,1 both work (with appropriate bitstream)

Fixes 2598

Signed-off-by: Mark Hayter <mark.hayter@gmail.com>