[syn] Minor refactorings in parse-syn-report.py
This was triggered by a PR to remove master/slave terminology, which
touched the "master_key" arguments of the script's _match_strings and
_match_fp_number functions.

I think this version is a little simpler to follow, and gets rid of
the problematic terminology by getting rid of the arguments entirely.

  - The _match_strings function (which only had one call site) is now
    inlined

  - The _match_fp_number no longer operates in-place on a dict and
    instead returns the stuff it extracted, which is inserted into the
    "document dict" at the call site. I've also expanded the docstring
    somewhat.

There should be no functional change. The logic in _match_fp_number
could be simplified a bit if we allowed an empty list of results for a
key with no valid matches.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>