[sw, dif] add C/C++ guard to dif_warn_unused_result.h
This header is used included by other polyglot headers, and must
also be a polyglot header, as per
https://docs.opentitan.org/doc/rm/c_cpp_coding_style/, "Polyglot
headers":

"Moreover, all non-system #includes in a polyglot header must also be
of other polyglot headers. (In other words, all C system headers may
be assumed to be polyglot, even if they lack guards.)"

In practice this header is not likely to cause any issues, however, it
easier to add the guard than treat it as an exception.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>