[otbn] Correct wmask assertions in OTBN top-level
This fixes the assertion about imem_wmask_bus so that it only triggers
on a write (a partial wmask on a read is allowed, but ignored).

This patch also removes the imem_write_o port from otbn_core: the core
cannot write to IMEM. We keep the imem_write_core signal in otbn.sv,
because it makes things look a bit more uniform, but it's forced to
zero.

The patch completely removes imem_wmask_core from otbn.sv: it was just
assigned to and then ignored in otbn.sv, so we may as well get rid of
it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>