[tool/script] delete clean section in make files
As discusses in PR #2626, we are not using the "clean" function in
makefile. This PR removes the clean section in syn.mk, lint.mk, sim.mk

To make sure the directory is clean before running regression, dvsim has
a switch `-purge` that uses `os.remove` to delete the scratch_root.

Signed-off-by: Cindy Chen <chencindy@google.com>