[dv] flash ctrl DV, testplan updates
This adds the initial testplan for verifying the flash ctrl. It is
likely not fully complete at this point and may need an additional round
of vetting  / updates before opening it up for a review.

The DV plan is also updated with the correct rtestbench block diagram
and some explanation of the env components.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>