[otbn,model] Add basic OTBN model specifics
Add the variant, model and basic instruction support. Not all
instructions are implemented, but it can run the basic mutiplication
of sprint 2.

It also adds the assembler (otbn-asm) that is a minimal assembler that
generates programs for the model as binary or C array.

The C example runs a single multiplication on static operands. It
reads back the result and compares it to the expected result. You can
analyze the trace by inspecting the file /tmp/otbn_XXXXXX/trace, where
the actual path is printed every time the otbn is triggered in the
simulation output: "Execute OTBN in /tmp/otbn_XXXXXX"

Signed-off-by: Stefan Wallentowitz <stefan.wallentowitz@gi-de.com>