[aes] Add 2-share implementation of the cipher core for security hardening
This commit adds a first version of the first-order masked, 2-share
implementation of the AES cipher core. The masking can be enabled
via a design time parameter. If enabled, both the cipher datapath and
and key schedule are masked.

The generation of the masks is not part of this commit but future work.
At the moment, constant masks are used. This is of course not secure.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>