[mask rom] Initial Pseudo-Code for Mask ROM
This is the initial high-level overview of what the Mask ROM needs to
accomplish, and in what order. It is based on the (in-progress) Secure
Boot specification.

It has been written as C-like pseudo-code so that we can more easily
understand the flow of how the mask rom boot should work.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>