[mask_rom] Mask ROM Initial Commit
This is the initial framework which we will continue to develop the Mask
ROM upon. The Mask ROM is the first stage of the Secure Boot process. The
aim of the Mask ROM code is to have as few dependencies as possible, and
for all the code to be as simple as possible, so we re-implement some
structures in other parts of the `sw` tree, rather than pull them in
entirely.

This code is based on the implementations in `sw/device/boot_rom` and
`sw/device/exts/common`.

The Mask ROM currently does not do anything, and certainly cannot load
a flash image to execute yet. These are future additions that will be
made.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>