[reggen] Add bus port to `inter_signal_list`
Make bus_device, bus_host to be inter-module signal. reggen, when it
reads the comportable IP hjson, adds entries to inter_signal_list if bus
port exists in the IP hjson.

This commit itself breaks the top level stuff. `topgen` and `tlgen`
should be modified to support this.

This commit is related to Issue #3031

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>