[otbn] Fix .data VMA in OTBN linker script
I had thought that the "imem" and "dmem" regions in the MEMORY command
above would have separate counters, but that's not how it works at
all!

It seems that ld actually works by building a usage bitmap for the
memory space and if you don't specify an address for an output
section, it gets placed in the "first available address matching the
specified region". With the code before, the "first available address
in dmem" was... the first address after the end of imem. Not very
Harvard! :-)

This patch also changes the spacing in the AT> output section
attributes to be more conventional.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>