[regtool] Add `_` separator between multireg base name and index
This commit causes the regtool to insert a `_` to separate the base name
of the multireg from the register index. This improves readability when
dealing with multiple multiregs, especially if the base name ends on a
number.

For example a multireg named `KEY_SHARE1` now leads to registers
- `KEY_SHARE1_0`, ..., `KEY_SHARE1_7` instead of
- `KEY_SHARE10`, ..., `KEY_SHARE17`.

This is related to lowRISC/OpenTitan#3006.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>