[dvsim] Change cores-root to avoid conflicts with autogen'd core files
The `$SCRATCH_ROOT` defaults to `$PROJ_ROOT/scratch` the way we invoke
dvsim, if `$SCRATCH_ROOT` is not defined. This may cause core file
conflicts with autogen'd core files in that build scratch, since all
dvsim flows use `$PROJ_ROOT` as the `cores-root` path. This patch
changes the `cores-root` to `$PROJ_ROOT/hw` to avoid that from
happending.

Fix #3049

Signed-off-by: Michael Schaffner <msf@opentitan.org>