[sw] Vendor in llvm_clang_rt_profile
This change adds a vendor.hjson file and a patch to be able to build
the LLVM profiling runtime support library for OpenTitan. This library
relies heavily on file I/O, which we don't have, and supports multiple
platforms that we are not interested in. Thus, the hjson file excludes
all files that are irrelevant to OpenTitan while the patch removes the
headers that we don't need and adjusts the paths of several headers.

Signed-off-by: Alphan Ulusoy <alphan@google.com>