[dv/esc_agent] Align with design behavior for ping timeout
According to PR #3034, we decided to leave the issue as WONT FIX and
DV will align:

1. Since prim_esc_sender does not know ping_timeout, so in esc_monitor,
we will check the resp_p/n until it reaches complete state or
signal_int_fail state.

2. A minor style change: change `check_esc_resp` task from `if else
if` statements to `case` statement to improve the code readability.

3. In task `check_esc_resp` if it is a ping and a ping is interrupted
by real escalation, move the priority to the highest, and skip checking

Signed-off-by: Cindy Chen <chencindy@google.com>