[otbn] Clarify what happens on out-of-bounds or unaligned load/store
Allocate an error code for "bad data access" and refer to it in the
docs: this can totally change as we implement the design, but it's
much easier to say something and change it later.

This patch also slightly expands the general notes about what happens
on an error ("We stop and set an error flag": see _index.md). I think
this should be pretty uncontentious. Obviously, there are more details
to add in future!

The patch adds a check for top-of-memory to the pseudo-code that
describes how the bn.lid and bn.sid instructions work. Hopefully, this
pseudo-code will eventually be pulled from the simulator, but it can't
hurt to add that check for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>