[dv/alert_driver] semaphore fix
The semaphore is used to limit the access to use send alert, and make
sure each alert handshake are two clock cycle apart. But previous
implementation has scheduler issues as below:

```
fork
  thread 1: send_alert
  thread 2: reset
join_any
disable fork;
1) semaphore try_get
2) semaphore release ---> if other thread get it during line 1) and 2)
here, then it will release the semaphore from other thread.
```

The reason i implemented line 1) and 2) is to handle the logic when
reset is issue during send_alert, then the code is still able to release
semaphore. However, I did not count for the case when alert_send (thread 1) is
finished, and other thread is trying to acquire the semaphore.

The solution in this PR is to move the try_get and release to a separate
reset thread.

Signed-off-by: Cindy Chen <chencindy@google.com>