[dvsim] Do builds smartly
This is a small enhancement that saves compute resources by eliminating
identical builds.

It is possible for tests to specify unique builds via the `build_mode`
key in the HJson. Without that option, tests are run with the `default`
build mode, which is the vanilla build with no extra (specific) compile
time options. It is possible the the uniquely specific builds may differ
from the default only under certain conditions, such as coverage being
enabled. If coverage is not enabled, then it is wasteful to run two
completely identical builds, expecially in time/resource-sensitive
environments such as the private CI. This enhancement checks if newly
encountered builds are unique to an existing one in the list by
comparing the final `cmd` and the environment exports. If they are, then
the new one is discarded and tests belonging to the new build are
remapped to run with ths existing build.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>