[dvsim] Force an ordering when constructing command in Deploy object
A Deploy object has an effect by running some command. This command is
constructed as a string in Deploy.construct_cmd by iterating over
"mandatory command attributes". These are stored in a dict. We don't
really care about the exact ordering, but it does need to be
consistent (and not depend on Dict's internal hashing).

One way to do this is to make mandatory_cmd_attrs an OrderedDict, but
there are several definitions of mandatory_cmd_attrs, which all need
to be updated correctly. Since we don't care about the order, and only
care that it's consistent, it seems simpler just to iterate over the
attributes in sorted order.

Note that sorted(my_dict.items()) does the right thing (ordering by
key) because the default sort on tuples is lexicographic, so we'll
alphabetically sort the list of keys. Indeed, this has no duplicates,
so we'll never inspect the values.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>