[otbn] Add support for unconditional jumps in otbn-rig
For a tiny example, run:

  hw/ip/otbn/util/otbn-rig --asm-output gen.S --size 10 -o gen.json

and gen.S will contain:

  /* Section 0 (2 instructions) */
  .offset 0x0
  addi          x30, x0, 1269
  jal           x6, 3320

  /* Section 1 (4 instructions) */
  .offset 0x2d4
  lui           x29, 29447
  srl           x18, x30, x30
  slli          x17, x24, 5
  ecall

  /* Section 2 (2 instructions) */
  .offset 0xcfc
  andi          x24, x0, 4
  jal           x3, -2604

Note that the .offset directives in the generated assembly don't
actually work for generating different sections: that, together with
the rather ugly command line, will be fixed in a separate commit.

Finally, this commit changes the comments generated in the assembly
output from "Snippet" to "Section": the things being described might
come from several snippets. I think I just had the word snippet on the
mind when I wrote the code in the first place!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>