[sw, runtime] Introduce "DIF" like PMP library
This change introduces a low level library for PMP hardware
manipulation. It follows DIF format and design principles, however
PMP being a part of RISC-V specification and part of the Ibex core,
this library is not compliant with DIF guidlines (must use inline
assembly to access Ibex CSRs, ...).

Otherwise this library matches the DIF, in that it is meant to be
used by higher level code. Primarily Mask ROM Lockdown module, and
post Mask ROM configuration stages, such as ROM_EXT.

!!!NOTE!!!

This change has contains outstanding actions and review comments,
which will be addressed in the follow-up changes. Please refer to

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>