[otbn] Split up JSON dumping/loading in otbn-rig
This is a bit verbose, because we're properly parsing JSON data back
to ProgInsn objects, so have to do various sanity checks on the way to
make sure all the types match up.

The core of the patch is in the otbn-rig script, where we change the
command line to have two subcommands: 'gen' and 'asm'. Running 'gen'
will generate some JSON that describes the snippets in a random
instruction stream. Feeding that JSON to 'asm' generates assembly code
that can be assembled into a binary for testing.

This isn't quite so handy to use if you "just want to generate a darn
asm file!", but it's the shape we're going to use in the future and
running things in two stages like this will spot any silly mistakes in
the JSON parsing.

Here's an example of how to run it to generate a small example:

    hw/ip/otbn/util/otbn-rig gen --size 10 -o tst.json
    hw/ip/otbn/util/otbn-rig asm -o tst tst.json

This example can be assembled and linked with:

    hw/ip/otbn/util/otbn-as tst.S -o tst.o
    hw/ip/otbn/util/otbn-ld -o tst -T tst.ld tst.o

If you just want to see some assembly, the 'gen' phase will print its
JSON to stdout if not given an '-o' argument, the 'asm' phase will
read from stdin if not given a JSON file, and will print just the
assembly to stdout. So you can run:

    hw/ip/otbn/util/otbn-rig gen --size 10 | \
      hw/ip/otbn/util/otbn-rig asm

Of course, this doesn't generate a linker script, so this is only
useful for quick checks. Chaining like this does:

    hw/ip/otbn/util/otbn-rig gen --size 10 | \
      hw/ip/otbn/util/otbn-rig asm -o tst

(but you no longer get to keep the JSON intermediate file).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>