[otbn] Avoid getting stuck in otbn-rig
This avoids painting ourselves into a corner with straight-line
instructions. Once we start generating jumps, there's a possibility of
ending up with e.g. a current PC of 100 and some other instructions
already defined at address 104. Then we *have* to generate a jump or
an ECALL, or we'll crash into stuff that's already defined (causing
infinite loops and other badness when you try to run the test!).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>