[sw] Conditionally Provide Some libc Symbols
`memory.h` provides some symbols which are also provided by libc. When
these are used by DIFs, we would prefer to use the host's libc
implementations rather than our own.

This approach is used so that DIFs can depend on `memory.h`, but also be
built for host-side software.

OpenTitan does not access volatile memory using the `memory.h` functions
-- volatile memory must be accessed via the functions in `mmio.h` -- so
this should not affect the functionality of any tests.

Closes #1806.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>