[otbn] add code snippet for 384 bit modular multiplication
This is a generic implementation of modular multiplication based
on the Barrett reduction algorithm without requirements on the form
of the modulus, only the range of the modulus is limited. For
special formed moduli (e.g. generilzed Mersenne primes in p384)
we might want to provide an optimized version.
The implementation mainly follows the description of the algorithm
in the "Handbook of Applied Cryptography" (Algorithm 14.42) and
re-uses its terminology. Some changes have been made to improve
register and multiplier efficiency.

This is related to lowRISC#3192.

Signed-off-by: Felix Miller <felix.miller@gi-de.com>