[sw/bitfield] Provide byte inspection builtins
These builtins are needed to implement parts of some DIFs. They are
implemented in a way to be compatible with the builtins style guide
changes.

In particular, this avoids hand-rolling these functions in each DIF, as
`bitfield.h` is a file that DIFs can use.

Providing these builtins was suggested by Silvestrs.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>