[otbn] Print error messages on failure in otbn-objdump
subprocess.run(...., capture_output=True) captures both stdout and
stderr for the child process. We need to capture stdout because we
want to alter it if all went well. But we need to make sure that
everything makes it to the console if something explodes and I'd
forgotten the captured stderr.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>