[top] Select Ibex register file implementation based on target
This commit propagates the register file selection parameter of
rv_core_ibex up to the top to allow selecting a variant based on the
target. The register file itself has a big impact on the overall area
of Ibex (~30%) and we have different implementations FPGA (LUTRAM),
minimum size on ASIC (Latches), and a generic FF-based implementation.

This commit modifies all FPGA tops to use the FPGA register file. We
keep using the FF-based generic implementation for ASIC targets and RTL
simulation (see lowRISC/OpenTitan#2650).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>