[CI] License checker: show full command line
If the license checker fails, it shows an error message in CI:

"Licence header check failed. Please check output of
util/lowrisc_misc-linters/licence-checker/licence-checker.py on
the noted failures.".

This commit makes the message more actionable for the user confronted
with this message by showing the full invocation, including the config
file.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>