[otbn] Be more explicit about "operand value" and "encoded value"
This commit tries to straighten out the two integer values that we
have for an operand. The "operand value" is the value that you see
referred to in the ISA. It might be signed and might have been shifted
up by some number of bits. If the operand is encoded relative to the
current PC, the "operand value" is the absolute address.

The "encoded value" is the non-negative number that should be assigned
to the bit field(s) in the encoding.

While straightening this out, I also changed EnumOperandType and
OptionOperandType so that they no longer derived from ImmOperandType.
While they *are* immediates, they behave completely differently for
our purposes. It's much cleaner to make them separate. If we need some
base class that contains the three in the future, we can do so, but
deriving from (the current contents of) ImmOperandType turns out not
to be very helpful.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>