[dv] Fix up prototype of dv_base_reg_block::build
This gained a "csr_excl" parameter in commit 85aca77, but we forgot to
update all deriving classes to have the same prototype. VCS seems to
allow us to be slapdash like this, but Xcelium shouted. We "fixed"
that by giving the base class implementation a default argument in
commit 1b1f2ec. This shut Xcelium up, but seems not to actually follow
the IEEE 1800-2017 standard (see section 8.20 "Virtual methods").

This patch removes the default value from csr_excl (since it didn't
help anyway, we may as well remove the magic) and fixes up the
deriving classes.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>