[dv/hmac] fix regression failure
Hmac regression found one corner cases regarding fifo_empty.
After hash_process is issued, if message input is not a multiple
of 4, there is a period of time sequence won't check fifo_empty
interrupt because it is hard for scb to align.
Now since prim_packer can hold one more data, the ignored period of time
needs to be extended.

Signed-off-by: Cindy Chen <chencindy@google.com>