[dv/gpio] fix regression failure
The daily regression GPIO failure on `gpio_sanity_no_pullup_pulldown`
was due to a round up corner case:
For the statement in GPIO scb line 82:
`((crnt_time - data_in_update_queue[$].eval_time) / clk_period) > 1`
In the corner case, the crnt_time is 434ns, the eval_time is 395ns,
and the clk period is 20ns. So (434-395)/20 = 39/20 is 1.
However, the eval_time should be: 394499ps which should round down to
394ns instead of 395ns. I think it might because of async rst_n caused a
little bit delay.
So the solution I propose in this PR is to: change the `clk_period` from
`int` to `real` and allow rounding for the division value. In this
corner case, 1.95 will be round up to 2.

Signed-off-by: Cindy Chen <chencindy@google.com>