changes requested by review
Software side fixes (mostly style) from review. One question
was using define set the PHY state, changed to set the gpios
allowing a script to run verilator sims with all four versions

Hardware side changes to restore the input synchronizer in iomux
(this is the correct place because it wants to be at the pin, although
it makes it a bit hidden from the fs_nb_pe code). Simulation shows
that this has the expected result of moving things a clock time and
occasionally moving a transaction by a bit time. Removed the manual
pin changes for testing pin flip mode and added a flip in the top
wrapper for nexysvideo, thus the same FPGA image can be used for both
flips.

Will squash these (or split changes) later