[util] Rejig how we load hjson configurations for dvsim.py
The main new feature from this patch is that a dvsim configuration
should now set a "dvsim_flow" value. This value is used to decide
which subclass of FlowCfg to construct (in CfgFactory.py).

There are two upsides to this:

  1. You can now run a lint or synthesis run without specifying the
     tool on the command line (before, the code made a SimCfg unless
     it recognised the tool).

  2. If you run dvsim.py on some other random hjson file, you get a
     somewhat helpful error message. Before, you'd get something
     cryptic about expanding verbosity flags.

There is also a downside:

  1. Every configuration needs to specify dvsim_flow. In practice,
     this isn't so bad, because this can be done in the included
     common_*_cfg.hjson.

Note that "every configuration" here includes primary configurations.
This is kind of silly, because a "primary configuration" is really
just a list of other things to run. In later patches, we can split
these out into their own type, which should clean up quite a lot of
the code, and get rid of this requirement. However, you can't do that
splitting without the change in this patch (I tried!), so I've done
this patch first.

To make sense of how this all works:

  - dvsim.py calls CfgFactory.make_config

  - This uses CfgJson.load_hjson to load an hjson file and everything
    it includes.

  - After loading the file, make_config looks at dvsim_flow (which
    must have a value) to decide which subclass of FlowCfg to
    instantiate.

  - The constructor for FlowCfg gets passed hjson_data. It sets up
    a whole list of attributes, then calls self._merge_hjson to merge
    the data from hjson_data into itself. It then calls self._expand
    to expand all the wildcards. Subclasses can hook in to these two
    methods if they need things to happen at specific times.

The only slight complication is from primary configs: configurations
that have a list of children to be loaded and run. These need to load
up some new hjson files. They can do so by calling back to the
factory (passed in as an argument to avoid circular dependencies).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>