[OTBN/DIF] Cleanup: Name enum type and values consistently
The error enum type is named `dif_otbn_err_code_t`, use the same prefix
also for the enum constants, also matching the use of "err" over "error"
in the hardware register name.

Also use hexadecimal numbers in the case statement to match the base
used in the hardware documentation.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>