[otbn] Add a stepped version of otbnsim
This communicates using stdin/stdout, reading commands and writing the
results. The idea is that it can be used by the RTL simulation to run
the ISS in lock-step. We'll probably change the "trace format" too,
but this should do for a start.

For now, a run looks something like this:

  $ hw/ip/otbn/dv/otbnsim/stepped.py
  load_i /tmp/otbn_4UemTt/imem
  LOAD_I '/tmp/otbn_4UemTt/imem'
  .
  load_d /tmp/otbn_4UemTt/dmem
  LOAD_D '/tmp/otbn_4UemTt/dmem'
  .
  jump 0
  JUMP 0x000000
  .
  step
  EXEC 0x000000
    x4 = 00000000
  .
  step
  EXEC 0x000004
    w0 = cc3f07bf17f7f2f2603a253cac0e011bae35732e201dddcfd3eb86ebb986bf4c
  .
  step
  EXEC 0x000008
    x4 = 00000001
  .

The un-capitalised commands were on stdin (echoed by my console). The
results on stdout are a "header", then some other results and finally
a line with just a ".".

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>