[otbn] fix specification of bn.subm instruction
With the `bn.addm` instruction the result is expected to
be in the range of `0` to `MOD-1` even if the non-modular
result is only in the range of `0` to `2*MOD-1`. For this,
in case the result is greater or equal `MOD` it needs to be
corrected by subtracting `MOD` once.
So far we had the same defintion for `bn.subm` which is wrong.
Here, the non-modular result is tolerable in the range
of `-MOD+1` to `MOD-1`. In case it is smaller than 0 it
needs to be corrected by adding the result to `MOD` such that
the final result is also between `0` and `MOD-1`.

Signed-off-by: Felix Miller <felix.miller@gi-de.com>