[hmac/fix] corner case for error sequences
While modifying hmac scb I found a corner case where in hmac error
sequence.
In sanity sequence, if hmac msg stream in msg is partial (length is not
a multiple of 4), there is one clock cycle that is wavied to check
fifo_empty interrupt. So in sanity sequence I used `msg_q % 4` to gate it.
But in hmac_error seq, we are not always stream in `msg_q`, if we found
the sequence causes an error, we will resend some other random messages.
This could potentially stream in partial message. So this PR resolves
this issue by gating the error_sequence as well.

Signed-off-by: Cindy Chen <chencindy@google.com>