[docker] Ensure Containers use UTF-8
We've had issues in the past with trying to build documentation with
non-ASCII characters in the document comments.

This sets the locale to UTF-8, so that python chooses utf-8 by default
too. This ensures that parsing files in python with non-ascii characters
will succeed, as long as those characters are valid utf-8.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>