[ci] Add Task To Build Docker Containers
We currently do no checking or automated updating of the Dockerfiles
committed to the repository. This makes a step along that path, with a
task that will at least attempt to build the three containers.

In the future, we may turn on some kind of automated updating for
commits to the master branch, but that's a lot more dangerous than this.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>