[otbn] Add a model of externally visible register block to ISS
This is recovered from otbn.hjson, using reggen. We don't bother
tracking addresses for now: I suspect that name-based addressing will
probably work just fine, and might be a bit easier to understand.

Here's an example of running the stepping simulation when you hit an
ECALL instruction:

  EXEC 0x000080
    otbn.INTR_STATE |= 0x000001 (from HW) (now 0x000001)
    otbn.CMD &= ~ 0x000001 (from HW) (now 0x000000)
  .

Obviously, we'll probably finalize on something a little more
machine-readable for the update format in the end, but this should do
for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>