[kmac] Remove unused ports from sha3pad and more FPV
entropy interface in sha3pad is not used. Also block pushed indicator is
not used outside as `run_o` can be used as the indicator.

SHA3 padding logic assumes the incoming message is already XORed with
the entropy at upper level. And the padded bits such as prefix, function
pads don't need entropy as they are not security sensitive information.

Also, this commit revises FPV assertions to not have "unreacheable"
assertions and assumptions. Previously, the FPV couldn't reach the
Vector assertion to check the SHA3 result.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>