[reggen] Do not declare non-local parameters in *reg_top.sv
Local parameters are only defined inside *reg_top_pkg.sv. They are
used as constants inside 1) the register file and/or 2) the actual
module.

Non-local parameters are module but not register file parameters:
They are used inside the module but not in the register file. The
register file itself is not parameterized. For this reason, there is
no point in declaring non-local parameters as module parameters in
the register file top. No existing IP core was using this.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>