[otbn] Simplify verbose tracing in otbnsim
Rather than keeping a verbose property on OTBNModel, we move it to an
argument on the simulator step and run functions. Once we've done
this, we can also avoid having to create the model explicitly in the
scripts.

This also means we can cleanly print disassembly in the stepped
model (which was the main motivation!)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>