[doc] Remove review lines from non-final HW stages
The "OpenTitan Hardware Development Stages" documentation at
https://docs.opentitan.org/doc/project/development_stages/ distinguishes
between a signoff review to reach the last development stage in
hardware and verification (D3 and V3) from the review required to move
between the in-between stages (e.g. D0 to D1).

The final signoff review is done by the Technical Committee with
clearly assigned reviewers (and typically in a review meeting).

Other stage transitions are done through normal PR reviews.

Currently, the hardware checklist template doesn't distinguish between
these types of review, but adds a review date and reviewers field to all
stages. This PR removes this information from the checklist for all but
the final stage, under the following motivation:

* Clearly spell out reviewers for the final stage transition, and
  indicate that this review is different from other stage transition
  reviews.
* Ease the stage transitions for all other stages by avoiding to capture
  information twice: the date and the reviewers of a stage transition
  are recorded in the GitHub pull request and the commit itself which
  makes the transition; updating the text file after it has been
  reviewed to capture who reviewed it feels unnecessary.
* Remove the wording "signoff" from all non-final stages; the
  documentation sees "signoff review" as something that is only done for
  the final stage.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>