[otbn] Run the ISS in lock-step with the RTL in otbn_top_sim
And (very excitingly) they actually match for the smoke test! There's
not yet much observability for errors: all we do is check that the
simulations take the same amount of time and that DMEM matches when
the simulation finishes.

At the moment, error reporting is pretty rubbish: we just print an
unhappy message to stderr. Figuring out how to pipe these errors
around the internals of the SystemVerilog simulation is probably the
next significant job.

As a future patch, I intend to add register file dumping to the
simulator, so we can compare register data at the end of the
simulation (basically the same check as the golden register values
that are in the smoke test at the moment).

A more significant piece of work is cycle-by-cycle checks. The plan is
to add a tracing interface to the core itself, render that to a
string (in SystemVerilog) and then textually diff with the trace
coming out of the ISS on each instruction.

Obviously, that's all in the future: for now, we just check we can run
the darn thing :-)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>