[dif_plic] HW Review Complete
As part of the DIF S2 signoff, we ensure that the peripheral is being
used correctly by the DIF. This review happened today with Eunchan Kim,
Silvestrs Timofejevs, Garret Kelly, and myself.

This PR includes the documentation clarification around Claim/Complete,
as sketched out in that meeting, as well as an update to the PLIC
checklist.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>