[otbn] Add support for CSRs/WSRs to ISS
This gets rid of the SingleRegister dictionary that was defining MOD
and ACC and defines a new class (WSRFile) to hold the values. These
are stored internally as unsigned integers, but that's hidden: you
have to access them via functions like read_signed() or
write_unsigned().

We also move the flags inside the CSRs. I'm not completely convinced
this is the right way to do it: alternatively, you could have the
flags sitting inside the state and pass them (as well as the WSRs) to
the CSR functions. If we decide to do that, it should be reasonably
easy to change back. The flags code also gets moved into its own
file (to break a circular include dependency).

There's a certain amount of noise in insns.py from updating how we get
to the flags and how MOD/ACC work. Incidentally, I think the previous
code was probably wrong for MOD as it treated it as a signed number.
While touching this, I also got rid of some bogus assertions (which
only pass because 1 == True in Python: aargh!) and added explicit int
casts to go between bool and int objects.

Finally, this gets the CSRRS/CSRRW instructions working (and, of
course, ports the wide equivalents to the new framework).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>