[verilator] Move Verilator Stop Address out of RAM
The Verilator Stop Address was at `0x10008000` - but RAM ostensibly runs
from `0x10000000` to `0x10010000`, so when Mask ROM zeroed all of RAM
before executing, it was causing the verilator simulator to finish.

This change moves the stop address out of the extents of RAM, to
`0x30000000` (which is currently unassigned in the address space).

Signed-off-by: Sam Elliott <selliott@lowrisc.org>