[sw/device] Move Device-specific Addresses
This takes a slightly more principled approach to device-specific
addresses, as used by DV (we've already done the same for verilator).

Now test status reporting just looks at the potential address to write
into, rather than the exact device kind, and the same for the DV logging
which bypasses the UART. This should allow us to replicate this
functionality in more simulation environments, without having to update
lots of different parts of the codebase.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>