[verilator] Migrate Stop Address to Test Status
This test modifies the verilator toplevel to use the DV Test Status
reporting interface rather than the ad-hoc manner it was using to
terminate before. It does so using the same logic, which should be
easier to keep up-to-date. The VERILATOR_END_SIM_ADDR is renamed to
match its new functionality.

On the software side, this removes the Device Stop Address and uses only
the Device Test Status Address. Various functions and assembly that used
to use the stop address, now do not, including `abort` (which now spins
forever). This is to keep the test status code out of non-testing code.

This should make it harder to accidentally trigger verilator to stop.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>