[otbn] Add simple ISS testing infrastructure
After breaking things a couple of times and starting to get specific
on exactly how some instructions should work, it looks like a good
idea to add some "smoke tests" for the simulator.

This patch adds a couple (checking we implement 'add' and 'addi'
correctly) but the idea is that we'll flesh that out for more
instructions in later commits.

The tests are extremely simple: each has a single assembly file,
together with a list of expected register values. We check that
assembling and linking the assembly file and running the simulation
ends up with the same values as expected.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>