[fpga] Reduce the main system clock frequency to 10 MHz
This commit lowers the main system clock frequency on FPGA to 10 MHz.
Consequently, the peripheral clock frequency derived inside clkmgr
is lowered to 2.5 MHz, and the UART baudrate needs to be lowered to
115200 baud.

With this change, the FPGA build again meets the timing and thus the
timing check for the FPGA build during CI is re-enabled.

This resolves lowRISC/OpenTitan#2508.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>