[aes] Add parameter to allow forcing all-zero masks during SCA analysis
If the newly introduced SecAllowForcingMasks parameter is set to 1, the
FORCE_ZERO_MASKS bit in the control register can be set to 1 to force
all masks to constant 0. If at the same time, the mask share of the key
is 0 (KEY_SHARE1_0 - KEY_SHARE1_7), this effectively allows to disable
the masking at runtime which can be useful for SCA analysis.

If the parameter is not set, the corresponding bit in the control
register is tied to 0 and in addition has no effect.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>