[dv, verilator] Simulation memory prototype impl
This change set implements a small memory used for chip level
simulations. An empty space in the chip address map can be used for this
memory. It is primarily used for the embedded CPU in the design (Ibex)
to signal information to the testbench, for example indicate when the SW
is done and the simulation can exit.

Currently, a couple of addresses in our on-device SRAM are being used
for this purpose. With this patch, those are completely gone.

The simulation SRAM intercepts a TL interface to instantiate a 1:2
socket and the (adapter + ) SRAM. This must NOT exist during synthesis,
hence takes the following strategy:
- An interception point is created in `rv_core_ibex` - this is done by
disconnecting the output of the TL fifo and the module port. This
disconnection happens only if `VERILATOR` or `UVM` is defined.
- THis allows the user to instantiate the SRAM in the testbench and
connect its inputs and outptuts to the interception point in
`rv_core_ibex`.
- This enables us to preserve the original design hierarchy without
pulluting with simulation components (or adding forces which is not
supported in Verilator).

Please see sim_sram documentation for more details.

Once approved, the address locations for signaling the end of test and
logging will be switched over to use this address space.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>