[otbn] Use libcst to extract implementations for docs
This isn't the full solution: we are going to want to do some
prettifying to make this more "pseudo-codeish". For example, we
currently have lines like

  val1 = state.gprs.get_reg(self.grs1).read_unsigned()
  state.gprs.get_reg(self.grd).write_unsigned(result)

which should probably be something more like

  val1 <- gprs[self.grs1]
  gprs[self.grd] <- val1

or similar. However, it does get rid of the duplicated definitions
from the YAML files and, if ugly, the definitions in the spec are now
at least clear.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>