[otp_ctrl] First round of adding synchronizer stages
In particular, this adds a synchronizer and output flops for the pwrmgr
signals, and synchronizes the command and data signals to/from the OTP
macro using 4-phase req/ack handshakes. To that end, the OTP clock mux
and synchronizers, as well as the prim_otp instance are moved into a
separate submodule for clarity.

Signed-off-by: Michael Schaffner <msf@opentitan.org>