[otbn] Add checks for li expansion to ISS testsuite
This looks a little strange at first. "This is a test of the
assembler: why is it going in the ISS testsuite?", but the point is
that the easiest way to check whether LI has been expanded correctly
is to run the result and see whether we get the value we wanted. The
ISS testsuite is a really easy place to do that!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>