[otbn] Allow hex constants for big number operands
This part of the code is in charge of converting a string to a number.
If the string doesn't look like a number (we have something like
"bn.addi w1, w2, cabbage") then it punts and hopes that we can use the
binutils assembler to resolve the argument as a symbol.

However, we don't want to punt on "bn.addi w1, w2, 0x123" and Python's
int() function needs a base argument of 0 to do auto-detection of hex
constants.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>