[kmac] Revise CFG to be read-only in HW
This commit is to fix the CFG register behavior. It was external and
`hrw` type for HW to direct control the write permission.  After
utilizing REGWEN feature, it doesn't have to be write-able from HW view.
The permission was remained and caused the unknown value to kmac_en and
rest of the values, which creates an error on FPV.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>