[sw] Deny Most GNU Extensions
This includes fixups for empty struct initializers (which should just be
omitted), and an empty struct declaration (which was unused).

This change does not enable warnings about comma pasting in variadic
macros, because we use it in one place, and the problematic macro is
complex enough that a refactor might take some time. I have added it to
the list of warnings to enable in the future.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>