[otbn, dv] Write an initial DV plan for OTBN
This describes a testbench that doesn't exist yet, so there will
probably be changes. One obvious question is the number of agents
needed: for now, I've described just one called "otbn_agent", but I'd
be surprised if that's the final answer.

There is also an open question about how the testbench will get
instruction streams to load, but I think it probably makes the most
sense to try implementing that and document what we've built after the
fact.

Finally, neither a list of test sequences nor a list of coverage
points is defined: these will be addressed by a testplan and an
initial list of coverage points in future commits.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>