[otbn] Alter BN.MOV[R] implementation
BN.MOV[R] was implemented using an add by forcing one operand to 0. This
alters it to use the same data path as `BN.SEL` with the condition
forced to true. This reduces the logic needed to generate the operand B
input to the bignum ALU (only BN.MOV[R] needed the force to 0 option).

Signed-off-by: Greg Chadwick <gac@lowrisc.org>