[dvsim] Run cov_merge / cov_report as part of the main set of jobs
Now that dvsim has dependency support (where a job depends on zero or
more earlier jobs), rather than "job spawning" (where a job may spawn
some other jobs on completion), we can express the correct
dependencies for the cov_merge job. This should only run if all the
tests passed, so depends on all of them.

One nice side-effect is that the timer count shown on the console no
longer restarts when we get to coverage. The code is also much
simpler.

Finally, we no longer need to store print_legend as a Scheduler class
attribute (because we only actually run the scheduler once, so don't
need to avoid printing the legend multiple times).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>