[util] Add C Enum Support to sparse-fsm-encode.py
This adds support for generating C enums to the sparse encoder script.
We intend to use this when hardening the software we're writing in C,
including in particular the Mask ROM.

The script has not been improved much, as the intention is this is run
once for a specific enum, and the result is copied into a C file and
modified, rather than the script producing a perfect enum when run.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>