[sw] Initial Hardened Bool Type
When hardening software, the generic boolean type is not good enough, as
`true` and `false` are very similar in representation. This means that
it is reasonably easy for `true` to be changed into `false` with very
few bitflips.

This change adds a hardened boolean type, where the values have a much
greater hamming distance, which makes it far more unlikely that a truthy
value will be changed into a falsey value or vice-versa.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>