[kmac] Fix critical syntax errors.
This commit is to fix a few critical errors preventing the design from
compiling in the simulator. Mainly the KMAC has been checked in FPV not
in the simulator until now. So some parts of the design assumed `FPV_ON`
is set always.

Also, this fixes parts of the lint error such as typo, bit-width
mismatches.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>