[dif/otbn] Allow 32b access to DMEM
OTBN's data memory (DMEM) can be accessed 32b at a time, or 256b at a time.
When accessing the memory through the device bus interface, 32b accesses
are possible. Up until now we still disallowed 32b accesses throught the
DIF and only allowed 256b accesses on the grounds that DMEM internaly is
256b wide. However, that restriction is rather arbitrary, and requires
RMW-cycles when software (on Ibex) only wants to write a single 32b word
into DMEM. This happens for example as part of the RSA (`modexp.S`
calling convention) when setting pointers at the beginning of the
routine.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>