[sw, ROM_EXT] Extend the ROM_EXT start-up and print "Hello World!"
This change extends the `rom_ext_start.S` including the C runtime,
as well as adds a proof-of-concept C `rom_ext_boot` configuration
routine that at this point simply prints out "Hello World!".

Most of the `rom_ext_start.S` is identical with the Mask ROM
counterpart, however some of the configuration that Mask ROM does
such as SRAM scrambling is not necessary.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>