[doc] Add short-form contributors guide
This document is a short-form companion to the upcoming detailed
contributors guide. It gives an overview of how we work, how to create
bug reports, our Security Issues Process, and the usual process for
contributing code.

The contribution guide effort was kicked off by Pirmin and has since
seen suggestions, edits, and comments from a wide number of people
across the project.

Signed-off-by: Alex Bradbury <asb@lowrisc.org>