[memutil] Split out the non-verilator part of verilator_memutil
The point of this patch is to define DpiMemutil, a DPI-based interface
to backdoor load memories with the contents of VMem or ELF files, in a
way that can be used by non-Verilator simulations.

To do this, we split out everything but command line parsing and
SimCtrlExtension bits into a DpiMemutil class. The VerilatorMemutil
class is now a wrapper around a DpiMemutil, providing the
SimCtrlExtension interface.

This wrapping is done with a member (rather than multiple inheritance)
to allow users to subclass DpiMemutil and then wrap that. Doing that
with multiple inheritance, you end up with the usual diamond diagram
and bad things happen...

The split is cunningly designed to stay API compatible for C++ code
that uses VerilatorMemutil, but we don't keep API compatibility for
SystemVerilog code because we rename

   simutil_verilator_memload -> simutil_memload
   simutil_verilator_set_mem -> simutil_set_mem
   simutil_verilator_get_mem -> simutil_get_mem

One slight wart on the VerilatorMemutil class is the
RegisterMemoryArea forwarding functions. These are needed for API
compatibility, but we could get rid of them in future by changing the
callsites to use the GetUnderlying() accessor.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>