[otbn] Add stall on first cycle to ISS.
The RTL issues its first instruction fetch the cycle after the start
signal is raised. So actual instruction execution does begin until the
cycle after that. A stall is added to the first cycle of the ISS
execution to match this behaviour.

In addition the done signal from OTBN is registered. This is required to
ensure it syncs up with the ISS done signal with the new stall on first
cycle behaviour and makes more sense than adjusting the ISS to match the
previous RTL behaviour as it removes a feedthrough path from the RTL.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>