[dv/otp_ctrl] enable memory auto test
enable memory automation test for otp_ctrl. The test_access memory can
only be accessed when lc_dft_en is on. So this PR added a pin interface
for this port to be turned on while running automation tests.

Also PR #3890 update the test_access window to be size 16. This PR also
updates the parameter

Signed-off-by: Cindy Chen <chencindy@google.com>