[otbn] Make otbn_core_model compilable with VCS
VCS correctly complains if you compare a chandle with zero or assign
zero to it (this should use null instead). It looks like Verilator is
doing the wrong thing, so I've filed an upstream issue here: [1].

More confusingly, VCS also complains if you call a method on a string
parameter. I'm not sure whether that's implied by the spec, but
there's an easy solution here: I just want to know whether a string is
empty, so can compare with "" rather than checking that its length is
zero.

[1] https://github.com/verilator/verilator/issues/2602

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>