[util, cargo] Rename `build_tock.sh` into `invoke_cargo.sh`
This file is a workaround to several meson shortcomings.

* The absence of `cat` command, which potentially can be solved
  in a cleaner way with something like:
  ```
  files = run_command(
    'cat', files('toolchain.txt'),
  ).stdout().strip()
  ```

* Inability to set the environment variables for some backends, for
  example - ninja (if I undertsnad it correctly).

We are intending to use rust for other features apart from Tock, and
are likely to be limited by the things stated above. As the result, it
is a good idea to reuse this script for different `cargo` invocations,
until (if) we find a better solution.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>