[rom_ext] ROM_EXT Parser: Image Ranges
This change adds an accessor for the image ranges, which computes
useful offsets within the image:
- The Image Start
- The Start of the Signed Area
- The End of the Image (which is the End of the Signed Area)

The latter one of these is the complex issue, because it is calculated
from the image length, which is stored in the image and user-provided.
This code does no checking of the calculated value, a caller must check
this value has not overflowed, and is within the expected extents of the
slot.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>