[otbn] Fix building earl grey with OTBN model
When passing OTBN_MODEL, Earl Grey can be built with the OTBN model
instead of the RTL implementation. This sets DesignScope == "", which in
turn doesn't make the otbn_rf_peek() function available through DPI
as SV export.

To avoid a compile error in this case make otbn_rf_peek() a weak symbol,
and check for its availability at runtime.

Fixes #3929

Signed-off-by: Philipp Wagner <phw@lowrisc.org>