[dv/alert_handler] exclude alert_handler loc_alert_cause
This commit excludes alert_handler loc_alert_cause to avoid top-level
csr automation test failure. Because CPU clock is disabled in csr
automation test, this will cause escalation not responded to the
alert_handler and cause a loc_alert fail.
This is a temp solution, to add back this register, we will hook up an
escalation receiver driver for csr automation tests.

Signed-off-by: Cindy Chen <chencindy@google.com>