[sw/otbn] Integrate OTBN code building into meson
Build OTBN code with the meson build system. The primary use case is the
integration of OTBN binaries into device (Ibex) binaries.

OTBN is a separate architecture based on RV32I. Since meson doesn't
support compiling (and combining) software for multiple targets in the
same run, we have to cheat by using an external script, which
effectively does what meson would do for "normal" device software: call
the assembler, linker, and binutils to produce the desired output files
from the sources.

However, since we know that OTBN is based on RV32I, we make at least
sure that we reuse the toolchain discovery results from meson and pass
them on to the OTBN toolchain. The OTBN toolchain itself is a "frontend"
to the GNU Assembler and the GNU linker, callable through otbn-as and
otbn-ld.

Fixes #2936

Signed-off-by: Philipp Wagner <phw@lowrisc.org>