[otbn] Add a model of the processor to the UVM testbench
This OTBN model module wraps the Python ISS and will be used as a
golden reference. The model is attached to the rest of the testbench
through an interface. The interface signals are tracked with a passive
agent, which sends transactions to the scoreboard.

This patch also has minor tweaks to the documentation, updating the
list of interfaces in the block diagram and correcting the agent
descriptions.

This is a partial implementation: we don't yet check timing on the
"done" signal, nor do we connect up the error checking.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>