[sw] Cleanup Device Image Linker Scripts
This change is intended to cleanup how we use our linker scripts and our
linker flags so that we end up with a more controlled set of device ELF
files.

The main way this is done is by extracting out all the informational
sections into a separate linker script fragment, that the other scripts
include.

This change also:
- Disables emitting unwind information.
- Adds some linker arguments to emit warnings if problematic sections
  are emitted (COMMON, and orphan sections).
- Defines entry points for all linker scripts (not that we ever use this
  information, but the ELF format requires it).
- Extracts some common device link arguments that are duplicated in the
  device tree.
- Lays some groundwork for investigating how we can use LLD for
  OpenTitan Device Builds.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>