[sw] Minor fix for crt code
Without this fix, the zero code gets to 0x1000_0010 (kTestConfig)
and just skips zero-ing the last entry.

This causes the test to fail later when the processor reads back the
config and gets unknown data.  This is not likely to be an issue
on verilator, as unknowns will probably come back as 0s.

It is not clear at the moment why this escaped DV CI private.

Signed-off-by: Timothy Chen <timothytim@google.com>