[sw] Degrade unused functions to warnings
Currently we abort the build if a function (in C) is unused. This is
quite inconvenient during development when one tries to comment out code
or return early to iteratively develop or debug a feature. To increase
developer productivity, this commit reduces the compiler message from an
error to a warning.

As we get closer to finalizing the software we can think about
re-enabling it, or using a different build configuration in CI that
overrides this setting.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>