[sw] Prevent use of COMMON Sections
Old versions of GCC default to using `-fcommon`, but this changed in
GCC version 10.0 to default to `-fno-common`, which Clang matches.

The DV environment was using an old GCC version with the wrong default,
which was causing issues when exposed to the changes in
lowrisc/opentitan#4014. This ensures we avoid using COMMON at all costs,
instead favouring using real, defined sections for our C code.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>