[sparse-fsm-encode] Disallow Complementary Encodings
In software, it is usual to check that a queried state is equal to the
expected state, and also that the complement of the queried state also
equals the complement of the expected state, like so:

  if (read_current_state() == VALID) {
    if (~read_current_state() == ~VALID) {
      // Only in this block should it be assumed that `current_state`
      // is actually `VALID`.
    }
  }

With this in mind, we want to avoid producing state encodings where one
state is the complement of any other state. This helps avoid confusing
states in this kind of check (programmer error), and also helps to avoid
security issues.

This check may make it difficult to produce state encodings with higher
inter-state hamming distances.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>