[dv] Don't spit out extra spurious errors if build phase failed
If the (top-down) build phase fails in a test class with a
DV_CHECK_FATAL or similar, we won't run build_phase for the
environment. However, we will then go through the check phase.

There's not much point in printing further error messages in the check
phase in this case (it just makes debugging more confusing), so allow
null FIFO objects, ignoring them.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>