[sw] Exporting spiflash frames for DV tests
DV tests need a binary file containing the frames from the DV binary
image so they can flash them onto the device over SPI. This change adds
a method for doing so, which is configured on a test-by-test basis.

This moves to using dictionaries, which aren't perfect, but are an
improvement for extra information.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>