[otbn] Fix otbnsim bitwise operations with signed immediates
In Python, XOR'ing a signed number gives a signed number (inheriting
semantics from C). We store unsigned values in our registers in
otbnsim, so we need to convert the signed immediate into a u32 before
doing the bitwise operation.

Before this change, we hit assertions when performing bitwise
operations with negative immediates. After the change, we'll see trace
lines like this:

    xori         x10, x0, -728          | [x10 = 0xfffffd28]

(incidentally, this is the right interpretation - the RISC-V ISA
specifies that these signed immediates are sign-extended).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>