Update to Verilator 4.102
Require Verilator 4.102 (over 4.040) to fix a bug we encounter in OTBN
tracing code.

Besides the typical set of bug fixes and improvements we don't expect
any visible change to existing testbenches.

Note that this version of Verilator now requires a C++11-capable
compiler on the host; this is not a new requirement for OpenTitan users,
as we had such a host compiler requirement for a while now and building,
e.g. spiflash, requires even newer C++ features.

Fixes #4170

Signed-off-by: Philipp Wagner <phw@lowrisc.org>