[otbn] Remove stale TODO in sim hjson
As Sri pointed out, there's a simple way to point tests at a different
directory of binaries. Remove the TODO and add a comment explaining
how to do so.

Also get rid of the intermediate "otbn_elf_dir" variable: it's not
needed otherwise, and we don't want anything else to use it (because
that wouldn't play well with the plusargs override).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>