[dvsim] Allow a "setup" phase in a Regression
This patch allows a "regression" mode in a sim_cfg to define a setup
key. The intial use-case is something like this:

    {
      name: smoke
      tests: ["otbn_smoke"]
      setup: ["rm -rf {sw_build_dir}",
              "{proj_root}/hw/ip/otbn/dv/uvm/gen-binaries.py {sw_build_dir}"]
    }

The idea is that the commands in "setup" will run in order as part of
the build phase.

I've put this into the regression object for now, because that's where
we need it. I'm not sure that it would work so well on e.g. a build or
run mode, because it becomes less easy to see when exactly the command
should run.

This patch is the absolute minimum change I could make to dvsim to add
the needed functionality: I intend to make follow-up PRs to tidy this
up, but need to get OTBN DV up and running pronto, so thought a
minimally invasive patch was the best way to start.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>