[otbn] Teach the "smoke" regression to build some binaries to use
The idea here is that we'll generate binaries as part of the build for
"regression"s (here, the name means a collection of tests that should
be run together).

We won't generate binaries as part of individual tests. This makes it
a bit easier technically in the dvsim backend (which would otherwise
need to work out which mode should apply to what build), and also kind
of makes sense: if you want to "run the tests", you use a regression
target. If you want to run one test (possibly on some binaries you've
made yourself), you can do that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>