[sw/rom_ext] Add manifest.h to dependency sources
`sw/device/lib/rom_exts.h` is used in both the header and the
implementation of the ROM_EXT Manifest parser. In this case, we need to
declare that the generated manifest.h is also a dependency of any
libraries that depend on the parser, not just of the parser itself, as
any dependencies will use the library's header, which uses manifest.h.

This solves an intermitent build error we've been seeing with
manifest.h.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>