[otbn] Teach instruction generator to generate BN.LID instructions
This took quite a bit of work because it meant generalising
KnownMem.pick_lsu_target. The problem is that we have a base address
and a range of possible offsets and we need to pick a target for the
load/store operation. However, both the target address and the offset
need to be aligned, possibly with different alignment requirements.
Time for some (simple) number theory!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>