[otbn] Allow mulqacc to set flags
The bn.mulqacc.wo and bn.mulqacc.so instructions now set M, L and Z
flags in a specified flag group. We needed a little bit of shuffling
in the encoding to make this fit.

The changes are as follows:

 - There's now a flag group field at bit 31 (the same as other
   instructions that take a flag group)

 - The "wb" field, which chooses between the three writeback modes,
   has moved from bits 31-30 to bits 30-29 and has been split into two
   parts: "so" and "wb0". "so" is just set for bn.mulqacc.so (just like
   the top bit of "wb"). If "so" is not set, then "wb0" selects
   between bn.mulqacc and bn.mulqacc.wo. If "so" is set, "wb0" selects
   the destination halfword.

This commit includes all ISS, RTL and specification changes required to
implement the above.

Fixes #2979

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>
Signed-off-by: Greg Chadwick <gac@lowrisc.org>