[dvsim] Replace `sw_test` with `sw_images`
This change is part 1/4 commit series that updates the way SW tests are
build for chip level. For the chip level SW tests to pass, all 4 commits
are needed.

This change reduces dependency on DVSim to process how SW images for
each tests are set and built. The goal is to be able to supply an
arbitrary list of images to build, rather than assuming a fixes set (ROM
and SW test). The need for this change arose from one more addition -
OTBN that requires an image to be built. The processing is now handled
by `sim.mk` and testbench code. All SW build steps (specific to DV) are
integrated into meson.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>