[otbn] Update error handling documentation
This was sitting in a stub "Design Details" section. I've renamed and
re-purposed that: if we have other general design details in the
future, I propose that we give them their own sections.

This description matches what I understand from the discussion on
issue #4208. Obviously, the patch doesn't update the RTL to match.
There shouldn't be any ISS changes required.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>