[topgen] Fall back to template hjson for initial validation if necessary
This commit changes topgen to use the template hjson for generated IPs
when performing the initial validation in case the auto-generated hjson is
not yet available. Previously, topgen would simply run into a chicken egg
problem and abort: Before calling the IP generators, it must validate the
top level for which all IP hjson files are needed.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>