[otbn] Hush warning messages when skipping non-ELF files
When I wrote this, I imagined that we'd have a directory that just
contained ELF binaries, so you might want a warning message if you'd
accidentally dropped in some "foo.alf" or similar.

However, the current plan is to build everything in place, so we'll
have a load of (random) source code and object files in the same
directory. All very well, but we don't want to spam the console.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>