[otbn] Implement fatal/recoverable alerts
There are currently two different types of alert. A "fatal" alert is
caused by things like ECC errors. In the future, it will also be
caused if we notice FSM glitches. This sends an alert, which latches
on until reset. The fatal alert has an associated FATAL_CAUSE
register, with one bit per cause.

A "recoverable" alert is sent just once. At the moment, we don't
actually have anything that generates one of these, but we probably
will in the future.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>