[present/dv] Test iterative chaining
This PR adds an iterative test configuration to the PRESENT testbench,
similar to how the cipher is used in the OTP controller.

For each `1 < N < NumRounds+1`, N chained PRESENT instances are
generated, each running a single physical round of the block cipher.

e.g.: N = 3 --> 3 PRESENT instances are generated, with the output of
                the first connected to the input of the second, and the
                output of the second connected to the input of the third.

PRESENT should still produce the same results when configured like this,
so we use the same reference model for comparison checking.

Signed-off-by: Udi Jonnalagadda <udij@google.com>
Signed-off-by: Michael Schaffner <msf@opentitan.org>