[dv/otp] Support SW partition write lock (write digest)
This PR has two main changes:
1. Update function `is_secret`. Current is_secret function returns 1
when it is NOT secret partition. This logic is not correct and this PR
fixed it.
2. Support lock SW partitions by using DAI to write digest. Because the
digest check is depend on SW (OTP won't check if the digest is correct),
this PR writes random data to SW partition digests.
Also there is an issue related to this PR #4363, so in this PR, scb only
checks digest_0 for sw partitions.

Signed-off-by: Cindy Chen <chencindy@google.com>