[dv/push-pull-agent] fix xcelium hanging issue
In push-pull agent monitor, we set `valid_txn` to 1 and to 0
sequentially without any timing delay, then in a separate thread we are
looking for a negedge for `valid_txn`.
It works fine for VCS, but for xcelim, it could not detect the negedge.
This PR fix it by using the design signal req/ack instead of valid_txn

Signed-off-by: Cindy Chen <chencindy@google.com>