[dv common] Enable DV macros in non-UVM components
This change enables most of the DV macros to be used in non-UVM
testbenches, by replacing calls to UVM logging macros with our
platform-agnostic ones.

It also updates the platform-agnostic logging macros (for the
UVM-enabled case) to invoke UVM report functions directly underneath,
scoped with `uvm_pkg::` so that uvm_pkp does not have to be manually
imported in the madule / interface where they are invoked.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>