[aes] Parameterize LFSR default seeds and perms
This commit changes how the LFSR default seeds and permutations for the
clearing and masking PRNG of AES are generated and propagated. In line
with other OpenTitan IPs, we now use the `gen-lfsr-seed.py` script. The
values are exposed as IP parameters.

We do not yet expose them to the top-level/`top_pkg.sv` as the type
declarations and dimensions are likely to change in the process of SCA
hardening. This allows to continue the SCA hardening work without
constantly updating the top level.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>