Change VerilatorSimCtrl::Exec to handle --help properly
If a verilator-based toplevel wants to do some post-simulation
checking or reporting, it needs to know whether VerilatorSimCtrl::Exec
actually ran the simulation or not.

Before this patch, it could do so by replicating the contents of the
Exec method. This patch changes the prototype of the Exec method so
that it reports whether the simulation actually ran or not.

Use it like this:

    auto pr = simctrl.Exec(argc, argv);
    int ret_code = pr.first;
    bool ran_simulation = pr.second;

    if (ret_code != 0 || !ran_simulation) {
      return ret_code;
    }

    // Run post-simulation checks
    ...

The include-order changes are to satisfy clang-format on the changed
files.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>