[aes] Add SVAs to check correct extraction of the masking PRNG output
The output masks and further pseudo-random data (PRD) for the individual
S-Boxes is extracted from the PRNG output. If this extraction is not done
correctly, multiple S-Boxes might use the same output masks or PRD which
could be exploited during SCA.

This commit adds a check for that by repacking the output masks and PRD
and comparing that with the original PRNG output during simulation.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>