[dv/lc_ctrl] enhance smoke test
These PR did a few enhancement for the lc_ctrl smoke test:
1. Move the constraints to next_valid_lc_state request to
`lc_ctrl_env_pkg.sv` so we can reuse that in scb. Thanks weicai for the
help and suggestions. We found for an associative array, index cannot be
randomized with constraint, so we spilt it into a separate randomize
function within the sequence.

2. Minor clean up and add some comments.

3. Use all three registers for lc input token instead of one.

Signed-off-by: Cindy Chen <chencindy@google.com>