[otbn] Allow concurrent runs of otbn_build.py on the same files
Because the Meson-based build system doesn't really allow multiple
toolchains, we use a hack to build OTBN code through "custom
commands". Doing this, in turn, means that we have to assemble and
link every assembly file that goes into an ELF in a single script.

If two different ELF files use the same assembly file, this means
we've got two different runs of the script that are both trying to
create the relevant object file at the same time, which causes
confusing errors.

This patch should fix the errors by assembling the objects to
temporary files and then (atomically) renaming them to put them in the
right place. Obviously, it would be nice to teach the build system to
do this properly, but that looks like it's a bigger job, so we're
applying the hacky fix for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>