[otbn] Fix logic for picking LSU target in rig
This wasn't quite right before (there was a bug that meant we
calculated k_min/k_max "backwards" from offset_range), and the
comments were a bit confusing, because I'd changed notations a few
times when writing it without tidying everything up properly.

This should be correct now and, possibly more importantly, it's rather
easier to follow the comments to convince yourself that it makes
sense!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>