[otbn] Don't jump to the instruction we're choosing in rig
The code that picks a jump target is careful to pick somewhere that
doesn't yet have any instruction data, but it didn't know not to pick
the address whose instruction we're currently defining.

We'll probably need something a bit cleverer in the future (when we
start generating branches and function calls), but this is a simple
fix for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>