[kmac] Revise Endianess in the Message FIFO
Problem:

    When the software writes partially and proceed the SHA3, the
    assertion inside sha3pad module fires. The assertion is to check
    when it is partial, the MSB of the strobe is always 0.

The strob and the data in partial completion was not expected value.
lower bytes of the message and lower bits of the strobe are 0 rather
than MSB as 0.

The data in between prim_packer and prim_fifo inside kmac_msgfifo module
was swapped always. The code was derived from HMAC.

It is correct to swap the data in HMAC. HMAC implements the hashing
logic in big-endian way. So the lower most bytes should be placed at the
higher most byte.

In KMAC, the message in SHA3 hashing module is converted into 3-D cube
form. The conversion logic assumes the data to be little-endian. Bit 0
is the actual bit 0 in the SHA3 specification. So, no need to convert
again.

This commit is to fix the issue above by assigning the data without
endian conversion.

The issue is reported by @udinator

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>