[ci] Remove pinning to setuptools < 50
We did have to pin setuptools due to Debian issues, which are now
resolved and we can remove the pinning. We did the same for Ibex a
while ago and it worked fine since, so we're reasonably confident that
this change will work in OpenTitan as well.

Also remove the explicit installation of six, it's not needed any more
on Ubuntu 18.04.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>