[vendor/ibex] Remove duplicate check tool requirements files
This also changes the fusesoc root path to {proj_root} in DVsim

Our builds only worked before because there was a duplicate
"check_tool_requirements.core" inside the vendored Ibex repository.

Now that the duplicate has been removed, we depend on the actual
"check_tool_requirements.core" which is located in the project root
directory.

Signed-off-by: Michael Schaffner <msf@opentitan.org>