[AON timer] Initial rtl commit
This commit adds the basic timers, registers and CDC for the AON wakeup
and watchdog timers.

There is one functional change to the HJSON, which is to add a wakeup
cause register. This register is written to zero by software to clear
the wakeup request (similar to pinmux wakeup).

CDC logic for register writes is currently not very sensible (a four
entry async fifo for every register). There are TODOs covering this and
it will be improved in subsequent updates.

Signed-off-by: Tom Roberts <tomroberts@lowrisc.org>