[dvsim] Simplify recursion in find_and_substitute_wildcards
I was hunting through this code when trying to track down a completely
unrelated bug, and realised that the double recursion in this function
is much more complicated than it needs to be.

If the first argument is a dictionary, this should behave almost
exactly the same as the previous version. The only slight change is
that this now "slices" the type of OrderedDicts. However, I believe we
only use OrderedDicts for their insertion ordering, which is
guaranteed for dicts by Python >= 3.6 (which we now use) anyway, so
there should be no change to dvsim's behaviour.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>