[otbn] Report errors properly from ISS to RTL
Most of the plumbing is pretty obvious. The only challenging bit is
getting multiple values back from the C++ model wrapper to the SV code
in otbn_core_model.sv. We were already passing 3 flags as the bottom 3
bits of a uint32_t. Now we pass a 16-bit error code as the top 3 bits.
If we ever need more than that many bits, we might have to do
something cleverer with out parameters, but I thought the quick and
dirty solution was probably best for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>