[ci, topgen] Add CI tests for English Breakfast top
For the beginning, we build:
- a couple of binaries,
- the Verilator simulation,
- the FPGA bitstream.
Executing binaries in simulation and running tests on the FPGA will follow.

In topgen, we need to skip fixing the C header guards if the generated C
files are not within the source tree. This is always the case if topgen is
invoked through FuseSoc in CI. However, in this case the C files are not
needed anyway.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>