[ci] Don't generate binaries for top_englishbreakfast
It seems we can't have binaries for different tops in $BIN_DIR.
Generating the bootrom binary for top_englishbreakfast seems to break
the execution of tests on the NexysVideo FPGA board (top_earlgrey).

This commit disables the generation of binaries for top_englishbreakfast.
At the moment, we can accept that the corresponding bitstream generated in
CI is not functional.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>