[kmac/dv] Update testplan wording
This commit updates the KMAC testplan wording.

Previously, it stated that we only manually squeeze more output data if
the desired output length is greater than 1600 bits.

This has been fixed to reflect that we only manually squeeze output data
if the desired output length is greater than the keccak block size (168
  or 136 bytes).

Signed-off-by: Udi Jonnalagadda <udij@google.com>