[otp_ctrl] add script to use present crypto C lib in python
This PR tries to reuse present crypto C lib in python by:
1. Write a crypto_present.c function and compile it input a
crypto_present.so library.
2. In otp_ctrl/util folder, add a crytpo_present.py script to load c
library and use it for 31 round of python calculation.
3. Add a Makefile for later reference when we try to generate `.so` lib
file on the fly

The crytpo_present.py script includes an encrypt and a decrypt function.
Both functions' inputs are 128 bits key and 64 bits data. Both outputs
are 64bit in hex form

Signed-off-by: Cindy Chen <chencindy@google.com>