[otbn] Fix a bug in RIG jump target selection
The pick_branch_targets function picks a "gap": a range of addresses
from which we can choose a target. Once it has done so, gap_vma is the
start address of the range and gap_len is its length (a positive
number of bytes). If there is only one valid address in the gap,
gap_len will be 1.

To choose the address within the gap we calculate max_insn_off, a
"maximum offset for an instruction" (in units of instruction size).
This calculation wasn't quite right, and could give a negative value
when gap_len was 1.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>