[lc_ctrl/otp_ctrl] Update mmap and state generation scripts
This moves the OTP scrambling constant generation (for keys and
digests) into the memory map generation script instead of using
the topgen RndCnt mechanism.

Likewise, the generation of the RAW unlock token is moved into
the life cycle state generation script.

These changes have a couple of advantages later on:
- The scrambling keys are conveniently available in the OTP memory map
configuration, which eases generation of the OTP preload image later on.
- The RAW unlock token does not have to be "statically" hashed in RTL,
but can be hashed in Python as a preprocessing step before dumping it
into the template.
- It will be easier to replace the token hashing algo with KMAC later
on.

Note that we do not expect to have multiple instances of the OTP or life
cycle controllers in our top-level, hence moving away from the RndCnst
mechanism (which uses instantiation paramters) is safe.

Signed-off-by: Michael Schaffner <msf@opentitan.org>