[reggen, dv] Add ability to set DV register prefix
This came out of a discussion with Nuvoton to add support for generating
UVM reg models that are extended not from `dv_base_reg*` classes
but from custom classes that add Nuvoton-specific functionality on top of
`dv_base_reg*` classes. With this PR, this can be achieved by setting
the switch `--dv-base-class-prefix foo`. This approach assumes that the
following exist, given the arg value of `foo`:

- A fusesoc core with the name `lowrisc:dv:foo_reg` that sources the
following files at minimum:
  - `foo_reg_pkg.sv`: includes the `foo` reg classes below.
  - `foo_reg.sv`: register class
  - `foo_reg_field.sv`: field class abstraction
  - `foo_reg_block.sv`: register block class abstraction
  - `foo_mem.sv`: memory abstraction
- These are required to derive from their corresponding `dv_base_reg*`
classes.
- If any of these specialized abstractions is not needed, they need to
be typedef'ed to the `dv_baes_reg*` one in the pkg.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>