[aes] Simplify clearing functionality
This commit simplifies the clearing functionality by reducing the number of
trigger bits from 4 to 2:
- One bit for clearing key, IV, and input data registers, and
- one bit for clearing output data registers.

This way, software can already trigger the clearing of the key, IV, and
input data registers while the last block operation is still running. After
the operation is finished, the output registers can be read in parallel to
the clearing operation. Having more, separate clearing triggers is not
useful but makes design verification a lot harder as it adds many
difficult corner cases.

This resolves lowRISC/OpenTitan#4653.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>