[otbn] Tidy up how we pass err_code from ISS to SystemVerilog
Rather than packing the error code into the top bits of status, pass
it out with an explicit output argument.

The only "clever" bit is in otbn_core_model, where raw_err_code_d and
raw_err_code_q update at the same time, but with a blocking and
non-blocking assignment, respectively. If there is just one
"raw_err_code" value, Verilator complains that the inout argument of
otbn_model_step acts as a blocking assignment to the value, but the
initialisation on reset acts as a non-blocking assignment to the same
variable.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>