[doc] Work around delayed math rendering
With MathJax 2 math rendering is delayed by a couple of seconds in many
page loads (I wasn't able to determine what exactly triggered the
behavior). The console shows

```
Uncaught DOMException: Node.removeChild: The node to be removed is not a child of this node MathJax.js:19
    c https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/MathJax.js?config=TeX-AMS-MML_HTMLorMML:19
```

(This error message is from Firefox, but Chrome shows the same
behavior and a similar error message.)

I wasn't able to determine the root cause for this behavior, but
switching to MathJax 3 (a complete rewrite) fixes the problem, so I went
with that.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>