[sw, pmp] Add new enhanced PMP C API
Add an API to create, configure and commit entries to the PMP
control registers. Updates are performed using 'transactions'.
Transactions must be initialized before use with the number of
regions that are to be configured. This will then be checked
when the transaction is finalized and the CSRs are updated.

Unlike the existing PMP API we specify all region types the same
way using a struct containing `start` and `end` addresses. This
makes it easy to be explicit about the region boundaries as shown
in the example below.

The new EPMP API assumes that most `mseccfg` updates will occur in
assembly and that the API will only be used to disable the Rule
Locking Bypass (`msecffg.RLB`) feature. In particular this API
assumes that Machine Mode Lockdown (`mseccfg.MML`) will be
disabled (though that can be verified by a call to `epmp_check`.)

The API has been designed so that Machine Mode Lockdown support
can be added at a later date if required by adding an additional
set of `Lockdown` permission enum values. It would make the API
clunkier though.

Example (configure and set in a single transaction):

  ```
  // Initialize the library with the current state of the PMP
  // CSRs.
  res0 = epmp_init();

  // Enable Rule Locking Bypass (RLB).
  res1  = epmp_set_rule_locking_bypass(kEpmpRlbToggleEnabled);

  // Update 4 PMP entries in a single transaction.
  res2 = epmp_transaction_start(kEpmpEntryCount4);
  res3 = epmp_entry_configure_na4(kEpmpEntry0,
             (epmp_region_t){ .start = 0x0, .end = 0x4 },
             kEpmpPermLockedReadWrite);
  res4 = epmp_entry_configure_napot(kEpmpEntry1,
             (epmp_region_t){ .start = 0x40, .end = 0x48 },
             kEpmpPermLockedReadExecute);
  // pmpaddr2 used as base address for entry 3
  res5 = epmp_entry_configure_tor(kEpmpEntry3,
             (epmp_region_t){ .start = 0x100, .end = 0x200 },
             kEpmpPermLockedReadWrite);
  res6 = epmp_entry_configure_tor(kEpmpEntry4,
             (epmp_region_t){ .start = 0x200, .end = 0x300 },
             kEpmpPermLockedReadOnly);
  res7 = epmp_transaction_end(kEpmpEntryCount4); // update CSRs

  ```

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>