[aes] Align alert naming and signaling with latest guidance
This commit aligns the naming and signaling of alerts inside AES with the
latest guidance in the comportability spec. In particular, the alerts
are prefixed with either `fatal` or `recov` to indicate the severeness, and
the alert name itself is indicative of the alert cause (update error in
Control Register or detection of a fault).

In addition, the recoverable alert is high for just one clock cycle to
signal a single alert event, and the fatal alert is additionally latched
inside the alert sender module.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>