[otbn] Define helper functions for snippet merging in RIG
It turns out that "try to merge the next snippet and make a
SeqSnippet otherwise" is a reasonably standard operation. Make
Snippet.merge do that instead.

The only time we need the "clever" function is when building up a big
list of snippets. Factor that logic out of SnippetGens into a static
merge_list function.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>