[otbn] Tidy up how we deal with fuel in RIG
We used to have quite a bit of complicated juggling, where every
generator made sure to leave enough fuel for a following ECALL
instruction. Stop doing that: it's way more complicated than it needs
to be. Generators still have to make sure they don't paint themselves
in a corner (by running out of actual space), but they can now exhaust
fuel without worrying.

This makes it much easier to do recursive generation. It's needed in
particular for sane loop generation, where we want to constrain
model.fuel dramatically when we're generating loop bodies.

While we're changing this, we also refactor the methods in SnippetGens
a little, to reflect what's going on a bit more obviously in the types.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>