[dv/hmac] reduce runtime for sha_vector test in smoke regression
Two purpose of this PR:
1. The sha_vector test reads a input txt file into sequences input and check if
the output is expected from the txt file. This process took 408s CPU
time. Due to the current private CI timeout issue, we will only randomly
run 2 sequences. This reduce the runtime to 20s.
Thanks @weicaiyang for the suggestions.
2. Move the `smoke_test` bit from rv_timer to `dv_base_env_cfg.sv`. In
this way, all other IPs needs to reduce the run time can use this
variable.

Signed-off-by: Cindy Chen <chencindy@google.com>