[otbn] rework parts of 384 bit Barrett code
This reworks some steps of the Barrett reduction
algorithm.

It removes the truncations in step 2 of HAC 14.42
uses the full size of 2 WREGs for the compuation
of r, r1 and r2 instead. This allows us to ommit step 3 of
HAC 14.42. Additionally we do not need to compute
the mask anymore that was used for this step.

This is related to #4548.

Signed-off-by: Felix Miller <felix.miller@gi-de.com>