[sw/sca] Add 'b' (batch encrypt) simple serial command to improve capture rate
This command is designed to maximize the capture rate for side-channel
attacks. Instead of expecting a plaintext and sending the resulting
ciphertext from and to the host for each encryption, this command
repeatedly encrypts random plaintexts that are generated on the device.
This eliminates the overhead of UART communication and significantly
improves the capture rate. The host must use the same PRNG to be able to
compute the plaintext and the ciphertext of each trace.

Signed-off-by: Alphan Ulusoy <alphan@google.com>