[dv/alert] update shadow_reg alert naming in DV
This PR finishes the first step report in Issue #4895.
Previously shadow_reg has a naming convention but current change does
not adpot that convention anymore. So in order to automatically check
shadow_register's update and storage error, it requires to user to
manaully link the related alerts in hjson file.

Signed-off-by: Cindy Chen <chencindy@google.com>