[otbn] Initial framework for configuring weights in RIG
This patch only allows the configuration to specify weights between
the different generators. It also assumes that the configuration comes
from the rig directory.

The plan is to do much more, and to allow something more general for
custom configurations (passing them by path too, and allowing
configurations to inherit from each other). This is the "MVP" version
to get the wiring right.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>