[util/design] Extend LC state scripts to also generate LC state types
This moves the entire life cycle state type definition into the life
cycle state specification Hjson.

The life cycle state generator script now translates this definition and
prints it into the SV package.

This is a preparatory step for adding the OTP preload image generator,
which needs to have access to the entire life cycle state encoding
definition to create the correct life cycle state vector to be put into
the image.

Signed-off-by: Michael Schaffner <msf@opentitan.org>