[aes] Improve DOM S-Box
First experiments on FPGA revealed that the previous design suffered from
substantial leakage resulting from mixing new input data and randomness
with previous intermediate results. Basically, the output of the DOM S-Box
is only valid in the 5th cycle. Before that, outputs shouldn't toggle and
internal nodes should only toggle when the corresponding stage is actually
evaluated, i.e. if the data produced during that cycle contributes to the
output.

To counter those issues, this commit adds pipeline registers to the DOM
multiplier primitives (optional according to paper) and inserts additional
register stages inside the GF(2^8) and GF(2^4) inverters to only present
new input data to the multipliers when they actually need it. Similarly,
fresh randomness is only presented to the multipliers when it is actually
needed. This requires local PRD buffering but at the same time allows to
reduce the width of the masking PRNG by a factor of 4. Also, the PRNG is
operated during 4 instead of just 1 cycle per round which is beneficial
in terms of SCA resistance.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>