[systemtest] Decrease spiflash timeout to 30s on FPGA
On FPGAs spiflash should be rather quick to load a software binary to
the device. If it takes longer then probably something went wrong. To
avoid waiting for a timeout for a very long time reduce it to 30s on
FPGAs, while keeping the original timeout for Verilator platforms (where
spiflash takes a very long time, and currently doesn't work at all).

Signed-off-by: Philipp Wagner <phw@lowrisc.org>