[topgen] Don't generate empty clock_reset_export entries
The check_clk_rst_export function was essentially the same thing as
setdefault(). Since topgen's merge.py is actually the only thing that
reads this key, keep the same defaulting behaviour but don't actually
generate extra entries.

If we don't do this, we have to add clock_reset_export to the optional
fields list for every type in topgen/validate.py to avoid a warning.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>