[topgen] Allow eflash objects to have name, clock_connections
In practice, they always have a name (we generate exactly one in
top_earlgrey.hjson, which is an element in the memory list with name
and type both "eflash").

They also always have a clock_connections field, which is generated in
amend_clocks (I wonder whether this should actually just overwrite
clock_srcs, but that's definitely for a different patch).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>