[ci] Use Nexys Video bootrom in CW305
The CW305 FPGA bitstream build needs a bootrom. None of the available
bootroms work, as we'd need to produce a specific one for this board and
hence he resulting bitstream is known to be broken. Either way, the
comments say we use the Nexys Video bootrom, so let's actually do
that.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>