[keymgr] remove one cycle drop at op_accept after keymgr_en is off
op_accept is low for one cycle when LC turns off KM
When it's low, kmac key is switched to sideload key. And this cycle may
expose the good sideload key. It's probably not a security issue but scb
checks any good key shouldn't be expose at this point.
It may be easy to update design so that we don't need to hack scb to
skip checking this cycle.

Signed-off-by: Weicai Yang <weicai@google.com>