[reggen] Define a Register type for use in reggen
The big change in this commit is to define a Register type (in
util/reggen/register.py), to represent registers as parsed from hjson.
Once you do that, it makes sense to define objects to represent things
inside a register like allowed accesses (in access.py), fields (in
field.py), bit ranges (in bits.py) and enum entries (in
enum_entry.py).

The original plan for this commit was to stop there but multireg
objects, which wrap a register, are a bit fiddly so we also define a
MultiRegister class (in multi_register.py).

The motivation for this change is that the code in validate.py now
actually parses its input to MultiRegister or Register objects.
Before, it checked the data was good and left it as a nested
dictionary. Downstream code then has to do silly things like re-parse
integers from strings. This rather violates the "parse, don't
validate" mantra.

**

The most obvious user-visible change from this commit that there
should be slightly better error messages BUT there will be fewer of
them. Rather than collect up error counts, as validate.py does, this
code raises a ValueError. The disadvantage of doing this is that we
don't get to generate multiple error messages. The advantage is that
the code gets much cleaner. This trade-off seems like a good one to
me: this is a configuration parser, not a C compiler!

A slightly less obvious change is that the parsing code is now more
opinionated. Before this commit, there were lots of configuration
errors that you could make that caused the parsing code to emit a
warning (that no-one notices...) and then to change the input to look
more sensible. Now, we just raise an error: in my opinion, there's no
real reason not to force a user to Do It Right.

To see how this commit changes things for code that now uses the
Register types, see the make_intr_alert_reg() function in validate.py.
This has got a little shorter, and no longer needs to reproduce the
logic to generate fields like "genswwraccess" and similar. Another
place to look is check_wen_regs() in the same file, where
multi-register handling has got much simpler.

**

There are various ways in which this is work in progress. Firstly,
there are other objects that are currently not wrapped (same-address
register lists, windows, skipto) and the top-level "wrapping object"
is still a dictionary. Assuming we're happy with this approach,
wrapping those should be easy enough.

Secondly, there are duplicate "multireg", "register" and "field"
classes. The code in gen_rtl.py constructs actual objects to represent
these. It was doing this by re-parsing the dicts. This commit
simplifies that code a bit, so it's now just re-shaping things like
Register objects. An obvious job for a follow-up commit is to get rid
of the types that are currently in data.py.

Finally, the wrapping of access enums in access.py is a bit ugly. I've
left this similar to the original for now, to avoid yet more changes
in the codebase.

**

As part of fixing everything up to match again, I also removed
genbasebits support, which seems to have lain unused since the initial
import.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>