[reggen] Allow paragraphs in regtool not-quite-markdown
It's obvious what to do with descriptions for fields and enum entries
are obvious (surround the paragraphs with <p></p> tags). The remaining
difficulty is for register descriptions.

For now, we keep the first paragraph of the register description where
it was (in bold). Subsequent paragraphs, if they exist, get placed in
a separate table row following the header. This means they don't get
centred or appear in bold, which I think is probably the right
behaviour.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>