[dv/sram] SRAM smoke test, bring to V1 status
This PR implements the smoke test for `sram_ctrl` as laid out in the
testplan.

Note that this PR requires #4794 to be merged first, CI will fail until
then.

This PR also incorporates the changes made by @weicaiyang in #5122 to
fix the RAL HDL hierarchy, and adds the SRAM configs to the nightly
regression.

Finally, this PR also updates the `sram_ctrl` to V1 status as all items
have been completed.

Signed-off-by: Udi Jonnalagadda <udij@google.com>