[reggen] Delete duplicate Field type in data.py
We've recently moved to representing registers, fields etc. in
classes, rather than dicts. This duplicated some work done in
gen_rtl.py and data.py and this commit removes the duplicated Field
type.

Most of the commit is reasonably mechanical. There's a bit of cleaning
up in reg_pkg.sv.tpl (because it actually made this change easier to
make). There's also a bit of "data modelling" work, defining accessors
for the HWAccess and SWAccess classes. I expect this to change a few
more times before they settle down.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>