[reggen] Delete duplicate Register/MultiRegister types in data.py
We've recently moved to representing registers, fields etc. in
classes, rather than dicts. This duplicated some work done in
gen_rtl.py and data.py and this commit removes the duplicated Register
and MultiRegister types.

There's a certain amount of tidying up in reg_pkg.sv.tpl, which I
think makes the implementation a bit easier to read (as much as is
possible for template code like this).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>