[packages] Fix CentOS package names
Some of the packages listed in `yum-requirements.txt` seem incorrect.
This PR fixes that. I unfortunately do not have access to a CentOS
machine to test these. They were pointed out by our IT. Our workflows
are in the process of switching over to a new dispatch system, so there
may be more fixes needed. In the meantime, if someone could test these,
that will be great.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>