[dvsim] Schedule jobs by dependency
The current Scheduler implementation splits jobs by targets - builds,
runs, cov etc. All jobs added to each target are scheduled ONLY after
all jobs in the previous target are complete. What that means for
example is, AES tests cannot run immediately after AES simulation
executable is built - they have to wait until ALL other builds are
complete. This results in the max parallelizable queue to starve, making
the process slightly inefficient.

The changes in the PR reverts back to the original approach, where jobs
are enqueued in the dependency order independently for each DUT
(HJson config). The dispatch queue size is still limited to the
`max_parallel` setting, but at least it is no longer starved. The
benefit of this may not be visible if max_parallel setting is << number
of jobs in a given target. At Google, we set `max_parallel` to 200,
which far exceeds the number of builds (~25). So regressions now finish
a lot quicker (~40% speedup).

Most of the changes revolve around segregating the items and maintaining
target specific status queues (since jobs in several targets my be running
in a given time). Apart from that, I moved the code around a bit to make
things more prominently visible. Other updates are as follows:

- TargetScheduler is removed
- Helper methods added to find a finished item's successors
- A failed item recursively cancels its successors (with exceptions)
- `_cancel_item()` recursively cancels all successors of a canceled item
- Likewise for `_kill_item()`
- `kill()` recursively kills / cancels item and all successors
- On SIGINT (with above bullet), `Scheduler::run()` returns the final
  results back rather than exiting immediately - this allows the full
  result table to be printed - which might be useful info to look at.
- Linter fixes

Signed-off-by: Srikrishna Iyer <sriyer@google.com>