[site] Generate complete DIF description
Currently the text produced for the DIF description stops at the first
subelement. This is apparent if the description contains a variable
enclosed in backticks. The text up to this variable is returned but not
the rest.

Instead of using the `text` member of the element, which only contains
the text up to the first subelement according to the ET documentation,
create the string manually.

Create the string for the XML element and process the result for usage
in HTML.

Another option to create the string would be to use `itertext` to get
the text of all subelements, but properties would be lost so no `code`
highlighting would be possible.

Fixes #5001

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>