[reggen] Define classes to represent parameters in a block
This is supposed to represent the thing at obj['params_list']. The
Params class is essentially just a list of the parameters. These
parameters are now split into 3 classes, inheriting from BaseParam.

LocalParam and Parameter are for "normal" parameters and localparams.
RandParameter is used for random netlist constants.

This change has a small impact on topgen.py: the code that pulls
parameters up to the top-level actually gets a bit cleaner, but it now
needs to convert back to dict explicitly. That shuffling, in turn,
slightly changes the ordering of fields in (the checked-in)
top_earlgrey.gen.hjson.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>