[ci] Remove --fork-point from git merge-base command
The --fork-point argument enables a complicated mode for git
merge-base, which tries to cope with situations where the branch from
which we forked (origin/master) has had rewrites.

This uses the reflog when searching for ancestors, which isn't
appropriate in CI (where we're working on fresh checkouts). It seems
that this fails occasionally, leading to rather mysterious CI errors.

Since we don't rewrite the master branch, just use vanilla merge-base.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>