[dvsim/syn] Update parsing script and area reporting
This updates the area reporting of the DVSIM parser script.
Submodules can now be expanded recursively by name, or by
setting a maximum expansion depth. This allows us for instance
to report two levels of hierarchy, where we only expand the
top_earlgrey module, but not the padring and ast modules.

Further, the area report is extended to also include logic and macro
percentages for each module.

Signed-off-by: Michael Schaffner <msf@google.com>