[rtl/kmac] fix `right_encode(output_len)` for KDF
this PR fixes the output length encoding that is padded by the
kmac_keymgr module.
The values should be exactly the same as the encoded keylength in
kmac_core.sv, except the byte that contains the byte-length of the
encoded value should be in the MSB position for right encoding instead
of in the LSB position for left encoding.

Signed-off-by: Udi Jonnalagadda <udij@google.com>