[dvsim] Fix --cov + --build|run-only bugs
The `SimCfg` prunes the number of builds if there is no uniqueness among
the existing build objects. If a simulation is rerun with --run-only
switch (the assumption here being there is no need to rebuild the
simulation executables since the source code did not change), with the
old code, the build pruning is not performed, causing the tests to
reference non-existent builds and fail. This commit fixes that issue.

If `--cov --build-only` is passed on the command line (with the intent
to pass `--cov --run-only` later), builds get prematurely pruned out
because we turn off the coverage switch. This prevents us from  doing
`--cov --build-only` and `--cov --run-only` as two separate steps.

Also, the coverage table and summary table end up throwing errors. This
change fixes that.

In addition, the `--cov-merge-previous` was also broken (appending Path
vars instead of str). The `odir_limiter` also has a minor fix such that
if `odir` does not exist, it still prunes and returns the old dirs.

The cov merge task also needed information from the builds, which do not
exist if `--run-only` switch is passed. The CovMerge code is updated to
seek the same information from runs instead.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>