[csrng] Select Canright S-Box implementation for AES cipher core
This commit changes the default selection for the S-Box implementation
inside the embedded AES cipher core from the LUT-based implementation the
unmasked Canright design. This is more suitable for ASIC implementations
because of the lower area footprint. The LUT-based implementation should
be used for FPGA targets only.

This change got already merged with lowRISC/OpenTitan#5215 but it got
accidentally reverted when rebasing lowRISC/OpenTitan#5195.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>