[sw, mask_rom] Initialise Pinmux
UART pins used to be dedicated, however with recent changes these
became multiplexible. This means that in order to get any uart output
in Mask ROM or ROM_EXT - the corresponding pins must be configured.

This change adds call to pinmux_init at the beginning of
mask_rom_boot. This ensures that the UART pins are configured in both
Mask ROM and ROM_EXT.

NOTE:
DIF Pinmux is work in progress, however it is expected that Mask ROM
will use the DIF directly to configure the pins. This also means that
sw_lib_pinmux will cease to exist.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>