[topgen] Restructure how we elaborate blocks into IP instances
There's no significant functional change, but this code should be a
bit simpler to understand and work with.

Because we now iterate over the instances in a different order (by
instance, rather than by block), the auto-generated random constants
are generated in a different order, so they change in the checked-in
top_earlgrey_rnd_cnst_pkg.sv.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>