[topgen] Stop pretending it's not OpenTitan-specific
Topgen was initially designed to be a rather generic toplevel generation
tool. That's not what it is today: it is heavily biased towards
generating OpenTitan toplevel designs -- which is exactly what we need
in OpenTitan. Instead of trying to keep things generic, this commit goes
the first step towards an "opinionated topgen".

As first step, all templates are stored not in the toplevel that is to
be generated, but next to topgen: other components in OpenTitan expect
files generated from these templates to be around exactly like that, so
we cannot simply modify them at will.

If necessary later we can add an override/extension mechanism of some
sort.

This commit tries to be rather minimal in the changes to topgen.py;
there's more to clean-up, but since we have at least three people
working on the same file at the moment I'm trying to keep the diff to a
minimum.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>