[otbn] Tweaks to OTBN stalling logic
This doesn't alter architecturaly observable behaviour. It makes more
uniform use of the stall signal (rather than doing special case things
for load/stores) and alters read commits to the register file so they
happen on the cycle the read value is used.

It also lays the groundwork for RND CSR reads being able to stall for
arbitrary periods of time.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>