[lint] Waive Verilator UNUSED warnings for packages
By default, recent versions of Verilator complain if not every
parameter that was defined in a package gets used somewhere. This
might be reasonable if there's a single top-level, but it doesn't work
at all if you have lots of different top-levels (several chip
top-levels, then each IP block), all of whom use different
combinations of the parameters.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>