[sw, rom_exts] Split ROM_EXT generation into binary and base library
This is a first step of switching to running on-device tests via
Silicon Creator (MaskROM + ROM_EXT) code.

- This change prepares for the next step - adding a "null" ROM_EXT
implementation, similar to what has been done in:
d2bbbe117d6c92be7715172401892d21910d73bb

- Splitting out the base library out of the binary generation,
allows to use it as a dependency in both production and test
binaries.

- As the entry point into smoketests is through `main()`, `dummy.c`
file has been added for the "production" binaries generation. It
is likely to become redundant when `rom_ext_null` is introduced.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>