[reggen] Promote FOO_SIZE to an int unsigned in *_reg_pkg.sv
With the previous version of the code, there was a problem if the last
window happened to finish on the same power of 2 boundary as the
block's address space. In that case, we ended up with something like

    parameter logic [9:0] BLOCK_FOO_SIZE = 10'd1024;

which is, of course, zero.

It turns out that the only code in the tree that uses these _SIZE
parameters is using them to compute int parameters (rather than
something involving the address width). As such, we can just promote
the parameter to an int unsigned, fixing various width mismatches as a
bonus!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>