[topgen] Remove alert_module and interrupt_module from top_*.hjson
These lists actually had code to populate them in topgen/merge.py's
amend_alert function, but this didn't ever run because it already has
a value. That value means that you need to remember to update it
whenever you add a new instance of an IP block.

The original motivation behind listing these explicitly was that the
top-level might not want to wire up all the alerts / interrupts from
the modules that it instantiates. We're not doing that at the moment
and, if we start doing that again, it's probably cleaner to explicitly
disable things rather than to have to add everything twice.

With this patch, we fix the code in amend_interrupt and amend_alert to
add everything. Now the only difference between that and the manual
list is the exact ordering of the bits in the signal. Since we access
these bits through auto-generated named constants anyway, we can
dispense with the manual step entirely.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>