[spi_device] Adding clock mux for DPSRAM
The DPSRAM requires to be asynchronous to support FlashMode and
PassThrough mode. This commit changes the DPSRAM type to
prim_ram_2p_async_adv and add clock mux for sram B port.

As SPI_CLK does not toggle when it is in idle, glitch-free clock mux
can't be put in. SW is required additional cautions when it changes the
sram clock.

Suggested programming sequence is added in the newly introduced register
field `sram_clk_en`

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>