[reggen] Make Params class a MutableMapping
A MutableMapping is effectively a dict and can be used in practically
the same way.

The special implementation of `get()` is gone now and replaced with the
default implementation, which has an additional, optional second
parameter to give a default value. The existing behavior remains
unchanged (return `None` if the key was not found).

Signed-off-by: Philipp Wagner <phw@lowrisc.org>