[dv/sram] implement pipelining test and update scb
this PR adds the SRAM pipelining test as laid out in the testplan.

in this test, we choose a random mem address and send a series of
back-to-back transactions to that address, to stress the internal
pipelining and forwarding logic.

this requires an overhaul of the scoreboard as the SRAM pipelining logic
means that while TL memory requests are handled in-order, the underlying
memory macro is updated in an out-of-order fashion, leading to several
tricky edge cases.

NOTE: this PR depends on #5530 to be merged first, as that contains a
      fix for an issue uncovered by this test.

Signed-off-by: Udi Jonnalagadda <udij@google.com>