[csrng] Remove duplicated parameter
The existing code defined the NHwApps parameter in csrng.hjson (which
flows into csrng_reg_pkg.sv) and also defined it as a parameter in the
csrng and csrng_core modules.

This causes a namespace conflict, violating the SystemVerilog spec. We
can't remove the parameter from the csrng module because it's used in
DV to allow the testbench to test different configurations from what
ends up in the chip. Fortunately, NHwApps doesn't alter the top-level
port list so isn't needed by the reggen/topgen machinery, meaning that
we can just remove it from the hjson file.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>