[dv] Get blocks with multiple device interfaces working with chip DV
The big change here is that we might now generate several *_ral_pkg.sv
files for a block: one per device interface. To support this,
gen_dv.py now has a loop over device interfaces (just like gen_rtl.py
and gen_fpv.py).

What's more, we have the same problem as gen_fpv where we need to list
the generated files in the core file properly. To get that right, we
pull core file generation out of ralgen.py and put it into gen_fpv,
where we know what files we've created.

Finally (and this is the biggest part of the patch), we split up the
top-level and block-level versions of uvm_reg.sv.tpl. The top-level
and block-level templates share functions from uvm_reg_base.sv.tpl,
and I think the result is a bit easier to follow. There's also a
liberal sprinking of big "##"-style comments to try to make the
template code a bit easier to follow.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>