[doc] Use relative links in Hjson-related shortcodes
The testplan, hwcfg, and registers shortcodes currently take a single
argument referring to the IP description file in Hjson format.

Before this commit, the Hjson file path was relative to $REPO_TOP.
After this commit, the path is relative to the file using the
shortcode. The previous behavior can be achieved by using absolute
paths, which are rooted in $REPO_TOP.

With this change users of the short codes do not need knowledge
about the overall file structure, making the IP directory "relocatable."

To avoid doing the same change three times in the testplan, hwcfg, and
registers shortcodes, this commit unifies them into a single shortcode.
(Unfortunately "inheritance"/"nesting" isn't really easy with Hugo
shortcodes.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>