[doc] Make URLs in prj.hjson files relative to the containing file
Paths to documentation content, such as the test plan, the design
specification, etc., were given in the ipname.prj.hjson file as relative
path, which was resolved relative to REPO_TOP. This makes IP blocks
non-relocateable.

With this PR the paths become relative to the file they are written in,
i.e. relative to the ipname.prj.hjson file. Absolute paths are resolved
absolute to REPO_TOP, effectively producing the previous behavior.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>