[util] Don't explode on unexpected format in fix_include_guard.py
The existing code assumed that there was an #ifndef line somewhere in
the file and exploded if not (calling None.group()). This patch
handles these cases more gracefully by having a notion of "unfixable"
files. It also changes the output format a bit to make the CI
wrapper's job a bit easier.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>