[dv/otp_ctrl] Fix killing push-pull agent seq
When lc_escalate_en is On, it will turn all state machines to ErrorSt
and all output to defaults. Thus if pull request is sent, if won't get
any responses.

Previous code handle it by directly kill the sequence, but this is not
recommended, as it could casue some error in sequencer. Current method
set the push-pull sequence to non-blocking mode when lc_escalate_en is
On. Then scb will ensure the request never gets reponses. And eventually
we will issue reset to clear the push-pull sequence.

Signed-off-by: Cindy Chen <chencindy@google.com>