[otbn] Track the operation of the idle_o interface
This is possible in the scoreboard, but it's probably easier to just
model with an FSM. Apart from anything else, getting the DONE signal
out with a proper monitor would be rather difficult.

Note that this commit slightly changes the behaviour of idle_o,
removing a combinatorial path from the TL input to idle_o. There
doesn't seem to be any need for it, and it makes the behaviour
slightly easier to describe.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>