[rom_ctrl] Instantiate scrambling primitives for rom_ctrl
Also update SW build scripts so that we install the scrambled binaries
to build-bin as well and update all the documentation examples to show
how to use the new filename.

Note that we don't update top_englishbreakfast to match: that doesn't
have a rom_ctrl, so shouldn't take a scrambled ROM image.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>