[flash_ctrl,lint] Fix width errors for parameters in flash_ctrl_pkg
Most of these changes are pretty mechanical, silencing Verilator
warnings by making casts explicit.

The only non-trivial change is in flash_ctrl_info_cfg.sv, where a
page_addr_t structure (normally used for addressing pages within a
data partition) is used to address pages in an info partition. This
means that the "CurPage" index needs expanding from InfoPageW to
PageW. In practice, Info partitions are no bigger than data
partitions, so this should be fine, but we also add a comment and an
ASSERT_INIT to make sure this holds.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>