[otbn] Teach RIG not to choose GPRs pointing at invalid WDRs
A previous version of the OTBN spec said that the instructions
BN.LID, BN.SID and BN.MOVR would just look at the bottom 5 bits of the
value of the GPR used to choose the WDRs they'd read or write. We're
changing the spec so that an invalid index causes an error; teach the
RIG about this to avoid generating lots of spurious streams.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>