[sva] csr assertion dependency update
This PR follows the CSR assertion updates. Since the new assertions do
not depend on RTL package anymore, we can remove this dependency and use
macro defines (if later on FPV and DV has different path)

The rest of the changes are to update this change to each IP and
template file.

Signed-off-by: Cindy Chen <chencindy@google.com>