Require a newer (still modified) version of FuseSoC
We have relied for a long time on a patched/customized version of
FuseSoC. This version has now been updated to use FuseSoC 1.12 as
baseline, with our patches on top of that.

Update the (reasonably meaningless) version number in our
python-requirements.txt to reflect that. Users still need to install
whatever the relevant Git branch has to offer.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>