[otbn] Wrap non-synthesizeable code with defines
Some dv code in otbn is included as part of the design core files
to do side by side comparisons.  These files need to be wrapped with
the SYNTHESIS flag to ensure they do not get parsed for synthesis.

Applying this fix to otbn_memutil_pkg.sv (the same fix is already applied
to various other model files).

Signed-off-by: Timothy Chen <timothytim@google.com>