[rom_ctrl] Fix backdoor path
We assume u_reg is the instance name of reg top in RAL model.
Need to find a better way to handle multiple regfile in one IP.
This still works for one CSR regfile and one mem regfile in a IP, as we
treat the mem backdoor separately. But can't handle if there are 2 CSR
regfiles in a IP.

Signed-off-by: Weicai Yang <weicai@google.com>