[reggen] Fix name for "u_reg" backdoor in RAL code
The generated RAL code needs to know where to find the generated
register in the design. Before this patch, we had it hardcoded as
"u_reg" but that isn't going to work if there are multiple register
blocks in the design(!).

Here, we use the same naming convention as the rest of the reggen
code: if a device interface has an explicit name, "foo", we assume
things relating to the interface get a "_foo" suffix.

Also fix the name of the one example in the tree (in rom_ctrl.sv) to
match this convention.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>