[dv] Split out the logic to connect a sequencer to a monitor
This connection is handy, but not very generic. In particular, it
implies some slightly unnecessary parameterisations for types in the
dv_lib base classes: not impossible, but a cost to users who don't
want the feature.

This patch splits that logic into a new "dv_seq_agent" core file. This
subclasses the parts of an agent that need to know about the special
behaviour (the agent itself, its configuration, the monitor and the
sequencer). Now, the code that uses this feature (csrng, i2c, spi and
push_pull) can use it explicitly, and the rest of the code doesn't see
it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>