[sw, dif_kmac] Add implementation of SHA-3 (blocking only)
Add support for SHA-3 operations to the KMAC DIF. These operations
do not support returning partial results (i.e. they block until the
operation is complete). Support for non-blocking operations will be
added in a future PR.

Note that some aspects of the implementation will need to be
modified when support for XOF operations, such as SHAKE, are added.
I have marked these with 'TODO: XOF:' for now.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>