[otp_ctrl] Reverse LC counter/state positions in memory map
In principle, this is not strictly necessary, since the life cycle
controller already sequences the counter update first, before repeating
the programming request with the updated life cycle state.

However, by reversing the positions of the life cycle state and counter
within the memory map we can make this a bit more secure since that
ensures that the counter words will always be touched first - no matter
the sequencing within the life cycle controller.

Signed-off-by: Michael Schaffner <msf@opentitan.org>