[flash_ctrl] Correct behavior when buffer not enabled.
When buffer is not enabled, the flash_phy_rd may erroneously return data
to back to back transactions when it is not supposed to.

This happens because even when buffers are not enabled, the read data is
written into the holding FIFO between read and descramble stages.  As a
result, the return path falsely thinks the data is available and returns
it.

This causes an issue because even though the data is returned, the front door
logic has already created 2 transactions to the flash, and a result, we have
extra data returning.

The buffer not enabled case can be caused by otp_ctrl not returning the
flash controller's request for a key.  This in turn can happen because
entropy is not yet enabled.  This latter point deserves a wider discussion
as to the right solution.

To fix this, the forward hint is used to distinguish when the data in the
FIFO is valid vs when it is not.

Signed-off-by: Timothy Chen <timothytim@google.com>