[dv/otp] force some signals to improve FSM error coverage
This PR tries to improve OTP_CTRL's FSM coverage below:
1. sw partition check_fail: Since sw partition does not really do
otp_checks, the check_fail comes from internal ECC reg calculation
error. In this sequence, I force the error to bit 1 in order to trigger
check fail.

2. LC check fail: This is done when we program lc partition via LC
interface but did not perform a reset. This is not recommanded in OTP
spec so I did not support that in scb, but just use this sequence to
check.

Signed-off-by: Cindy Chen <chencindy@google.com>