[padring] Scan role parameters
Ideally, the scan role (none, scan-in, scan-out) would be captured as
part of the pinout Hjson description.
However, due to the need to keep this information in the foundry
repo, an approach with a separate SV package is taken.

The open-source version is just a generic assignment of scan roles.
The scan role parameters for the synthesized ASIC target are defined in
the foundry repo and will be read in when the "fileset_partner" flag is
defined in the build flow.

Signed-off-by: Michael Schaffner <msf@google.com>