[flash_ctrl] Simplify defn of buf_invalid_alloc in flash_phy_rd
The existing expression caused a Verilator UNOPTFLAT warning, because
bit k of the signal depended on bits 0..k-1. Rather than telling
Verilator to schedule things more cleverly, rewrite the definition to
something equivalent which doesn't cause a problem.

buf_invalid_alloc[i] is supposed to be true if it's the first bit that
is set in buf_invalid[i]. The existing code checked that no previous
bit was the first set bit. The new code checks that no previous bit is
set at all. This is equivalent. I'd expect any half-decent synthesis
tool to generate the same code from both versions but, if it doesn't,
the new formulation should be more efficient.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>