[usb / top] Hook-up usb rx enable
- Currently the rx enable is assumed to be dynamic, but this may
  change based on Nuvoton feedback.

There was a discussion regarding whether we should make most of the usbdev inter-signals instead of pinmux signals.
See below for rationale.

In addition to dp/dn, sense and dp/dn_pullup_en all have to remain pinmux-able signals.
Sense may need to come through pinmux since it is used to detect vbus power.  As long as the vbus voltage is not too high (or is clamped to a safe range), it can be muxed on nay pin.
dp/dn_pullup_en make use of pinmux's sleep capture as part of usbdev's suspend / resume feature.

This ends up leaving a few scattered signals that may or may not fall into the pinmux category.
Ultimateily, it is difficult from the perspective of the usbdev to know whether a signal should be pinmuxed or directly signaled out.

The ideal solution are two-fold

1. Make all signals inter-signal, and use a top specific wrapper that connects the necessary signals to pinmux.
2. Make all signals pinmuxable, and manually handle the pad connections where required.

In our design, we currently have most of the support needed for 2, (see #6042) for possible enhancements

Signed-off-by: Timothy Chen <timothytim@google.com>