[spi_dev] Attempt to fix combo loop
A combo loop was observed through the following path

data_i -> dp_sel -> data_o, which then loops back to data_i.

The issue is because in cmdparse the upload case immediately
selects a new datapath because there are no more guaranteed clocks
for certain commands.  However, this same combo selection does not
have to be used for the read return path, so the fix just flops the
read return on spi_clk_out (spi_clk_in might also be okay)

Signed-off-by: Timothy Chen <timothytim@google.com>