[dv/otp_ctrl] align the behavior for fatal error
This PR aligns the behavior for PR #6077.
What changes here is - now if any alert triggered, the otp_ctrl will go
to terminal stage for all partitions (previously only the partition that
has error will go to terminal stage)
This PR works on scb and interface. The coming PR will work on sequence
that does not enable scb.

Signed-off-by: Cindy Chen <chencindy@google.com>