[keymgr,lint] Waive some Verilator warnings about array accesses
"Fixing" this in the RTL would be hard. The problem is that you'd need
to make some 2-bit signals to represent the enum entries.
Unfortunately, AscentLint (the lint tool that we use for signoff)
doesn't accept code that extracts bits from enum values without an
explicit concatenation. So you end up having to write something like
this for each enum entry:

  localparam bit [2:0] ErrInvalidOpBits = {ErrInvalidOp};
  localparam bit [1:0] ShortErrInvalidOp = ErrInvalidOpBits[1:0];

which is getting a bit ridiculous. Just waive the warning.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>