[hmac,lint] Ignore an unused output in a way that linters understand
To ignore outputs from a module, the usual practice is to just connect
them to nothing (with an explicit empty port connection). Connecting
up a signal and then ignoring that signal gives a lint warning, which
we were waiving for AscentLint but not for Verilator.

Switch to the more normal way of doing things and remove the
now-unneeded waiver.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>