[mask_rom] HMAC driver fixes for digest reading and endianness
This change fixes `hmac_sha256_init` and `hmac_sha256_final` to match
the output of `openssl` and `sha256sum` and preserve the numerical value
of the digest.

It also fixes IRQ clearing in these functions so that the driver can be
used more than once. This is required to be able to verify the second
slot if the first one does not succeed.

Signed-off-by: Alphan Ulusoy <alphan@google.com>