[csrng,dv] Break long lines in csrng_if
I've done this by changing the assignments with ternary operators into
an always_comb block.

The advantage is that the formatting gets a bit clearer and it's no
longer difficult to squeeze everything under 100 characters. The
disadvantage is that we now have to make sure everything gets assigned
on both branches of the "if" statement.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>