Convert C defines for param names to snake_case
Many module parameters are converted from UPPER_SNAKE_CASE in their
hjson files to a variant of CamelCase for use in SystemVerilog files.
These parameter names are then converted to the all-caps form of the
CAMLECASE name in C header files.

The UPPERCASEWITHOUTUNDERSCORES form is too hard to read.  This change
converts back to UPPER_SNAKE_CASE for C header files and fixes all
existing uses for the previous naming scheme.

Signed-off-by: Chris Frantz <cfrantz@google.com>