[pwm,lint] Move unused regen signal across CDC; waive lint warning
This signal is here as a reminder to implement reg-enable based
register locking, but lint tools warn both that we're not using it and
that we're not using the regen field of reg2hw_sync.

Move the TODO message to the other side of the CDC primitive and
rename it "unused_regen", which means Verilator and AscentLint should
stop complaining about it (but the TODO message is still clear).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>