[flash_ctrl,lint] Move unused cfgs_i waiver into RTL
The fact that we only use some bits of cfgs_i was waived for
AscentLint, but still caused warnings with Verilator. We could waive
it for Verilator in the same way, but instead let's define an
"unused_*" signal in the RTL to waive just the cfgs inputs that we are
actually trying to ignore.

This should be slightly safer, because it means the warning would come
back if we messed up one of the other branches in the if/else tree for
some reason.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>