[mask_rom] Move all executable code into .text section
All executable code in the mask ROM needs to be in a single
contiguous PMP region. Moving all executable code into the .text
section allows us to define unambiguous `_text_start` and
`_text_end` symbols that can be used to set up a PMP region for all
of the executable code.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>