[tl,dv] Allow bits to be set in responses regardless of mask
In our design, a memory with ECC integrity protection enabled will
always respond with complete words, ignoring the a_mask field. We
don't have much choice about this: it would be impossible to check the
integrity bits in the response otherwise.

Looking through the TL spec (version 1.8.1, specifically the commands
in section 7.2), I don't believe that the check in mem_model.sv is
mandated by the specification. That is, I think our design is still
conformant (phew!). So remove the check.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>