[rom_ctrl] Enable connection to KMAC
This changes the meaning of the SkipCheck parameter once again.
Before, it disabled the connection with KMAC (essentially making KMAC
seem to return after a single cycle, with a digest of '1) and then
also disabled the comparison with the expected contents.

After this commit, rom_ctrl does the handshaking with KMAC and
calculates whether it matches. The only thing that SkipCheck now
controls is the "good" signal sent out to the pwrmgr. We can't
get *that* working yet, because we don't yet model the cSHAKE digest
calculation in the Python code, so can't put the right values in the
top of ROM.

This change should hopefully mean that synthesis results give roughly
the right area, even when SkipCheck is set: since the calculated
digest is exposed as a "DIGEST" register, the synthesis tool will be
forced to generate all the flops to hold it's (not feasibly
predictable) value.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>