[dv/lc_ctrl] Fix regression error
This PR fixes LC_ctrl regression error.
The issue comes when lc_ctrl smoke sequence pushes expected token data
to the push-pull data queue, but in the meantime, the pull request could
be interrupted by other lc_ctrl errors. So we push some data to the
queue that is not being used.
To solve this issue, I add a function to clear push-pull agent's data
when user do not need it.

This PR also has some small fixes in lc_ctrl testbench.

Signed-off-by: Cindy Chen <chencindy@google.com>