Update Verible to latest upstream
Verible is providing lint and formatting tooling and is evolving
rapidly in response to OpenTitan requirements.

It is an optional productivity tool; developers are recommended to have
it installed to run lint and early/experimental formatting on their own
machine. Lint is also done in CI.

Update Verible to the latest upstream version to unblock other work,
e.g. #4613 and to include recent lint improvements.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>