[dv/common] add ECC support to mem_bkdr_if
This PR adds full ECC support to the memory backdoor interface.

To do this, secded_gen.py was modified to also output an enum of the
different ECC variants, which are used to parameterize the mem_bkdr_if
so that we can pick up the correct data widths internally, etc...

Some ECC-specific wrapper functions for `read()` have also been added to
mem_bkdr_if that return the syndrome and error information as well as
the read data - these can be useful for ECC error tests
(sram/flash/otp).

A followup PR will enable this functionality in the OTP and Flash
testbenches.

Signed-off-by: Udi Jonnalagadda <udij@google.com>