[push_pull agent] Driver code refactor
Noticed an opportunity to refactor the code a bit to make it more
efficient. Primarily, the host and device mode drivers check at runtime
on every cycle whether its in push mode or pull mode to decide which
signals to drive, which IMO can be improved.

The host and device driver classes really need to be broken further down
to 2 classes each, for push and pull directions (so have 4 classes) and
decide at the time of creation, which one to pick. That way, at runtime,
we will no longer need any conditionals.

Each of the driver variants are small enough - so they are all added to
the same file - `push_pull_driver_lib.sv`, which originally was
push_pull_driver,sv`. The `push_pull_driver` class (which previously did
not do much), now absorbs a bit more of the driver code (the
`get_and_drive` part) and hands off the actual interaction with the
interface to a `sub` driver, which is implemented as a pure virtual
class. The sub driver is then specialized to the 4 variants.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>