[otbn] Fix error when using the last x1 value as a base address
The pop_i input to otbn_stack was correctly using pop_stack, which
checks that the current instruction (which requests a stack pop) is
being committed. This avoids a double stack pop from code like

    lw  x2, 100(x1)

However, we didn't do the same conditioning on call_stack_err_o. This
means that if an instruction like the above happened to read the last
element of the stack then we'd see call_stack_err_o go high on the
second cycle (when rd_commit_i is false, but the stack is now empty).

To reproduce (without the fix), run:

    util/dvsim/dvsim.py \
        hw/ip/otbn/dv/uvm/otbn_sim_cfg.hjson \
        --fixed-seed 4100705798 -i otbn_single

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>