[dv/kmac] cycle accurate model bugfixes - pt 5
this PR makes more fixes to the cycle accurate model, this time relating
to the EDN entropy interface - specifically, overhauling how we time the
first round of a keccak hash.

Some related fixes have also been made to the base sequence, with how we
calculate a "static" entropy mode that is then applied to each hash
iteration.

Signed-off-by: Udi Jonnalagadda <udij@google.com>