[dv] Randomize ROM in chip_base_vseq
Because the rom_ctrl block reads the contents of ROM and passes them
to KMAC (which has X checks), we need it to have *some* architectural
contents. We don't really care what, though, so randomize it here.

Tests that actually have a ROM image use chip_sw_base_vseq, which
loads up the ROM image a bit later.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>