[sw] Implement prng for random testing
This adds the ability to generate and test randomized payloads from SW,
rather than test fixed payloads. Stimulus and hence, verification
quality is better when HW is programmed and checked against with
constrained random data. This will be mainly used in DV code where both,
SW image and the HW can be probed to extract the data and perform
checks.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>