[dv/sram] fix initialization checks
currently to trigger memory initialization the base sequence calls
`csr_wr` to fetch a new key, then calls `csr_update` to write the next
control bit, however due to the behavior of `csr_update`, a new key
request is also triggered at the same time.

This sometimes messes up the scoreboard timing model, but is also
entirely unnecessary, so this PR removes it and replaces with a single
call to `csr_wr` to write both bits at once to simplify things.

Signed-off-by: Udi Jonnalagadda <udij@google.com>