[sw, pwrmgr dif smoketest] Fixes to get it working
The test as it exists had a couple of bugs:
- It configures AON timer to issue wakeup after 5 200kHz clocks (25us)
which is too short - the whole pwrmgr config routine takes about 50us,
and the low power entry takes several clock cycles to full enter the low
power mode. Increased it to 20.

- AON timer programming is followed up with `usleep(100)` which seems
unnecessary. Adding that causes the wakeup to occur even before pwrmgr
is programmed.

- Low power configuration disables USB clock in active power - that
results in clkmgr not being able signal back to pwrmgr that it was able
to disable all clocks properly, causing pwrmgr FSM to hang. See #6504.
Updated the config to retain USB clock to be active in active power.

- Re-org'ed the code a bit such that AON programming is not done after
wake up from low power exit.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>