[tool_requirements] Set Verilator and Ninja checks to as_needed
Previously, we were always checking the tool version of Verilator and
Ninja in lint runs even when the tools were not used.

While this is typically not a problem (these are standard tools every
developer has to install locally), it can be an issue for Dvsim runs
that are dispatched to cloud runners which only load the required tool
image. E.g. Verilator lint runs were failing due to a missing Ninja
installation.

This change sets Verilator and Ninja to "as_needed", and modifies
the checks in the common checker core such that only the version of the
requested tool is being checked.

Signed-off-by: Michael Schaffner <msf@google.com>