[rom_ctrl] Fix gnt/rvalid timing just after initial check finishes
The rom_ctrl block is supposed to give access just to the checker
until the checker is finished, at which point the bus can start
accessing data.

To control this, there is a mux with a sel_i signal. When sel_i is 1,
the checker is still running. When it drops to zero (a one-time
event), the bus should have access.

Before this patch, there was a bug where the block would seemingly
respond to a request in zero time, raising the bus grant signal and
rvalid at the same time. The problem was just that we were incorrectly
using a registered signal for the grant: the ROM had taken the request
a cycle earlier but we didn't tell the bus.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>