[otbn,dv] Pass mnemonic in trace from ISS
This is a partial change, in the direction of adding some functional
coverage to OTBN. The problem is that we want to track things like
"did I execute ADDI?" but don't want to write (yet another) decoder in
the coverage model, which might be buggy or get out of sync with the
actual RTL and ISS, yielding bogus coverage data.

To solve this, the plan is to pass mnemonics from the ISS to the trace
comparison and then (over DPI) to the coverage model. We know that the
RTL trace items will be in sync with the ISS sync items (otherwise the
test fails), so we can just build two queues on the SystemVerilog side
to merge together instruction data as seen by the RTL trace interface
and information about decoded instructions as seen by the ISS.

This patch doesn't do any of the later work: it just adds a mnemonic
to the output from the ISS in stepped mode and teaches the C++ code
that receives it to understand what it's seeing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>