[prim_secded] Add C reference models for Hsiao encode
This adds generation of C code for the Hsiao code encode. This is
intended to be used by DV environments for backdoor memory loading,
which only requires encode.

No C code is generated for the Hamming codes. These are slightly more
complex as the final code bit factors in previously computed bits where
the Hsiao code does not. So they need a slightly different approach to
the C code. Hamming codes aren't currently used in OpenTitan so C code
for their encodes aren't required.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>