[otbn,dv] Monitor the trace interface directly; connect to coverage
This commit adds otbn_trace_monitor, a UVM monitor that snoops on the
RTL trace interface. This passes transactions to the scoreboard, which
already received the equivalent transactions from the ISS. The
scoreboard then pairs them up again and passes them to otbn_env_cov to
be used for coverage checking.

This patch adds the infrastructure that you need in order to cross
executed instruction mnemonics (from the ISS) with instruction
data (from the RTL). It doesn't add any new coverage points, but it
does add a cross-check to make sure that the instruction addresses
from the ISS and RTL match. Note: This isn't part of the verification
of the block, as such, but is needed to make sure that we are
collecting the coverage that we think we're collecting.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>