[ doc, build_docs ] Optionally bind to global wan when previewing
This commit adds an additional --bind-wan argument to build_docs.py.
This forces the hugo server to bind the http server to a port on the
global interface (as opposed to just on localhost, which is the default).

This is important when using hugo on build farms with older (or no) browser
support, and no support for ssh tunnelling.

For testing other options, there is also a --hugo-opts flag for passing
arguments into hugo directly.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>