[rom_ctrl] Only raise ROM req line when not in reset
The previous code assumed that the checker always requested ROM
accesses. This is logically correct, but meant that we were raising
the ROM's req line when in reset. Since the ROM doesn't itself have a
reset signal, this would mean spurious activations of the ROM
macrocell.

This patch explicitly registers the req signal in the checker's
address counter and passes it through the rest of the design.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>