[spi_device] Revise readcmd to use cmd_info CSR
This commit revises the `spi_readcmd` submodule in the spi Flash mode.
It previously used the pre-defined opcode values and the configurations
to parse the address, the existance of the dummy cycle, read IO modes
(Single/ Dual/ Quad).

Now it uses the SW configurable cmd_info CSR to get the command
structures.

It introduces mbyte_en field too. It is not fully implemented yet. It is
a stopgap to support MByte transition. Full functionality will be
implemented later to reduce the amount of the commit size.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>