[otbnsim] Make OTBNState.pc_next always available
pc_next was giving the next program counter if it wasn't just pc + 4,
i.e. pc_next was only set if a jump or branch forced it to be
non-linear. This setup required logic using pc_next to deal with that
and fill in "pc + 4" if pc_next was None.

This commit switches things around so that the next program counter is
always available by calling get_next_pc().

We still keep track of program counters that are explicitly set to
omit the next program counter information if it wasn't explicitly set.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>