[sha3] Recast an always_comb block to avoid a lint warning
This lint warning from Verilator is about an inferred latch. It seems
that Verilator version v4.108 doesn't spot that the three fields of
error_o are all assigned at the top of the always_comb block.
Recasting the statement as an equivalent assignment to error_o fixes
the problem.

This spurious warning is already fixed in Verilator, by commit 4f36e3e
which appears in release v4.202. So this won't be a problem long-term,
but there's no real reason to depend on newer versions of the tool at
the moment: if anything, the rewritten code looks easier to
understand.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>