[dpi_memutil] Fix bug in RegisterMemoryArea
If !name_to_mem_.emplace().second then we already had a memory area
for the given name and pr.first->second is a valid index into the
vector called mem_areas_. If not, it ain't.

We grab a reference to the already-existing memory area in order to
print out an error message, but were accidentally doing so outside of
the error path. This causes a segfault, but only some of the time. (In
particular, it seems to depend on the Verilator version, which is a
bit bizarre).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>