[pinmux/tap] Rework DFT TAP hooks
The current DFT TAP hooks expose the JTAG signals as a struct type,
and the pinmux_jtag_buf module is instantiated not only for the DFT
TAP but also for the other TAPs that are connected in RTL.

While this is all correct from a syntax and design perspective, the
struct type seems to confuse certain automatic DFT insertion tools.
Further having multiple pinmux_jtag_buf modules instantiated exposed a
uniquification bug in Tessent.

To solve this issue, this patch adds a breakout module called
pinmux_jtag_breakout for the DFT tap on the top-level that exposes the
DFT signal hooks as simple logic ports.

Signed-off-by: Michael Schaffner <msf@opentitan.org>