[test] Use longer timeout for maskrom tests
Getting through Mask ROM in a verilated simulation as we run them in CI
currently takes around 42 minutes, which exceeds the timeout of 2500s =
41.67 minutes. Extend this timeout to 55 minutes.

Jobs are killed by Azure Pipelines after 60 minutes, so staying under
that threshold gives us a chance to actually run into the timeout set in
the test harness.

The timeouts we have here remain fragile and very susceptible to
variations in the runners we get assigned from Azure. I filed #6684 to
look into more sustainable solutions.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>