[dvsim] Add verilator to nightly
This PR adds verilator based simulations to the CI and nightly runs.
We cannot sustainably keep adding more and more tests to `systemtest` -
it puts a strain on our CI runtime, which is already too high.

Plus, there are various benefits of running Verilator sims within DVSim:
- Testplanning for a more doc-oriented approach
- Report generation
- Triaging failures
- Parallelize deployment to external compute systems over LSF / GCP
- Define regression sets

For CI, the smoke regression is not enabled. This is because we already
have `systemtest` hooked up and running.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>