[otbn,dv] Implement coverage tracking for base insns, except loops
This commit adds covergroups and coverpoints for the instructions in
the base instruction set. It also adds notes to the DV document to
explain how the things we're trying to track map onto coverpoints and
crosses in the DV code.

The commit doesn't include code for LOOP and LOOPI, nor does it
include a couple of coverpoints where we want to see control flow
instructions at the end of a loop body. We'll add those in a
follow-up, but that will need to extend the trace monitor to look at
some other design signals in order to figure out the current loop
state.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>