[doc, testplans] Fix non-existent tests
Several testplans indicate tests that do not exist yet. It would be
better to set the name correctly when the test is actually developed.
Going forward, missing tests in the testplan and written tests that do
not map to a testpoint will be flagged as an error in the pre-submit CI.

In addition, there are some cosmetic fixes such as line length exceeding
100 chars.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>