[otbn,dv] Simplify how we bind in trace interfaces
Using a hierarchical instance for the bind target makes things a bit
clearer (no problem if there are two hits, for example). Also, it
means we can just use a hierarchical reference to install the
interface in the UVM config db and don't need to mess around with the
otbn_trace_uvm_injector module.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>