[alert_handler] Duplicate LFSR counter and update EDN reseeding logic
This duplicates the timer counters inside the LFSR timer to harden this
module against FI attacks. In case the LFSRs or counters do not agree,
the ping fail alerts are continuously asserted.

This also updates the entropy hookup to EDN - although this will have to
be finalized when the reverse ping mechanism for escalation receivers is
added (the reseeding needs to be properly synchronized with the timeout
counters inside the escalation receivers since this could otherwise lead
to spurious escalation action triggers).

See also #2049

Signed-off-by: Michael Schaffner <msf@google.com>