[ci] Avoid a spurious warning from grep in the non-ASCII check
We've got a symlink to a directory checked in to the repository (it's
hw/ip/aes/pre_syn/python). This appears in the "git ls-files" list,
but grep then gets a bit confused about being asked to search through
a directory. Use "-d skip" to tell it to silently ignore the target of
the symlink.

Also, fix the message in quick-lint.sh: apparently, I was too eager
with my copy+paste!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>