[rom_ctrl] Remove a bogus TODO comment
When I wrote this, I imagined that a bus access when the ROM checker
was running should result in an error response. What we've ended up
with is that the bus access will just stall until the checker is done,
which seems like a perfectly reasonable behaviour. Let's just keep
that!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>