[dv/otp_ctrl] Change used_addr from a queue to an associative array
As Matute suggested, I changed the data structure to collect used
addresses. Now we are using an associative array to collect written
addresses.

This PR also improves the sequence randomization. The old way is to try
randomize addresses and avoid write used address, now we will randomize
any address, but `dai_wr` task will skip the write if it is the address
is already written.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>