[mask_rom, pinmux] Add Pin Multiplexer driver for mask ROM
Add a simple pinmux driver for the mask ROM to use. Uses the
`abs_mmio` library for register accesses.

Right now the mask ROM is only using two pins: one for the UART RX
and one for the UART TX. The number of pins the mask ROM needs to
configure will grow however so the library has been designed to
make it trivial to add new pin mappings. The configuration the mask
ROM needs is static so we do all of the work in a single function:
`pinmux_init`. Pad attributes may need to be set dynamically so
those will be handled later.

The driver itself is table driven which makes it easy to see which
pins are mapped to which peripheral inputs and outputs. In the
future we may want to revisit this decision to reduce code size.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>