[doc] Fix the functional coverage requirements
Previously, the coverpoints and crosses were computed separately and
averaged to show the final covergroup coverage score. To illustrate with
an example, if a covergroup had a cp with 90% coverage and cr with 50%,
it would compute the total score as 90+50/2 = 70%. This could be
misleading, becuase the cross could have orders of magnitude more bins
to cover than the coverpoint. With `-group ratio` enabled in the
previous commit, the coverge score is computed more realistically as
total bins covered in all coverpoints and crosses / total bins in all
coverpoints and crosses, which gives a more realistic picture of the
remaining work to be done. But the side effect of that is the tool no
longer provides a separate score for coverpoints and crosses. Hence, we
need to update the V2 and V3 functional coverage criteria to look at the
overall score instead.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>