[sw/silicon_creator] Move struct definitions shared across boot stages to a new header
We plan to move the bulk of the signature verification code to
`sw/device/silicon_creator/lib` to be able to use it in multiple boot stages.
Once this work is completed, the only difference between boot stages
should be the public keys that they use to verify manifest signatures
(`sigverify_keys.c`).

To this end, this change:
- Moves `sigverify_buffer_t` and `sigverify_rsa_key_t` to
`sw/device/silicon_creator/lib/sigverify_rsa_key.h`, and
- Moves `sigverify_rsa_key_get` to `sigverify_keys.c`.

Signed-off-by: Alphan Ulusoy <alphan@google.com>