[rom_ctrl,util] Add a collision check to scramble_image.py
It matters if we get one of these collisions because
it (theoretically) might allow an attacker to glitch address lines to
hide a modification he or she had made to a duplicated word in ROM.

This is a kind of random process: we pick the RndCnstScrNonce and
RndCnstScrKey arbitrarily and they *might* just happen to cause a
collision. A back-of-the-envelope computation suggests this happens
<1% of the time, but that's still too frequent to fail the SW build
for. For now, we just complain very noisily if it has happened. Maybe
we'll add a "--be-very-careful" flag later that's used for final
builds (there are probably other similar checks that need to hide
behind something similar).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>