[dv, flash_ctrl] Fix the smoke test
- The smoke test writes and reads back a random chunk of memory within
the flash.
- The start address and the randomized number of words is bus-aligned
(i.e. 32 bits) and not flash word aligned (64-bits).
- Between each randomized configuration, the flash mem is invalidated.
This means if we write to a flash mem half word, while leaving the other
half word as Xs, that causes an X problem on reads, because flash ctrl
reads in the whole word, not just the half word.
This commit fixes that problem.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>