[reggen] Generate a single we/re signal per register in reg_top
Most of this (enormous) diff is auto-generated. The only manual
changes are to reg_top.sv.tpl, access.py and register.py.

The idea behind this patch is that the write-enable and read-enable
signal for a field are both determined at the register level (by
reg_we or reg_re, combined with an address check). Before this patch,
each field that needed a write-enable or read-enable signal defined
one. For example, if we had a writeable register R with fields F0 and
F1, we were defined both R_F0_we and R_F1_we and always set them to
the same values. This patch changes things so that we define "R_we"
instead and use it as the "we" signal for both R.F0 and R.F1.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>