[dv,cip] Guarantee ordering of sampled A and D transactions
When TileLink transactions are observed by tl_monitor.sv, it writes
items to "A" and "D" ports, depending on what channel observed them.
These writes race against each other, so there's no order guaranteed
between an A and D transaction sampled on the same clock edge.

These ports are connected to fifos in cip_base_scoreboard.sv. Before
this commit, the scoreboard used two tasks (process_tl_a_chan_fifo and
process_tl_d_chan_fifo) to sample from the two fifos. Once an item was
found on either fifo, the corresponding task called process_tl_access,
which itself is a task implemented in the subclass.

The problem is that the subclass implementation couldn't guarantee the
ordering that it would see if there was an A and D transaction on the
same cycle.

But why does this matter? Suppose that my block has a register, FOO,
whose value can be written by hardware but in a way that we can
predict. Then I need to somehow store an expected value when I see the
A channel transaction for a read of FOO, then check it matches the
transaction data for the following D response. The obvious way to do
this is to mirror the expected value in the RAL with a call to
uvm_reg::predict() when I handle the A channel transaction, and this
is what code like keymgr_scoreboard::process_tl_access does.

What happens if I see two back-to-back reads of FOO, where it changes
value between the two? That is, what if the following transactions
happen? (the (1) and (2) are indices to match up the A-side request
with the D-side response)

    Cycle | A channel   | D channel
   ---------------------------------
    0     | Get FOO (1) |
    1     | Get FOO (2) | FOO rdata (1)
    2     |             | FOO rdata (2)

Before this patch, there were two possible orderings for the resulting
calls to process_tl_access(). We'd either see

    A (1); D(1); A(2); D(2)

or we'd see

    A (1); A(2); D(1); D(2)

If the expected value of FOO changes between the two transactions, the
second ordering (AADD) is a disaster if we're trying to use the RAL to
mirror things: we'll end up checking the D(1) rdata against the A(2)
expected value.

This patch guarantees that, in the (common) case where the device
always responds on the next cycle, we always see the nice "ADAD"
order. To do this, we make sure to handle D before A, meaning that on
the "middle cycle" above, we see D(1) then A(2).

This isn't a completely general solution. For example, if the device
might take multiple cycles to respond, we might get something like this:

    Cycle | A channel   | D channel
   ---------------------------------
    0     | Get FOO (1) |
    1     | Get FOO (2) |
    2     |             | FOO rdata (1)
    3     |             | FOO rdata (2)

and no amount of cleverness can guarantee the "ADAD" ordering. In that
case, the device's scoreboard is going to need to be cleverer.

Similarly, if the device can respond on the same cycle, we might get
something like this:

    Cycle | A channel   | D channel
   ---------------------------------
    0     | Get FOO (1) | FOO rdata (1)

in which case we'll actually have a "DA" ordering, which is a bit
confusing! A future version of this code could be cleverer to avoid
this problem (by counting transactions to realise that the two items
were supposed to match). But that's harder and we don't have any
devices in OpenTitan that respond in zero time at the moment. The
behaviour with this patch is one of two behaviours that the previous
code might have had, so it's no worse at least.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>