[lc_ctrl] Add CSR for switching to an external clock
This makes the switch to an external clock source configurable through
the CSRs and LC TAP. It is similar to the additional vendor-specific
OTP test CSR that can be used to switch to external VPP in certain life
cycle states.

Note that the external clock option is not available in production
states.

Signed-off-by: Michael Schaffner <msf@opentitan.org>