[otp_ctrl] Add bit remapping option to OTP image generation script
The ECC + data mapping within a word in the technology-specifc OTP
macros may be different from the mapping employed in the generic model.

This adds a command line option that allows to specify a permutation
that is applied to the generated OTP image before writing the hexfile.

For example, the following command reverses the byte order for a word
width of 24bit:

util/design/gen-otp-img.py -o test.vmem --data-perm '[7:0],[15:8],[23:16]'

Signed-off-by: Michael Schaffner <msf@opentitan.org>