[sw/silicon_creator] Add minimal ePMP library
This is a lightweight API for configuring and tracking ePMP control
register values in RAM. This library does not provide a general
purpose API for modifying the hardware registers themselves since,
for efficiency reasons, the mask ROM will be configuring the
hardware registers directly.

The API contains a function to unlock the address space required for
reporting test status. All tests booting from the mask ROM will need
to explicitly unlock the DV address space or attempts to report their
status will be blocked and result in an access exception.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>