[otbn,dv] Correct "next PC" checks on stopping instructions
Normally, as part of executing an instruction, we also check that the
next instruction's PC is valid. However, we should only do that check
if there will actually be a next instruction. That is, we should skip
the check if this instruction is already causing the processor to
stop.

This came up in a randomly generated test, which happened to generate
a branch to 0xffc (the top word in memory) which contained an ECALL.
The fall-through address would be invalid, but that shouldn't matter
because we're stopping anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>