[otbn] Make CFI-enabled binaries available in meson
Produce CFI-enabled ELF files and embeddeable RV32 objects, in addition
to the "plaintext" ELF files and embeddable objects. This causes meson
to always generate CFI-enabled binaries, and allows authors of build
description files to choose if they want to embed the CFI-enabled binary
into an Ibex binary, of the "plaintext" ones.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>