[doc] Add detailed contribution guide
This document gives detailed guidance on contributing to the OpenTitan
project. It has gone through many rounds of review in GDoc form (and
this PR simply converts that to Markdown, with the odd grammatical fix
along the way).

The contribution guide effort was kicked off by Pirmin and has since
seen suggestions, edits, and commits from a wide number of people across
the project. Thank you to everyone for your input.

Signed-off-by: Alex Bradbury <asb@lowrisc.org>