[nmi_gen] remove the nmi_gen dummy module
Since we've now wired up the escalation channels in the design,
we can actually remove this dummy module.

Note that this removes the only module in top_englishbreakfast that
exposes a HW reset request.
This is currently not compatible with the pwrmgr templates (HW and SW)
which expect at least one HW reset request to be present.

To this end, this commit also instantiates the watchdog timer which
exposes a HW reset request. This should not add considerable overhead
to top_englishbreakfast.

Signed-off-by: Michael Schaffner <msf@opentitan.org>