[lc_ctrl] Add scrap state for main FSM
This adds a scrap state to the main FSM in order to break an unintended,
circular logical dependency.

I.e., the life cycle FSM decodes the SCRAP life cycle state as any other
state and keeps on sampling the life cycle vector coming from OTP.

However, since the SCRAP state asserts `lc_escalate_en`, OTP escalates
and resets all partition outputs back to the default, causing the LC
controller to latch an invalid LC vector encoding.

This has a few unintended side effects. For instance, it causes the
`lc_escalate_en` signals to toggle from ON -> OFF -> ON. Also, it leads
to an incorrect life cycle TAP readout, as the life cycle state reads
back as "INVALID" instead of "SCRAP".

This commit introduces the following fixes: 1) the ON -> OFF -> ON
toggle is fixed in the `lc_ctrl_signal_decode` logic, which now asserts
`lc_escalate_en` when the data valid indication from OTP unexpectedly
goes to 0.

2) the FSM is moved into a terminal `ScrapSt` right away when a SCRAP
life cycle state is detected. In that state, the life cycle vector
coming from OTP is not sampled anymore, thereby breaking the logical
loop mentioned above.

Note that 2) has the additional benefit that life cycle transition
operations are now explicitly disallowed when the device is in SCRAP.
Before, this was checked only after initiating a life cycle transition
as part of the transition checking logic.

This fixes #7286

Signed-off-by: Michael Schaffner <msf@opentitan.org>