[rom_ctrl] Cut the instantiated ROM width back down to 39 bits
We don't really have any need to keep it at 40 because our macro
wrapper can do the rounding up if needed for the technology library.
It also avoids an awkward situation where we were "encrypting" a zero
bit in counter mode. This didn't really matter for the ROM
application (where confidentiality isn't a requirement), but it's a
bit awkward and easy to avoid.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>