[dv/kmac] Update scb to match RTL updates in #7242
This PR updates the cycle accurate model to match RTL updates made in
PR#7242.

The scb model can now support fully dropping an App transaction and
returning to Idle state in 2 different scenarios:
- SW clears error immediately, before the App hash is completed.
  In this case, immediately return to Idle state waiting for the next
  transaction and force the `kmac_app_monitor` to signal that it is
  idling as well.
- SW clears error after waiting for the hash to be completed.
  In this case, there is no guarantee of output digest correctness, so
  we do not check this in the scb, and then return to Idle as usual
  afterwards.

In addition, the app_fsm model in the scb is updated with the new
`StError` state that we transition into one cycle after seeing an error
on the App interface to allow forward progression without entering a
terminal state.

Signed-off-by: Udi Jonnalagadda <udij@google.com>