[sw/dif] Add FW override support to entropy src.
The entropy src allows the firmware to disconnect the raw entropy from
the pre-conditioner block at the output of the health checks. This
allows firmware to access raw entropy.

The firmware can also write data into the pre-conditioner block to
provide firmware-controlled entropy at the output of the entropy src.

This functionality can be useful to:

1) Implement additional health checks in firmware.
2) Implement KAT tests of the pre-conditioner block.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>