[rom_ctrl] Add a parameter to stub out scrambling in rom_ctrl
This is going to be used for the EnglishBreakfast top-level, where we
want a cut-down "OpenTitan-like" design that fits on a smaller FPGA.

When the parameter is enabled, all the checker FSM code gets stubbed
out and the de-scrambling ROM gets replaced by a plain one.
Connections to KMAC get broken (so we never start an operation). The
connection to the power manager always asserts that the ROM has been
checked and is good. The connection to the key manager always asserts
that we have a digest and it's zero.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>