[doc] Clarify (UN)LOCKED_OWNERSHIP states
The use of OWNED and UNOWNED was confusing and I'd like to discuss
ownership transfer and similar flows with LOCKED_OWNERSHIP and
UNLOCKED_OWNERSHIP. This will distinguish it from UNLOCKED_ROM states
but will also clarify that there can be a Silicon Owner when the device
is in an "UNOWNED"->"OWNERSHIP_UNLOCKED" state.

Added a URL that should be sufficient to request access to the source
files for the svgs I edited to facilitate future edits.
Signed-off-by: Drew Macrae <drewmacrae@google.com>