[topgen] Be explicit about where hint clocks get connected
Despite the big patch(!), there's no functional change here. The trick
is that we split amend_clocks() so that some of the function runs
after we've loaded up the blocks' hjson descriptions. Doing that means
that we can point to "idle" signals explicitly in clocking items and
use the names to wire everything up together.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>