[otbn] Add Hugo shortcode to link to instruction description
And use it in the OTBN documentation.

The new shortcode also changes the formatting slightly from what we had
before (most of the time): Instructions are not displayed in a monospace
font any more. This aligns with how we render register references, which
also don't use a monospace font.

Fixes #4869

Signed-off-by: Philipp Wagner <phw@lowrisc.org>