[otbn] P256: Don't overlap blinding `rnd` with output `x1`
The ECDSA P256 routine takes an input argument `rnd`, which is supposed
to be randomness used for blinding. However, I was unable to find code
that makes use of this blinding constant. The calling conventions
assigned `dmem[4]` to the pointer pointing to `rnd`.

At the same time, the verify operation writes its output `x1` also to
the variable pointed to by `dmem[4]`. In the code, this pointer is
therefore sometimes called `dptr_x1`.

With no blinding being employed and the `dmem[4]` memory location used
exclusively for storing the pointer to the x1 result, this commit
removes the blinding parameter completely and updates the software to
name the corresponding field x1.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>