[prim_lfsr] Fix assertion issue occuring right after reset
There are cases where the LFSR enable signal remains asserted
by the instantiating design while in reset. In such a case,
the NextStateCheck_A may erroneously fire in the first
cycle after reset release due to a SystemVerilog scheduler
peculiarity when the reset is synchronized with a reset synchronizer
flop that releases the reset right on the active clock edge with 0
simulation delay.

If this happens, the assertion body may be out of sync with the design,
since the assertion disable_iff statement is NOT evaluated with a value
sampled in the PREPONED region. In other words: when the assertion is
evaluated in the OBSERVED region (after active and NBA), rst_ni may
already read as 1 due to reset deassertion - however the flops in the
design may not have seen that change yet, leading to an assertion
failure because the flops have not yet changed state, whereas the
assertion expects them to have changed state.

There are a couple of possible solutions:
1) ignore the first clock tick after reset release by shifting the
assertion by one clock tick to the right
2) use a sampled reset value for disable_iff (#sampled(!rst_ni))
3) make sure the surrounding logic never asserts the LFSR enable
signal while in reset.
4) make sure that there is an artificial propagation delay on
all reset synchronizer outputs to shift reset deassertion away from
the active clock edge.

This patch implements solution 1) since this should ensure that the
issue does not occur anymore (also in future designs that instantiate
this prim).

Signed-off-by: Michael Schaffner <msf@google.com>