[prim_subreg] Prevent lint error on string param
We operate on a string parameter (without an explicit data type for tool
compatibility). This causes Verilator lint warnings:

```
%Warning-WIDTH: ../src/lowrisc_prim_subreg_0/rtl/prim_subreg_shadow.sv:40:57: Operator EQ expects 24 bits on the LHS, but LHS's VARREF 'SWACCESS' generates 16 bits.

%Warning-WIDTH: ../src/lowrisc_prim_subreg_0/rtl/prim_subreg_shadow.sv:41:57: Operator EQ expects 24 bits on the LHS, but LHS's VARREF 'SWACCESS' generates 16 bits.

%Warning-WIDTH: ../src/lowrisc_prim_subreg_0/rtl/prim_subreg_shadow.sv:41:67: Operator COND expects 24 bits on the Conditional False, but Conditional False's VARREF 'SWACCESS' generates 16 bits.
```

This commit removes the "bit" data type from the inverted string
parameter (since there's no real value in having it if it doesn't solve
the lint problem) and simply waives the message.

AscentLint doesn't show errors on the previous code.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>