[otbn] Describe WSR error conditions in docs and ISS
This actually got added to the RTL back in September with commit
d3154ec2 but we didn't update the docs or the ISS to match. :-(

The slightly odd looking "check, then look up again" structure in the
two execute() functions is to get a nice documentation rendering. We
have to check before we start reading GPRs (to avoid side effects with
x1), but don't have a nice way to render "xxx = state.get_wsr(idx)" in
the documentation. Checking the index and then using it again avoids
both problems.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>