[opentitantool] Add meson build target for opentitantool
This change means that the CI infratructure will build the
opentitantool (and therefore check that it compiles) although it
won't actually use it for anything yet. This will at least help
us detect totally broken builds.

This is a straightforward copy of the meson build target for the
rom_ext_signer. The build is conditional on rust (cargo
specifically) being installed and will otherwise be skipped.

I'm not sure how we should model the opentitanlib dependency in
meson so I've left it up to cargo to handle for now. It seems to
do the right thing.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>