[sw/tock] Remove Tock
The tock directory was added in #1488 to support a downstream OpenTitan
board definition. This was done so Rust Tock drivers could be written
on top of the C DIFs. As we no longer intend for DIFs to be used in
production code, this approach no longer makes sense.

The existence of the tock directory and its downstream board definition
is no longer of use, and is only serving as a source of confusion.

Remove the tock directory, tock build scripts, and support for local
tock builds.

Signed-off-by: Jon Flatley <jflat@google.com>