[otbn,dv] Refactor how the RIG ends programs
We want to add some generators that end programs with something other
than an ECALL (by generating errors).

Generalise things so that a generator class can say it does so (with
the new ends_program class variable) and change some variable names to
reflect this.

Also, change how we generate the top-level program. We now decide how
long the "head" and "tail" should be, respectively, and then generate
them separately. This makes some of the logic rather simpler,
especially when we start adding snippet generators that cause errors.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>