[tlul_adapter_sram] Fix minor bugs in byte write logic
This fixes a few bugs in the byte write logic and adds a workaround for
potential RAW hazards that can occur when the byte write logic is
employed together with prim_ram_1p_scr.

This workaround should be removed once the DV environment of sram_ctrl
is able to handle these corner cases (#7461).

Signed-off-by: Michael Schaffner <msf@google.com>