[otbn] P256 test: Read nonce and blinding constant from RND/URND
Modify the ECDSA P256 to read the nonce (k) and the blinding constant
(rnd) from RND and URND, respectively, instead of supplying it from host
software.

This commit does not change the API of p256.s, but instead loads the
random numbers in the wrapper to keep the change minimal. This isn't
optimal from a performance perspective, e.g. as we cannot overlap
getting RND from the EDN with other operations. However, it makes
testing much easier until.

We also have to remove the checks for the expected values of the sign
operation, since they depend on the nonce, which is now random. We
consider the operation successful of a sequence of sign and verify
passes.

Fixes #6993

Signed-off-by: Philipp Wagner <phw@lowrisc.org>