[otbn] Rewrite case in bignum ALU to make all branches reachable
operation_i.sel_flag has type flag_e, which has 4 possible values (the
four different flags: C, M, L and Z). The default case is only
reachable in simulation if the operation_i.sel_flag is 'x, which
doesn't happen and gives a coverage hole.

In other situations, there might be non-X situations where the case
statement might have picked the default. For example, the enum might
have had 5 items, so there would be 3 invalid bit patterns. In this
case, we'd have to be careful about changing the behaviour of the
default case. That doesn't apply here.

Also, reorder the case items to match the ordering of the
enum (originally, there was disagreement in the code about whether M
or L came first: apparently, this file supported the losing side!)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>