[otbn,dv] Add a generator to RIG to blow the loop stack
This is currently pretty simple-minded: you end up with an instruction
trace like this:

    00000a3c | loopi          943, 1236
    00000a40 | loopi          643, 2841
    00000a44 | loop           x24, 757
    00000a48 | loopi          126, 518
    00000a4c | loopi          813, 1999
    00000a50 | loopi          900, 2354
    00000a54 | jal            x29, .+668
    00000cf0 | loop           x8, 484
    00000cf4 | loop           x29, 3281
    00000cf8 | loop           x8, 2773

There's a TODO note to add some more straight line instructions if we
decide we want something that looks less synthetic. Note the JAL
instruction in the middle of that trace: we emit jumps if we're
running out of space.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>