[otbn,dv] Add the "BadAtEnd" generator to RIG
This generates a branch, jump or loop as the last instruction in a
loop.

To do this, we also factor out the gen_head method in the Branch
generator. This is the bit that generates the actual branch
instruction, which is what we need for BadAtEnd.

Similarly, we factor out a _pick_head method in the Loop generator.
This is in charge of picking the LOOP/LOOPI instruction (it also
returns some auxiliary information about loop shape that we need).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>