[otbn,dv] Fix calculation of k_min when picking LSU targets in RIG
Before, we were missing some allowed values of k because k_min wasn't
being rounded properly. In particular, for WSR operations we were
computing

   (0 + (-1 - 1)) // (-1) = (-2) / (-1) = 2

when we really wanted

   ceil(0 / (-1)) = 0

This gave a k_min of 2, so we would only write to WSRs >= 2. Since we
don't currently support URND (on WSR 2), the result was that every
operation was to WSR 3. Oops!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>