[dv] Minor adjustment to prediction order.
In the old implementation, the dv code would actually predict the
value on the first write but skip the second.  This technically
works too, but is out of sync with how the RTL behaves.

This minor tweak just makes sure we skip the first write and mirror on
the second write.

Signed-off-by: Timothy Chen <timothytim@google.com>