[otbn] Specify reaction to memory accesses while OTBN is busy
OTBN has a data and an instruction memory, both of which are meant to be
"inaccessible" as long as OTBN is executing a command. The question is:
what happens if host software still tries to access the memory?

Currently, we vaguely specify that we'll "block" the access, i.e. hang
the bus.

In further discussions we agreed to instead issue a fatal error, which
follows the normal fatal error machinery: terminate the current
operation (if any), indicate the problem in the FATAL_ALERT_CAUSE
register, and sometimes in the ERR_BITS register.

Fixes #2696

Signed-off-by: Philipp Wagner <phw@lowrisc.org>