[otbn,dv] Finesse "barely_oob_addr_cp" to make it possible to hit
I'd been overly clever here and was looking for an aligned address
where the base pointed at a valid byte of memory but the top fell off
the top of memory. Unfortunately, aligned addresses are multiples of
4, and the load/store width is also 4 bytes, which means this isn't
actually possible!

Repurpose the coverage point for "just above the top of memory", which
might be worth seeing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>