[dv/kmac] loosen timing requirements in kmac vseq
This PR loosens the timing requirements in
`kmac_base_vseq::wait_for_kmac_done()` task.

The task is currently implemented using `csr_spinwait`, which puts a lot
of timing-related pressure on the scoreboard (especially with
zero_delays enabled), and doesn't necessarily add much to the quality of
verification.

This task is updated to now use a "before" and "after" CSR read, with a
long wait in between to allow the KMAC to finish hashing, rather than
send out a CSR request every few cycles with `csr_spinwait`.

Signed-off-by: Udi Jonnalagadda <udij@google.com>