[spi_device] Add `upload` `busy` fields into cmd_info
This commit adds two fields, `upload` and `busy`, into the command
information data structure. Those two fields will be used in the upload
submodule.

Also, the cmdparse module extends its lookup of the command info slot
from up-to ReadCmdEnd to the entire command information slot
`NumCmdInfo`. The extension is to check opcode and check if the logic
uploads the on-going transaction.

To support the upload function, the number of the command information
entries is increased by 8, from 16 to 24.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>