[sw/silicon_creator] Add reset reason enum to rstmgr driver
This attempts to both centralize the definitions of the reset reason
bitfields and also make them slightly more robust. In particular
this change should catch modifications to the layout of the HW_REQ
field at compilation time.

This change also adds the `rstmgr_reason_clear` function that can
be used to clear reset reasons that have already been processed.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>