[dvsim] Minor fixes to coverage extraction
The issue was reported by @rasmus-madsen - if the cleanup tasks in
Deploy::post_finish() fails, it attempts append the error message to a
Launcher class member called `fail_msg` which has not been created,
causing the dvsim invocation to bomb improperly.

The primary goal of this change is to allow the failure of the cleanup
tasks to also be factored into whether the job failed or not. This is
achieved by making the job `status` a member of the `Launcher` class
which is finally returned to the Scheduler. Cleanup tasks are run
regardless of the job's outcome, but now that invocation is updated to
catch any exceptions thrown, so that the `status` can be updated if the
cleanup tasks fail for whatever reason, within the
`Launcher::_post_finish()` invocation.

In addition, a minor change in the Xcelium coverage extraction code
fixes the issue of some coverage metrics not showing up correctly.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>