[prim_lfsr] Do not shadow |state| variable
A variable called `state` is used inside a function (with function
scope), and also in the module with module scope. This makes the
variable in the function shadow the one in the module. Verilator reports
this as lint warning. Since we build the OTBN smoke test in a way that
fails on warnings, we get a failing CI run for the OTBN smoke test.

This commit renames the variable inside the function to avoid the
shadowing.

```
%Warning-VARHIDDEN: ../src/lowrisc_prim_lfsr_0.1/rtl/prim_lfsr.sv:420:80: Declaration of signal hides declaration in upper scope: 'state'
  420 |                                                           logic[LfsrDw-1:0]    state);
      |                                                                                ^~~~~
                    ../src/lowrisc_prim_lfsr_0.1/rtl/prim_lfsr.sv:377:26: ... Location of original declaration
  377 |   logic [StateOutDw-1:0] state;
      |                          ^~~~~
                    ... Use "/* verilator lint_off VARHIDDEN */" and lint_on around source to disable this message.
```

Signed-off-by: Philipp Wagner <phw@lowrisc.org>