[keymgr/fpv] Fix formal compilation error
We found the `ASSERT_INIT` throws a compilation error saying the
`kmac_en_masking_i` is not a constant. We were thinking we can change it
to an assumption, but that might not be ideal for top-level.
I consulted with Weicai, if there is any issue related to this,
top-level tests could easily find the error. So we think it is safe to
just remove this assertion.

Please let uw know if you prefer a better solution.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>