[otbn,dv] Fix logic in "deep loop" generator
This generator tries to generate a deep stack of nested loops, until
it blows the HW stack. Since we don't actually want to complete any of
these loops, it's supposed to pick bodysizes so that the loops end on
addresses that are out of range or have already got instructions. The
trick is that, because we never go back to code that we've visited
before, this ensures the loops will never complete.

Unfortunately, I got the check backwards! Interestingly, it doesn't
matter most of the time, but it *does* cause explosions if guess == 1.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>