[tlul,rtl] Rewrite a_ack_q computation to avoid lint warnings
With the previous version, the a_ack_q flop was dead zero if
EnableIntg was false. This caused AscentLint warnings about a constant
flop. Re-jig things so that the flop is only generated when it will be
used and the output signal (a_ack_q) is a constant zero if not.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>