[otbn] Introduce locked state
Introduce a new locked state, represented as LOCKED in the STATUS
register to indicate if OTBN has detected a fatal error and is now
"dead" until a reset.

With this new state, OTBN guarantees a certain value of the STATUS
register after a fatal error was detected, simplifying the documentation
and the verification.

Furthermore, operation handling from Ibex got easier even if fatal
alerts don't reach Ibex for whatever reason: an operation always returns
with a done interrupt, and reading the STATUS register clearly indicates
what happened to OTBN. (Before, software would have to read
FATAL_ALERT_CAUSE as well if STATUS was non-zero to check if an
operation can be restarted [in case of a recoverable error], or if
OTBN is locked.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>