[dv, xcelium] Properly pass excl files to IMC
To the report and analysis stages, the exclusion files were not being
properly passed. This PR fixes that.

Now, the generated report and the IMC GUI (analysis) should show the
exact same coverage numbers.

The following changes are made:
- A standardized variable 'xcelium_cov_excl_script' is added to HJSon to
add exclusions as TCL scripts. I suggest making this the recommended and
the only way of setting exclusions.
- This is passed to IMC invocation during both, report and analysis
stages.
- The AES vRefine file is removed because it is not needed (those
exclusions are already captured in the common TCL file).
- Some minor fixups.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>