[topgen] Add ability to query the blocks in top
This commits adds a very basic ability to query, specifically the list
of blocks in a top level design described in the supplied HJSon.

The `--get_blocks` switch added to it in this commit serves that
purpose.

Ideally, `topgen` should not create any files if we are only querying
things. It should just return the info sought to stdout. Due to the way
`topgen` is written, it unfortunately does mandate that an output
directory be provided, and it does create some files. This is documented
in #7857.

It is recommended to pass `--outdir $REPO_TOP` with the `--get_blocks`
switch, so that the sources are generated relative to the repo itself,
which should be identical to what is already checked in. Though,
ideally, the necessicity to always supply an output directory needs to
be eliminated (discussed in #7857).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>