[tool,check_dif_status] Better rep of DIF progress
This commit enhances the check_dif_status.py script by allowing the user
to pass the top level HJSon file, from which we can extract the full
list of blocks in the design. This is better than the existing approach
where the list of available blocks is computed by looking at the sources
in the DIF area. If the DIF sources dont even exist, then it does not
get accounted for.

This commit adds a `--top-hjson` switch, which is used to pass the top
level HJson file (relative to repo_top). If passed, it invokes the
`topgen.py` script to get the list of IP blocks in the design. If not
passed, it falls back to the current behavior, while throwing a warning
stating the the table may not present the full picture. This switch is
thus optional (but highly recommended).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>