[prim] Edge Detector
This commit introduces the edge detector in the prim lib.
The edge detector detects the signal changes by latching the signal into
the register and comparing the latched value and the current input
signal.

It also provides the synch option which is useful when the input signal
comes from outside of the chip (PAD) or from the other clock domains.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>