[prim_prince] Mark "leaf" functions in prince_ref.h as static inline
This is C code, so we can't just use "inline", but if you just declare
the functions as "static", some compilers will warn about unused
function definitions if they aren't all used.

The "static inline" incantation is the closest we can get to C++'s
"inline" semantics (which are what we actually want).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>