[hw] Simplify how we narrow to blocks in hw/ip in hw/Makefile
Rather than having a phony target that uses a bash "if" statement to
decide whether to do something, start by filtering out the blocks that
we don't want to run on in the Makefile itself. The resulting code is
a bit easier, and the console output is much easier to copy-paste if
something goes wrong.

Also, use Make patterns to make the rules much easier to read (in each
case, we make sure that $* is the name of the block).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>