[rom_ctrl] Use FileSiz to determine ELF segment size
MemSiz includes memory that is allocated but not loaded. FileSiz
is the actual size of the segment in the ELF file.

For example, the mask ROM ELF file contains a segment with both
.data and .bss packed into it. The FileSiz is 0xc bytes (the size
of .data) while the MemSiz is larger at 0xa8 bytes because it
includes the .bss section that is not loaded from ROM and takes up
no space in the ELF file.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>