[otbn] Make termination on underflowing x1 more explicit in ISS
This will cause slight changes to instruction behaviour in cases where
we might have seen multiple errors. Until now, we haven't been testing
this against the RTL, so we can probably assume we're switching one
set of mismatches for another(!).

But now we are more explicit about what's going on, which is probably
helpful in the documentation. What's more, we can now model things
like "if I do an LW that loads from x1 and underflows, I shouldn't see
an error caused by an invalid address"

Note that, while we handle CALL_STACK errors on underflow in insn.py,
we still need the GPRs.err_bits() method in order to handle overflow
properly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>