[otbn] Alter ISS behaviour for invalid BN.SID/BN.LID instructions
For a sequence like the following:

    li      x10, 100
    bn.sid  x10, 10720(x0)

there are two reasons that the BN.SID instruction should raise an
error. Firstly, x10 is 100, which isn't a valid WDR index. Secondly,
the store address is 10720, which is above the top of memory. The
error bits for these two errors are ILLEGAL_INSN and BAD_DATA_ADDR,
respectively.

The RTL signals both errors, which seems reasonable but doesn't match
what the ISS was doing. Change the ISS to match.

The same argument applies to BN.LID.

Note that there are lots of examples of "multiple errors" in a cycle
that need checking. This is just urgent because we've suddenly started
generating streams with the RIG that happen to trigger this case quite
frequently.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>