[otbn,dv] Write unit tests to exercise each "multiple error" case
This addresses each of the situations where an instruction could
theoretically generate more than one synchronous error. The "expected"
files are set up so that the (ISS) test passes. But they probably
don't match the RTL at the moment: we'll follow up, amending the
ISS+test or RTL (or both) to make things a bit more sensible and get
everything matching.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>