[spi_device] TPM over SPI specification
This commit adds the TPM over SPI section in the SPI_DEVICE HWIP
Technical Specification.

TPM over SPI function is added in previous commits. This commit
describes the overall TPM submodule inside the SPI_DEVICE IP. The
document does not explain the details of the TPM spec.

It is assumed that the readers have enough knowledge in the TPM and the
SPI spec.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>