[otbn,dv] Add functional coverage for multiple errors
This patch defines coverpoints and crosses to track the multiple error
cases documented in the previous one.

It also does some refactoring for computed values in the covergroups.
For example, SW and LW need to compute an address from and some
immediate bits plus an operand value and all the "$signed(operand_a) +
$signed(offset)" stuff was getting rather hard to read. Factor it out
into the callsite for the covergroup's sample method. The nicest thing
would be to have a local variable in the sample method itself, but it
doesn't seem that the SV spec allows that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>