[prim_lfsr] Improve statistics of non-linear output
The non-linear output option makes use of an Sbox-4 layer to transform
the data. In order to ensure no statistical defects this patch
introduces a permutation layer before the Sbox-4 layer that is taken
from a block cipher. This permutation is constructed such that no
adjacent bits feed into the same Sbox, as together with the linear
shifting pattern of the LFSR this could create statistical defects.

Note that this change restricts the LFSR widths to 32bit and 64bit when
the output non-linearity is turned on.

Also, the optional StatePerm permutation that can be defined at
instantiation time is moved after the Sbox-4 layer to not interfere with
the block cipher permutation.

Signed-off-by: Michael Schaffner <msf@google.com>