[dv/csr] Update write exclusion wdata value
This PR fixes issue #8234. In current code, if a field is excluded from
CSR write, we mask out the excluded fields to 0. This is under the
assumption that all fields' reset value is 0, which is not always the
case.
So this PR adds a function to find the mirrored value of the excluded
field from write, and update the wdata value accordingly.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>