[prim_lfsr] Further permutation refinements for SBox layer
As pointed out in
https://github.com/lowRISC/opentitan/pull/8222#discussion_r711229598
the block cipher permutations from PRESENT might not be ideal, since
together with the LFSR shifting property, they cause neighboring SBoxes
to have the produce the same output value once cycle appart.

This commit modifies the permutation to break that shifting pattern, and
adds FPV assertions for verification of the desired permutation properties.

Signed-off-by: Michael Schaffner <msf@opentitan.org>