[otbn] Use correct bus_intg_violation signal in ERR_BITS
The `bus_intg_violation` signal is driven at the toplevel, and then
wired through to the controller to be assigned to the `err_bits` struct.
We then wire the whole struct back up to the toplevel to assign it to
the `ERR_BITS` external CSR.

Previously, we used the signal that happens to be around in the
toplevel, where we should have used the one coming from the controller
instead. Fix that.

(The FATAL_ALERT_CAUSE external CSR is using the signal directly from
the toplevel.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>