[otbn,dv] Load ELF on each iteration in otbn_reset_vseq
We were getting away with not doing so before (by fluke!) but it could
still go wrong if, for example, the loaded binary stored to an address
that we expected to initialise with data from the ELF.

As it turns out, there's another problem: while the memory contents
survive a reset, the loop warping rules don't. This caused a nightly
failure (seed 531020925 with commit 268280674).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>