[otbn,sw] Ensure entry points are at address zero
In order to remove the START_ADDR functionality, we need to make sure
that the software we want to run really does start at address zero!
This patch adds a new input section to the linker script called
".text.start". The idea is that you should put the entry point there,
which will force it to appear at address zero.

The patch also ports each software code snippet that we load from Ibex
so that it uses the .text.start section.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>