[cip,dv] Expose "abort" and "error" separately in tl_access_w_abort
The status signal that was there before was UVM_OK if the access
completed and didn't signal an error. This is perfect for answering
questions like "should I update the RAL for this access?".

Unfortunately, it doesn't allow you to spot "an access ran to
completion and caused an error" because you can't distinguish it from
"the access was aborted".

Split the two signals and teach run_mem_partial_access_vseq_sub (the
one callsite that used them) how to use them properly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>