[ipgen] Support object template parameters
Some ipgen templates require complex data structures as template
parameters: nested combinations of lists, dicts, string and integer
values. The parameters passed to the clkmgr are a good example: they
consist of information about all clocks, their associated resets, and
much more. Having individual string or int template parameters is
infeasible.

This commit introduces a "catch-all" template parameter type called
"object". It can be any data structure, as long as it can be serialized
to Hjson. We actually check for that to ensure round-trip safety even if
the template parameters are not written to a config file and reloaded,
but passed to the template directly.

The checking is using the JSON serializer, as extended by the Hjson
library. This makes it possible to simplify the implementation
considerably, at the cost of a slightly less beautiful and less
performand implementation (since we do a object -> json -> object
round-trip with associated string parsing).

Signed-off-by: Philipp Wagner <phw@lowrisc.org>