[otbn,dv] Add a simple predictor for the STATUS register
This register jumps to "LOCKED" when we see a write with an integrity
error. Ideally, we'd track this sort of thing with the predictor in
the scoreboard but the automated CSR tests disable the scoreboard and
its predictor (but still check for expected values of CSRs!).

Hack in the update logic that we need.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>