[otbn,dv] Properly mirror STATUS reg from ISS to SystemVerilog
Also, mirror all the other registers in a slightly less ad-hoc fashion
in ISSWrapper.

There's a bit of churn in the model connection, because we were using
a signal called "status" to mean "the current state of the model".
This signal is now called model_state.

The status_o signal isn't actually connected in the UVM or Verilator
testbenches yet (so doesn't really do anything): this will be
connected in a subsequent commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>