[otbn, dv] otbn_imem_err test added along with related changes
The following changes have been made in this commit:
1. otbn_imem_err_vseq.sv: This is the sequence required to run otbn_imem_err
test. This is a new file.
2. otbn_scoreboard.sv: otbn_scoreboard is taught to expect a fatal error
when the model status goes to StatusLocked.
3. otbn_base_vseq.sv: run_otbn function is made to stall until otbn status is not
StatusBusyExecute instead of waiting for otbn to go to idle.
4. otbn_model_if.sv: wait_status now waits on negedge of the clock
instead of posedge to avoid race condition between alert arrival and
model state change. It ensures that the scoreboard sees the state change
before the fatal alert changes and knows it has to expect a fatal alert
and will not throw a fatal error.

Signed-off-by: Prajwala Puttappa <prajwalaputtappa@lowrisc.org>