[util] clkmgr groups enhancements
- This commit is necessary to address #8405 and #8037

Modules are now allowed to delcare a different clock group per clock.
This allows for some flexibility in how things are hooked-up.

Further the idea of "internal" clocks are introduced for clocks that
are generated entirely inside a module but still used with the regfile.

This is especially useful for creating CDC structures for blocks like
clkmgr that has locally generated clocks.

Signed-off-by: Timothy Chen <timothytim@google.com>