[otbn,dv] Weaken trace comparison checks to allow IMEM errors
We want to allow the testbench to tell the ISS that IMEM has been
corrupted. When this happens, we're going to have the ISS generate a
line like this:

    E PC: 0x00000094, insn: ??

Of course, that's not exactly equal to what the RTL spits out. But we
can't match the RTL's output exactly, because the RTL's "insn" field
might have been corrupted. To support this properly, this patch
weakens the check in OtbnIssTraceEntry to require that everything
matches up to the '??'.

You could imagine something much more general: for example, we could
interpret the output from the ISS as a glob or a regexp. But that
seems overly hard: do the stupid thing for now!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>