[englishbreakfast] Add missing dependencies on crossbar
8f2cb38e922989cf76b7eff96bb193e882d05034 changed the dependencies for
the crossbar to use unique names per toplevel. The required change was
done for the earl grey toplevel, but not for english breakfast. Adjust
the two.

This problem wasn't apparent before because the use of the topgen
generator created a dependency on all cores in the output folder, i.e.
everything that topgen produced.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>