[prim_lfsr] Minor lint fixes
Since j equal to or less than LfsrDw and the variable it
assigns into is $clog2(LfsrDw), the variable should always
be large enough to accommodate j.  Remove the padding bits
as the lint tool thinks its a negative number and spits out a
pile of warnings.

Signed-off-by: Timothy Chen <timothytim@google.com>

[prim_lfsr] Minor lint fix

tie-off sbox output when lfsr width is greater.  In these
situations, not all lfsr bits will be used in the permutation.

Signed-off-by: Timothy Chen <timothytim@google.com>