[otbn,dv] Report IMEM fetch errors more helpfully in OTBN tracer
If we've seen an IMEM fetch error, this changes the trace from
something like

    E PC: 0x00000b10, insn: 0x84d72c7b

to something like

    E PC: 0x00000b10, insn: ??

Since the IMEM rdata is known to be bogus at this point, we don't
really want to display it. This also makes things easier on the ISS
side, where we don't need to model changes to the bits themselves
that are caused by 1- or 2-bit injected errors.

This patch also makes sure that we always print out an 'E' line.
Before, it was either 'E' or 'S', depending on whether the current
instruction was stalled or not. (Which would be very finicky to model
in the ISS and not really do anything useful).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>