[dvsim] Correctly set self_dir for included Hjson files
The `self_dir` variable (used as `{self_dir}`) in dvsim-processed Hjson
files is intended to contain the absolute path to the Hjson file it
appears in. This was the case for top-level Hjson files (i.e. a Hjson
config file which was directly passed to dvsim.py), but not for included
files. Fix that.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>