[dvsim] Add a project folder name when `scratch root` environment variable is set
This PR solves a small issue when the `SCRATCH_ROOT` environment variable is set, and the
workstation has more than one copy of the repository.
Then we add the repository name to distinguish the output.
For example: the local workstation has two repositories named
`opentitan1` and `opentitan1`, the `SCRATCH_ROOT` is set to `~/desktop`,
then the two scratch roots are: `~/desktop/opentitan1` and
`~/desktop/opentitan2`.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>