[otbn,dv] Fix how we install loop warps in otbn_top_sim
The code was previously installing them on the first clock edge, which
actually happens before pull the reset line. This was ok, because we
don't then reset the values in the model, but a bit confusing!

We want to make a change that *will* reset loop warps properly in the
model on reset, so we need to sort this out first.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>