[otbn,dv] Sample RTL signals on first cycle of instruction
We sample various internal RTL signals in order to do coverage
collection. These are supposed to be "the value of the state just
before the instruction ran". Most of the time, this is the same as
"the value of the state just before the last cycle of running the
instruction", but that's not quite always true.

In particular, consider the following instruction:

  BN.LID x1, 0(x0)

This reads from the call stack (x1) in order to get a destination
address. With the current RTL implementation, the call stack pop
happens on the first cycle which means that the previous code was
sampling too late and saw an empty call stack. We wouldn't have
noticed, except that the checks on "lockl_x1_uflow" in otbn_env_cov
were failing (because we thought we should have seen a call stack
underflow and the RTL disagreed).

Fixes #8587.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>