[util] Add --avoid-zero to sparse-fsm-encode.py
The sparse-fsm-encode.py script can be used for things other than
FSMs. One example is for a CMD register (for OTBN), where it seems
like a good idea to avoid magic values that are close to zero.

With this flag, three 8-bit magic values can be generated with:

    util/design/sparse-fsm-encode.py -d 5 -m 3 -n 8 --avoid-zero -s 1

yielding

    State0 = 8'b11001101,
    State1 = 8'b01111110,
    State2 = 8'b10110011

Without the flag, the same seed generates

    State0 = 8'b10010001,
    State1 = 8'b00011110,
    State2 = 8'b11100100

with a minimum popcount of just 3.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>