[opentitantool] Add hook for bootstrapping code on emulators
When running an OpenTitan emulator, we might want to use OpenTitan
tool to "flash" a new firmware image.  We would not want to do so by
SPI transactions like the real chips, except when testing the read
only bootloader code.  Instead, we want to reach into special support
from the emulation system itself, to be able to replace the entire
content of the emulated flash storage.

This CL adds a generic method to the Transport trait for specialized
operations like that.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>