[dif] Prepare templates for rv_timer modifications.
To adapt the templates to auto-generate the RV timer IRQ DIFs, the
template functions must be adapted to be more generic (in terms of
passing in the register offset macro/variable with the address).
Additionally, the "params_list" from the RV timer's HJSON will need to
be parsed, to exract the number of HARTs/timers, so that capability is
added to the make_new_dif.py tool.

This partially addresses #8142.

Signed-off-by: Timothy Trippel <ttrippel@google.com>