[otbn,dv] Maybe use IRQ to detect end instead of polling STATUS
Configuring this goes in two steps: firstly, you have to write to the
enable register. There's an enable_interrupts_pct knob that sequences
that care can get from the environment config to spot that it is
needed.

Then the run_otbn() task in the base class chooses whether to use
interrupts or polling, based on whether interrupts are enabled (if
not, it will poll!) and poll_despite_interrupts_pct.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>