[spi_device,lint] Remove double-import of parameters
In spid_status, we're already importing spi_device_pkg::* into the
module's scope, so this line was shadowing CmdInfoIdxW (with itself!)
and Verilator spat out a warning.

In spi_passthrough, we were importing spi_device_pkg, which exports
NumCmdInfo, shadowed by the NumCmdInfo parameter. We then instantiate
it (in spi_device.sv), setting the parameter to the thing it's
shadowing. Get rid of the parameter entirely.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>