[otbn,dv] Properly handle unknown instructions in coverage
In 38a36b5, I switched from "dummy-insn" to "??" to match the trace
output when we try to fetch invalid data from IMEM. What I didn't
realise is that we do actually generate "dummy-insn" as well. That
happens when we have valid data from IMEM that doesn't decode to a
valid word.

Renaming that situation from "dummy-insn" to "??" doesn't seem right:
it's probably better to have a name that makes it clear what's going
on. But an IMEM integrity error probably shouldn't talk about "dummy
instructions" either.

This patch allows both names.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>