[tooling] Remove clang-format invocation from topgen.py.
This fixes #8721 and #8863 by removing the the invocation of
`clang-format` on autogenerated SW sources that are generated as a
part of invoking `make -C hw`. While not formatting these files with
clang-format produces code that is slightly in-consistent with our
codebase, it makes it easier for contributors with slightly more complex
developement environment restrictions to still produce auto-generated SW
that passes CI checks for stale autogen'd code.

Signed-off-by: Timothy Trippel <ttrippel@google.com>