[dif] Autogen all `dif_<ip>_alert_force()` DIFs.
This completes all tasks and fixes #8879. Specifically this commit:

1. updates the DIF autogen *.c, *.h, and *_unittest.cc templates to add
   a `dif_<ip>_alert_force()` DIF and associated unit tests,
2. re-generates all auto-generated DIFs to include the new templated
   alert force DIF, and
3. deprecates existing (manually-implemented) alert force DIFs for the
   few IPs that implemented them.

Note, due to staleness checks in CI on auto-generated DIFs, these
changes had to be combined into a single commit. While the commit
appears to be quite large (especially the number of files changed), this
is purely due to the fact that the autogen DIF templates changed,
causing all the auto-gen'd DIFs to change too.

Signed-off-by: Timothy Trippel <ttrippel@google.com>