[dif] Add overwrite safety check to `util/make_new_dif.py`
If `util/make_new_dif.py` is run initially to auto-generate a DIF header
and checklist boilerplate code/markdown (respectively), and accidentally
run again shortly after manualy modifying the boilerplate code (before
the modifications are checked-in to the repo), the modifications would
be overwritten. This commit adds a check to prevent such a mistake from
happening, partially addressing an action item in #8142.

Signed-off-by: Timothy Trippel <ttrippel@google.com>