[otbn,util] Halve the DMEM region in OTBN's linker script
This is to match how only the first half of DMEM is bus-accessible.

To put data in the scratchpad (either for security, to ensure it's not
bus-accessible, or because the program doesn't fit otherwise(!)), put
it in section .scratchpad. This will ensure it gets placed in the
latter 2kb of DMEM and marked NOLOAD.

We also slightly tweak the objcopy command in otbn_build.py, so that
it discards the .scratchpad section when making the ELF to be embedded
into Ibex code. This has the nice side effect of causing link errors
if you accidentally try to point Ibex code at a symbol that's defined
in an inaccessible region.

The only change to the code itself is in rsa.s (duplicated in
silicon_creator), where local variables get put in .scratchpad to
avoid running out of space.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>