[prim] Add option to not clear the packer FIFO upon read
There are cases where leaving around the data just read doesn't hurt but
instead outputting a deterministic value after read should be avoided.
For example, if the packer FIFO is used to feed pseudo-random data into
an LFSR, having the packer output such a deterministic value most of the
time is bad as it may simplify an attack trying to load the LFSR with
deterministic instead of random values.

For this reason, this commit adds a new parameter to the packer FIFO
primitive to control this behavior on a per-case basis and disables the
clearing when used to feed LFSRs or inside prim_edn_req.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>