[otbn,sw] Use weak symbols to shrink data in rsa_verify_3072_test
The code in rsa_verify_3072.s expects various input and output arrays.
The code in rsa_verify_3072_test.s wants to initialise these arrays
with known values. To do that, it was declaring another copy of each
array (with the known values) and copying the values across.

Rather than doing that, use weak symbols and per-data sections in
rsa_verify_3072.s and then override them in the test file.

Something like this is needed if you want to get the total externally
visible data down below 2kb.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>