[prim] Tweak prim_sync_reqack_data assertion so it can be disabled
When we run $assertoff(), it stops any new assertion sequences from
being started. However, this part of the previous assertion:

    $stable(data_o) [*2]

consumed time. If we inject a reset on the DST side that causes the
value to change in the last cycle of the window, an $assertoff won't
save us because the assertion that started on the previous cycle will
still run to completion.

Unfold the $stable(..) calls into comparisons between $past() and
present and change the assertion so that it doesn't consume time,
allowing it to be disabled at any point.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>