[otbn,dv] Run Python model with -u
This disables buffering at the block level for stdout and stderr in
the Python subprocess. Without this argument, it seems that Python
3.6, at least, buffers sys.stderr at the block level if it doesn't
point at a console. When run under dvsim, it points at the run.log
file.

With this fixed, you can add debug prints to the model with something
like

    print('my message', file=sys.stderr)

and have it appear in the log output. Without the flag, you need to
follow with a sys.stderr.flush() to be sure to see the result.

This isn't a problem with recent versions of Python: it seems that
they changed the default behaviour at some point.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>