[spi_device] One clock early for the cmd_info timing
cmd_info & cmd_info_idx are registered output in the cmdparse module.
The upload module in SPI_DEVICE uses cmd_info to determine if the
address field exists or not.

The cmd_info value arrives to the rest of the module one clock late. It
results in the upload module to assume the command does not have the
address field.

This commit pulls in the cmd_info one clock early. It leads to longer
datapath as cmdparse cannot register the data.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>