[syn/cdc] Tune SDC file to fix all SDC warnings/errors in CDC
This introduces some more flexibility to the SDC file so that we can
reuse it for open-source-only CDC runs where the hookup paths have
different hierarchical names.

Also, there are a few constraints that CDC does not understand, and we
introduce a variable to exclude them conditionally for CDC flows.

Signed-off-by: Michael Schaffner <msf@google.com>