[opentitantool] Add Skeleton of host emulation driver
The Ti50 team will compile our code to either OpenTitan, to past
Google Titan chips, or to a third target which is x86 code running
inside Linux processes on the developer's machine.

The team hopes to be able to run some of the same tests across all
three targets, ideally using the same sequence of OpenTitanTool
invocations.

This patch creates an empty Transport implementation, for the Ti50
team to later populate. I am posting this pull request in part to
rubber stamp the approach of putting this code into the OpenTitan
repository.  We will strive to make the socket protocol generic enough
that other emulators can use this same transport driver.

Signed-off-by: Jes B. Klinke jbk@chromium.org