[spi_device] Data Manipulation logic
This commit implements the data manipulation feature in the passthrough
mode.

The data manipulation in the passthrough let SW change any bit of the
first 4 bytes of the write data to the downstream flash device.

To enable, SW may configure:

- PAYLOAD_SWAP CSRs: set mask, data to change certain bits of the
  payload.
- CMD_INFO.payload_swap_en: set to 1 to enable the feature (command
  specific)

Note that enabling the feature depends on the command basis and the SWAP
data is shared across all the commands.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>