[sw/crypto] Clarify bounds assumptions for Montgomery R^2.
The Ibex implementation, on closer inspection, doesn't actually require
R / 2 < M, so the comment is rephrased to not state that it does. The
OTBN implementation is greatly simplified by the assumption that R / 2
< M, so it is modified to assume so and justify the assumption by citing
FIPS.

This commit also includes a minor adjustment to the R^2 loop, making it
a loop instead of a loopi so as not to consume the loop stack
unnecessarily.

Signed-off-by: Jade Philipoom <jadep@google.com>