[sw, tests] Refactor sram_ctrl scramble test and test utils
The main purpose of this change is to make the test util more
generic. It also includes some suggestions from the other PRs, such as
avoiding dependencies on top_earlgrey in test utils.

- Renamed test utils API to match accepted style
  `sram_ctrl_testutils_<...>`.

- Testutils are more generic, they operate on the passed in address
  instead of top_earlgrey addresses calculated inside the util.

- Init functions have been removed from test utils.

- Unnecessary data structures have been removed from test utils.

- Test has been simplified to only perform a single check at the start
  of RAM, as opposed to former the start and the end of RAM. If
  required, we can always add the end of RAM check later.

In a nutshell, the test util API was simplified and some of this
complexity was moved to the test instead.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>