[top/pinmux/pads] Pull the IE signals into the pinmux
So far the input enable signals of the pads have been hardwired to 1.
This can burn power unnecessarily - especially during deep sleep.
Therefore, this change pulls the input enable signals into the pinmux so
that we can enable the input buffers only when the associated MIOs /
DIOs are selected (either by the mux matrix or a wakeup detector).

Long-term, the input enable signals should be pulled down into the
individual peripherals for even more granular control - but this change
would be too invasive at this point, hence it is deferred (see issue

Signed-off-by: Michael Schaffner <msf@google.com>