[sw/ottf] Reduce the size of the FreeRTOS heap.
The FreeRTOS heap is allocated as a static array. Currently, our flash C
runtime (that we use for testing) zeros out the .bss section of memory
before transferring control flow to`main()`. This results in a
performance hit for tests that do not make use of the FreeRTOS OS layer,
i.e., non-concurrency tests. This commit reduces the size of the
FreeRTOS stack to 0x800u bytes for the time being.

Signed-off-by: Timothy Trippel <ttrippel@google.com>