[spi_device] Passthrough logic to see intercept
This commit revises Passthrough FSM to observe the intercept event not
only in the Address state but also in returning the data.

The port name is still `mailbox_hit` but internal name is revised to
`intercept` to cover broader events (occupied by Read Status, Read JEDEC
ID, Read SFDP, or Read Cmd to Mailbox region).

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>