[sw/ottf] Add malloc failure and stack overflow hooks.
The OTTF uses FreeRTOS to enable running concurrent tasks within a test.
When a new task is created, memory for the TCB and stack are allocated
using FreeRTOS's malloc implementation. Currently the size of the heap
is fixed to 0x800u bytes. This commit adds FreeRTOS hooks that display
a useful message to test developers and abort the test when FreeRTOS
detects a malloc failure or stack overflow.

Signed-off-by: Timothy Trippel <ttrippel@google.com>