[dv] Fix shape calculations for replicated ECC
data_width is supposed to be the width of a word in bits, not counting
any ECC bits. The computation that was there before wasn't quite
right because it didn't allow for the possibility of multiple
subwords. We use this structure for OTBN, where we have 256-bit words,
split into eight 32-bit subwords where each is protected with the
Ecc_39_32 scheme.

Also bump up the maximum bytes per word (we need more!) and add
read256() and write256() functions.

Finally, fix the instantiations of imem_util and dmem_util in the OTBN
testbench (I'd misunderstood what the n_bits argument did).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>