[dv] Teach encrypt/decrypt_sram_data to support OTBN
This commit teaches those two functions to support replication of
keystreams (in OTBN, we re-use a 64-bit keystream by concatenating it
with itself as many times as needed for the very wide word).

It also generalises the functions to allow the caller to select the
width of the subst/perm network. Before this commit, there was a
"byte_diff" flag that allowed you to select between full-width and 8
bits. This commit changes the flag to an integer that gives the width
explicitly. Call-sites that set byte_diff should have an sp_width of
8; those that didn't should have an sp_width equal to data_width.
OTBN's dmem can use an sp_width of 39 (and a data_width of 312).

The implementation is a little finicky because you have to extract
"chunks" of an unpacked array. For the 8-bit case, you could do things
like "data[i*8 +: 8]" but you can't do that if you don't know the
width at compile time (because range selects need a constant for their
width). As such, there's a big loop on each side of the call to
sp_encrypt or sp_decrypt to grab a chunk of data, operate on it, and
write it back to the main word.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>