[reggen] This adds initial support for a cm list
This adds initial support for declaring an optional list of
countermeasures in the IP Hjson.

The list is parsed into a list of CounterMeasure objects, and the
hierarchical naming is checked against a predefined format.

The actual CM_ASSETS and CM_TYPES definitions are not declared yet as
this is pending further internal feedback.

Signed-off-by: Michael Schaffner <msf@google.com>