[kmac] Use Empty signal from FIFO
The issue has been discussed in #9202

KMAC MsgFIFO sees the FIFO depth to calculate the empty status of SYNC
FIFO. When the logic writes data and the receiver acked the read data at
the same time and the FIFO has been empty already, the empty status
won't be de-asserted as the `depth` wont be increased.

The issue is that SW will not get the fifo empty event due to the
behavior described above if the receiver logic is faster than the SW.

This commit revises the msgfifo logic to see the `rvalid` as an
inversion of the empty status. The scenario above will lower the `empty`
signal with the fix.

Co-Authored-by: Cindy Chen <chencindy@opentitan.org>
Signed-off-by: Eunchan Kim <eunchan@opentitan.org>