[otbn, rtl, dv] Add prefetch stage
The prefetch stage fetches instructions a cycle ahead of when they are
required. This is to enable flopped register file inputs to implement
blanking.

The prefetch stage isn't really speculative. It only prefetches where it
can be confident of the address, in situations where it can't
(specifically branches and jumps) it does not prefetch.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>