[dv] Add sub RAL block creation knobs
This commit was an attempt at a fix for #9323 which was unsuccessful.
The idea was to provide per-sub-RAL block creation knobs which are
controlled at runtime instead. So for instance, the generic AST register
block within the chip RAL model could be skipped from creation at
runtime. The provided callback functions `pre|post_build_ral_settings()`
would be used to achieve this. The extended (external) chip env cfg
class could skip the generic RAL blocks for instance and maintain
separate actual RAL blocks in their env instead. Chose to keep these
updates but not execute this idea.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>