[otbn,sw] Zero the work buffer in rsa.s
Commit bc424b2 switched the code to using a "work buffer" as a
destination for the encrypt/decrypt operation and then copying its
contents back to the "inout buffer" when the operation was done.

If we want the buffer to be in a scratchpad region and not initialised
by Ibex, we also need to zero it before use. Here, the problem isn't
that we care about its contents. Rather, we need to make sure we don't
have failing data integrity checksums when reading them at the end of
the operation to copy them back to the inout buffer.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>