[rom_ctrl] Make the mux select into a MuBi4
This strengthens the "did we switch back to the checker?" logic.
Firstly, we've got sel_reverted which corresponds to the check that
was in alert_o before. However, we also take a copy of sel_bus_q and
check for reversion *there*. The idea is that if an attacker somehow
managed to override sel_buf_q, there would be yet another check to
catch it.

The sel_bus_i signal that tells the mux to switch to giving access to
the bus comes from rom_ctrl_fsm and boils down to a check that the FSM
is in state Done. We bulk out the sparse encoding for that FSM by
adding the MuBi4 value that compares the state with Done. That way, we
can pass those bits straight through to the mux without ever having a
glitchable 1-bit intermediate.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>