[opentitantool] Fixes to bootstrap of legacy Titan
This newly added code has a few bugs.

1. A bug was introduced during review, when the code to set the
"final" flag on the last frame was move out of the loop.  The hash of
the block needs to be computed after the flag is set.  (This bug would
have been caught had I run the one and only unit test in this file.)

2. The above lead to the discovery that the existing logic (also in
primitive.rs) does not attempt to verify the ack of the very last
block, leading to the false claim of successful bootstrapping.

3. Further real-life testing revealed more shortcomings of the
existing logic.  Ordinarily, the bidirectional transaction is used
such that the boot rom will ack the previously transmitted frame, at
the same time as receiving the next one.  I.e. the Rust code will look
for ack of frame 3 while transmitting frame 4.  If an unrecognized
checksum is received, the code will go back and retransmit frame 3
repeatedly, until ack of frame 3 is received.  However, if the cause
of the original bad ack was not in the boot rom, but due to noise on
the CIDO line, then the bootrom may have successfully received frame
3, and possibly also frame 4.  If the latter is the case, the boot rom
(on H1Dx at least) will repeat the ack of block 4 while frame 3 is
being retransmitted.  This CL introduces code to recognize this case,
and others, rather than only looking for ack of the frame most
recently transmitted.

TEST=Repeatedly flashed H1Dx while manipulating CIDO capacitance

Signed-off-by: Jes B. Klinke <jbk@chromium.org>