[mubi] Factor out some duplicated work in prim_mubi_pkg.sv.tpl
Almost no change to the generated code (except sorting out the length
of some box quotes), but this moves all the string interpolation where
we construct the mubi${nbits}_foo names into a header at the top. This
should hopefully be a bit easier to read in the template file.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>